Add long country code to referendum data

Augments data with an additional column holding the current or former three-letter ISO 3166-1 alpha-3 code of the country in which the referendum took place (see ISO 3166-3 for former country codes).

add_country_code_long(data)

Arguments

data

C2D referendum data as returned by rfrnds(). A data frame that at minimum contains the column country_code (with ISO 3166-1 alpha-2 or ISO 3166-3 alpha-4 codes).

Value

A tibble.

See also

Examples

c2d::rfrnds(date_max = clock::date_today(zone = "UTC"),
            date_min = clock::date_today(zone = "UTC") |> clock::add_years(-1L)) |>
  c2d:::add_country_code_long() |>
  dplyr::select(any_of("id"),
                starts_with("country_"))
#> ℹ Fetching JSON data from C2D API...
#> Downloading: 16 kB     Downloading: 16 kB     Downloading: 16 kB     Downloading: 16 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 65 kB     Downloading: 65 kB     Downloading: 81 kB     Downloading: 81 kB     Downloading: 81 kB     Downloading: 81 kB     Downloading: 81 kB     Downloading: 81 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 150 kB     Downloading: 150 kB     Downloading: 160 kB     Downloading: 160 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 200 kB     Downloading: 200 kB     Downloading: 200 kB     Downloading: 200 kB     Downloading: 210 kB     Downloading: 210 kB     Downloading: 210 kB     Downloading: 210 kB     Downloading: 210 kB     Downloading: 210 kB     Downloading: 210 kB     Downloading: 210 kB     Downloading: 210 kB     Downloading: 210 kB     Downloading: 230 kB     Downloading: 230 kB     Downloading: 230 kB     Downloading: 230 kB     Downloading: 230 kB     Downloading: 230 kB     Downloading: 240 kB     Downloading: 240 kB     Downloading: 240 kB     Downloading: 240 kB     
#> ✔ Fetching JSON data from C2D API... done [199ms]
#> 
#> ℹ Converting JSON to list data...
#> ✔ Converting JSON to list data... done [14ms]
#> 
#> ℹ Tidying data...
#> ✔ Tidying data... done [675ms]
#> 
#> # A tibble: 84 × 4
#>    id                       country_code country_code_long country_name
#>    <chr>                    <fct>        <fct>             <fct>       
#>  1 63849a3aa52c3995043a928c CH           CHE               Switzerland 
#>  2 63849949a52c3995043a9285 CH           CHE               Switzerland 
#>  3 6384954aa52c3995043a9279 CH           CHE               Switzerland 
#>  4 6384934ea52c3995043a9271 CH           CHE               Switzerland 
#>  5 6357a495a52c3995043a924a CH           CHE               Switzerland 
#>  6 6357a342a52c3995043a9242 CH           CHE               Switzerland 
#>  7 6357a0e6a52c3995043a923b CH           CHE               Switzerland 
#>  8 63579f05a52c3995043a9232 CH           CHE               Switzerland 
#>  9 634d67c1a52c3995043a922a CH           CHE               Switzerland 
#> 10 634d6256a52c3995043a9223 CH           CHE               Switzerland 
#> # … with 74 more rows