Add short country name to referendum data

Augments data with an additional column holding the common English name of the country in which the referendum took place.

add_country_name(data)

Arguments

data

C2D referendum data as returned by rfrnds(). A data frame that at minimum contains the column country_code (with ISO 3166-1 alpha-2 or ISO 3166-3 alpha-4 codes).

Value

A tibble.

See also

Examples

c2d::rfrnds(date_max = clock::date_today(zone = "UTC"),
            date_min = clock::date_today(zone = "UTC") |> clock::add_years(-1L)) |>
  c2d:::add_country_name() |>
  dplyr::select(any_of("id"),
                starts_with("country_"))
#> # A tibble: 84 × 3
#>    id                       country_code country_name
#>    <chr>                    <fct>        <fct>       
#>  1 63849a3aa52c3995043a928c CH           Switzerland 
#>  2 63849949a52c3995043a9285 CH           Switzerland 
#>  3 6384954aa52c3995043a9279 CH           Switzerland 
#>  4 6384934ea52c3995043a9271 CH           Switzerland 
#>  5 6357a495a52c3995043a924a CH           Switzerland 
#>  6 6357a342a52c3995043a9242 CH           Switzerland 
#>  7 6357a0e6a52c3995043a923b CH           Switzerland 
#>  8 63579f05a52c3995043a9232 CH           Switzerland 
#>  9 634d67c1a52c3995043a922a CH           Switzerland 
#> 10 634d6256a52c3995043a9223 CH           Switzerland 
#> # … with 74 more rows