Add turnout to referendum data

Augments data with an additional column turnout containing the voter turnout calculated as:

votes_yes+votes_no+votes_empty+votes_invalidelectorate_total\frac{votes\_yes+votes\_no+votes\_empty+votes\_invalid}{electorate\_total}electorate_totalvotes_yes+votes_no+votes_empty+votes_invalid​

add_turnout(data, rough = TRUE, excl_dubious = TRUE)

Arguments

data

C2D referendum data as returned by rfrnds(). A data frame that at minimum contains the columns electorate_total, votes_yes, votes_no, votes_empty and votes_invalid.

rough

Whether to fall back on a "rough" calculation of the turnout in case any of the variables votes_empty or votes_invalid is unknown (NA), or to be strict and return NA in such a case.

excl_dubious

Whether or not to exclude obviously dubious turnout numbers (those > 1.0) by setting them to NA. Such numbers stem either from data errors or (officially) tampered numbers.

Value

A tibble.

See also

Examples

# rough turnout numbers
c2d::rfrnds(country_code = "IT") |>
  c2d::add_turnout() |>
  dplyr::select(any_of("id"),
                electorate_total,
                starts_with("votes_"),
                turnout)
#> ℹ Fetching JSON data from C2D API...
#> Downloading: 16 kB     Downloading: 16 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 65 kB     Downloading: 65 kB     Downloading: 81 kB     Downloading: 81 kB     Downloading: 81 kB     Downloading: 81 kB     Downloading: 81 kB     Downloading: 81 kB     Downloading: 81 kB     Downloading: 81 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 150 kB     Downloading: 150 kB     Downloading: 160 kB     Downloading: 160 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 200 kB     Downloading: 200 kB     Downloading: 210 kB     Downloading: 210 kB     Downloading: 230 kB     Downloading: 230 kB     Downloading: 230 kB     Downloading: 230 kB     Downloading: 250 kB     Downloading: 250 kB     Downloading: 250 kB     Downloading: 250 kB     Downloading: 250 kB     Downloading: 250 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 280 kB     Downloading: 280 kB     Downloading: 290 kB     Downloading: 290 kB     Downloading: 290 kB     Downloading: 290 kB     Downloading: 290 kB     Downloading: 290 kB     Downloading: 310 kB     Downloading: 310 kB     Downloading: 310 kB     Downloading: 310 kB     Downloading: 330 kB     Downloading: 330 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 360 kB     Downloading: 360 kB     Downloading: 380 kB     Downloading: 380 kB     Downloading: 390 kB     Downloading: 390 kB     Downloading: 410 kB     Downloading: 410 kB     Downloading: 410 kB     Downloading: 410 kB     Downloading: 410 kB     Downloading: 410 kB     Downloading: 420 kB     Downloading: 420 kB     Downloading: 440 kB     Downloading: 440 kB     Downloading: 440 kB     Downloading: 440 kB     Downloading: 460 kB     Downloading: 460 kB     Downloading: 460 kB     Downloading: 460 kB     Downloading: 460 kB     Downloading: 460 kB     Downloading: 470 kB     Downloading: 470 kB     Downloading: 490 kB     Downloading: 490 kB     Downloading: 490 kB     Downloading: 490 kB     
#> ✔ Fetching JSON data from C2D API... done [205ms]
#> 
#> ℹ Converting JSON to list data...
#> ✔ Converting JSON to list data... done [22ms]
#> 
#> ℹ Tidying data...
#> ✔ Tidying data... done [659ms]
#> 
#> # A tibble: 129 × 8
#>    id                    elect…¹ votes…² votes…³ votes…⁴ votes…⁵ votes…⁶ turnout
#>    <chr>                   <int>   <int>   <int>   <int>   <int> <list>    <dbl>
#>  1 63208effa52c3995043a…  5.09e7  6.78e6  2.68e6  721157  239995 <NULL>    0.205
#>  2 63208da3a52c3995043a…  5.09e7  6.72e6  2.71e6  752871  246373 <NULL>    0.205
#>  3 63208be1a52c3995043a…  5.09e7  6.97e6  2.54e6  682995  232129 <NULL>    0.205
#>  4 63208a15a52c3995043a…  5.09e7  5.27e6  4.25e6  677146  226959 <NULL>    0.205
#>  5 632085cda52c3995043a…  5.09e7  5.09e6  4.50e6  616268  220564 <NULL>    0.205
#>  6 5f6c6d3fd1291cc3961f…  5.10e7  1.79e7  7.69e6  218093  226569 <NULL>    0.511
#>  7 5d1cc245dc0aeb467bfc…  7.90e6  2.88e6  1.19e5      NA   23151 <NULL>    0.383
#>  8 5d1cba4edc0aeb467bfc…  4.07e6  2.27e6  4.39e4    5165    5865 <NULL>    0.572
#>  9 5d19ffd1fe65c2bfeba0…  5.08e7  1.34e7  1.94e7   83418  308728 <NULL>    0.655
#> 10 5cb86b29cb4865239961…  5.07e7  1.33e7  2.20e6  104413  168753 <NULL>    0.312
#> # … with 119 more rows, and abbreviated variable names ¹​electorate_total,
#> #   ²​votes_yes, ³​votes_no, ⁴​votes_empty, ⁵​votes_invalid,
#> #   ⁶​votes_per_subterritory

# strict turnout numbers
c2d::rfrnds(country_code = "IT") |>
  c2d::add_turnout(rough = FALSE) |>
  dplyr::select(any_of("id"),
                electorate_total,
                starts_with("votes_"),
                turnout)
#> # A tibble: 129 × 8
#>    id                    elect…¹ votes…² votes…³ votes…⁴ votes…⁵ votes…⁶ turnout
#>    <chr>                   <int>   <int>   <int>   <int>   <int> <list>    <dbl>
#>  1 63208effa52c3995043a…  5.09e7  6.78e6  2.68e6  721157  239995 <NULL>    0.205
#>  2 63208da3a52c3995043a…  5.09e7  6.72e6  2.71e6  752871  246373 <NULL>    0.205
#>  3 63208be1a52c3995043a…  5.09e7  6.97e6  2.54e6  682995  232129 <NULL>    0.205
#>  4 63208a15a52c3995043a…  5.09e7  5.27e6  4.25e6  677146  226959 <NULL>    0.205
#>  5 632085cda52c3995043a…  5.09e7  5.09e6  4.50e6  616268  220564 <NULL>    0.205
#>  6 5f6c6d3fd1291cc3961f…  5.10e7  1.79e7  7.69e6  218093  226569 <NULL>    0.511
#>  7 5d1cc245dc0aeb467bfc…  7.90e6  2.88e6  1.19e5      NA   23151 <NULL>   NA    
#>  8 5d1cba4edc0aeb467bfc…  4.07e6  2.27e6  4.39e4    5165    5865 <NULL>    0.572
#>  9 5d19ffd1fe65c2bfeba0…  5.08e7  1.34e7  1.94e7   83418  308728 <NULL>    0.655
#> 10 5cb86b29cb4865239961…  5.07e7  1.33e7  2.20e6  104413  168753 <NULL>    0.312
#> # … with 119 more rows, and abbreviated variable names ¹​electorate_total,
#> #   ²​votes_yes, ³​votes_no, ⁴​votes_empty, ⁵​votes_invalid,
#> #   ⁶​votes_per_subterritory