Assemble MongoDB query filter document

Assemble MongoDB query filter document

assemble_query_filter(
  country_code = NULL,
  subnational_entity_name = NULL,
  municipality = NULL,
  level = NULL,
  type = NULL,
  date = NULL,
  is_draft = NULL,
  date_time_created_min = NULL,
  date_time_created_max = NULL,
  query_filter = NULL,
  base64_encode = TRUE
)

Arguments

country_code
The country_code(s) to be included. A character vector.

subnational_entity_name
The subnational_entity_name(s) to be included. A character vector.

municipality
The municipality(s) to be included. A character vector.

level
The level(s) to be included. A character vector.

type
The type(s) to be included. A character vector.

date
The date(s) to be included. A character vector.

is_draft
TRUE means to include only referendum entries with draft status, FALSE to include only normal entries. Set to NULL in order to include both draft and normal entries.

date_time_created_min
The minimum date_time_created to be included. A date, a datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05").

date_time_created_max
The maximum date_time_created to be included. A date, a datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05").

query_filter
A valid MongoDB JSON query filter document which allows for maximum control over what data is included. This takes precedence over all the other parameters, i.e. if query_filter is provided, the parameters country_code, subnational_entity_name, municipality, level, type, date, is_draft, date_time_created_min and date_time_created_max)`` are ignored.

base64_encode
Whether or not to Base64-encode the resulting query filter document. Note that the query_filter argument provided to other functions of this package must be Base64-encoded.

Value

A character scalar containing a valid MongoDB JSON query filter document, Base64-encoded if base64_encode = TRUE.