c2d: Download Data from the C2D Database, Which Covers Direct Democratic Votes Worldwide

c2d provides access to the C2D Database (https://c2d.ch) from R. This database aims to record all direct democratic votes worldwide and is operated by the Centre for Democracy Studies Aarau (ZDA) at the University of Zurich, Switzerland.

See also

Author

Maintainer: Salim Brüggemann salim-b@pm.me (ORCID)

Authors:

Uwe Serdült uwe.serduelt@zda.uzh.ch (ORCID)

Other contributors:

Centre for Democracy Studies Aarau (ZDA) info@zdaarau.ch [copyright holder]