Count number of referendums

Counts the number of referendums per level in the C2D Database.

count_referendums(
  is_draft = FALSE,
  country_code = NULL,
  subnational_entity_name = NULL,
  municipality = NULL,
  level = NULL,
  type = NULL,
  date = NULL,
  date_time_created_min = NULL,
  date_time_created_max = NULL,
  query_filter = NULL
)

Arguments

is_draft
TRUE means to include only referendum entries with draft status, FALSE to include only normal entries. Set to NULL in order to include both draft and normal entries.

country_code
The country_code(s) to be included. A character vector.

subnational_entity_name
The subnational_entity_name(s) to be included. A character vector.

municipality
The municipality(s) to be included. A character vector.

level
The level(s) to be included. A character vector.

type
The type(s) to be included. A character vector.

date
The date(s) to be included. A character vector.

date_time_created_min
The minimum date_time_created to be included. A date, a datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05").

date_time_created_max
The maximum date_time_created to be included. A date, a datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05").

query_filter
A valid MongoDB JSON query filter document which allows for maximum control over what data is included. This takes precedence over all of the above listed parameters, i.e. if query_filter is provided, the parameters country_code, subnational_entity_name, municipality, level, type, date, is_draft, date_time_created_min and date_time_created_max are ignored.

Value

A named list with level as names and referendum counts as values.

See also

Examples

# the whole database (excl. drafts)
c2d::count_referendums()
#> $local
#> [1] 45
#> 
#> $subnational
#> [1] 14516
#> 
#> $national
#> [1] 3017
#> 

# only Swiss and Austrian referendums
c2d::count_referendums(country_code = c("CH", "AT"))
#> $local
#> [1] 1
#> 
#> $subnational
#> [1] 6722
#> 
#> $national
#> [1] 671
#> 

# only Swiss referendums created between 2020 and 2021
c2d::count_referendums(country_code = "CH",
                       date_time_created_min = "2020-01-01",
                       date_time_created_max = "2021-01-01")
#> $local
#> [1] 0
#> 
#> $subnational
#> [1] 104
#> 
#> $national
#> [1] 9
#>