Download file attachment

Downloads a file attachment from the C2D database. The necessary s3_object_keys identifying individual files are found in the files list column returned by referendums().

download_file_attachment(s3_object_key, path = ".")

Arguments

s3_object_key
The key uniquely identifying the file in the C2D Amazon S3 bucket. A character scalar.

path
The path where the downloaded file is written to. If a directory, the original filename returned by the C2D services API will be used.

Value

A response object, invisibly.

See also

Examples

if (FALSE) {
# get all file object keys...
c2d::referendums()$files |>
  purrr::map_depth(2L, purrr::pluck, "s3_object_key") |>
  purrr::flatten() |>
  purrr::flatten_chr() |>
  # ...select first three keys...
  magrittr::extract(1:3) |>
  # ...and download the corresponding files to the current working dir
  purrr::walk(c2d::download_file_attachment)}