Edit existing referendums in the C2D database

Edits existing referendum entries in the C2D database via its API.

edit_rfrnds(
  data,
  email = pal::pkg_config_val(key = "api_username", pkg = this_pkg),
  password = pal::pkg_config_val(key = "api_password", pkg = this_pkg),
  use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg)
)

Arguments

data
The updated referendum data. A tibble that must contain an id column identifying the referendums to be edited, an is_draft column setting their updated draft status, plus a valid column containing the new values for each additional database field that should be updated.

email
The e-mail address of the C2D API user account to be used for authentication. A character scalar.

password
The password of the C2D API user account to be used for authentication. A character scalar.

use_testing_server
Whether or not to use the testing servers instead of the production servers for C2D Services API calls etc.

Value

data, invisibly.

Details

Note that the following variables are not supported, i.e. simply dropped from data:

id_official

id_sudd

files

See also