Hierarchize tags

Reconstructs the hierarchical relations between the three tag variables tags_tier_1, tags_tier_2 and tags_tier_3. Can also be used to simply determine the parent tag(s) of any tag.

hierarchize_tags(x)

Arguments

x
The tags to hierarchize. Either a character vector of tags or a single-row data frame containing at least the columns tags_tier_1, tags_tier_2 and tags_tier_3.

Value

A tibble.

See also

Examples

c2d::hierarchize_tags("territorial questions")
#> # A tibble: 1 × 3
#>   tag_tier_1         tag_tier_2 tag_tier_3           
#>   <chr>              <chr>      <chr>                
#> 1 state organisation federalism territorial questions

# hierarchize the tags of all Austrian referendums
c2d::rfrnds(country_code = "AT") |>
  dplyr::group_split(id) |>
  purrr::map(c2d::hierarchize_tags)
#> [[1]]
#> # A tibble: 2 × 3
#>   tag_tier_1         tag_tier_2        tag_tier_3           
#>   <chr>              <chr>             <chr>                
#> 1 state organisation federalism        territorial questions
#> 2 state organisation national identity NA                   
#> 
#> [[2]]
#> # A tibble: 0 × 3
#> # … with 3 variables: tag_tier_1 <chr>, tag_tier_2 <chr>, tag_tier_3 <chr>
#> 
#> [[3]]
#> # A tibble: 1 × 3
#>   tag_tier_1         tag_tier_2        tag_tier_3
#>   <chr>              <chr>             <chr>     
#> 1 state organisation national identity NA        
#> 
#> [[4]]
#> # A tibble: 1 × 3
#>   tag_tier_1         tag_tier_2        tag_tier_3
#>   <chr>              <chr>             <chr>     
#> 1 state organisation national identity NA        
#> 
#> [[5]]
#> # A tibble: 2 × 3
#>   tag_tier_1         tag_tier_2        tag_tier_3           
#>   <chr>              <chr>             <chr>                
#> 1 state organisation federalism        territorial questions
#> 2 state organisation national identity NA                   
#> 
#> [[6]]
#> # A tibble: 2 × 3
#>   tag_tier_1         tag_tier_2        tag_tier_3           
#>   <chr>              <chr>             <chr>                
#> 1 state organisation federalism        territorial questions
#> 2 state organisation national identity NA                   
#> 
#> [[7]]
#> # A tibble: 3 × 3
#>   tag_tier_1         tag_tier_2        tag_tier_3           
#>   <chr>              <chr>             <chr>                
#> 1 state organisation federalism        territorial questions
#> 2 state organisation national identity NA                   
#> 3 state organisation political system  NA                   
#> 
#> [[8]]
#> # A tibble: 1 × 3
#>   tag_tier_1 tag_tier_2     tag_tier_3
#>   <chr>      <chr>          <chr>     
#> 1 energy     nuclear energy NA        
#> 
#> [[9]]
#> # A tibble: 1 × 3
#>   tag_tier_1     tag_tier_2      tag_tier_3
#>   <chr>          <chr>           <chr>     
#> 1 foreign policy European policy EU        
#> 
#> [[10]]
#> # A tibble: 1 × 3
#>   tag_tier_1      tag_tier_2 tag_tier_3           
#>   <chr>           <chr>      <chr>                
#> 1 security policy army       military organisation
#> 
#> [[11]]
#> # A tibble: 1 × 3
#>   tag_tier_1      tag_tier_2 tag_tier_3           
#>   <chr>           <chr>      <chr>                
#> 1 security policy army       military organisation
#>