Hierarchize tags (fast)

Reconstructs the hierarchical relations between the three tag variables tags_tier_1, tags_tier_2 and tags_tier_3. Other than hierarchize_tags(), this function assumes that the three tag variables are always complete, i.e. that no (grand)parent tags of lower-tier tags are missing. This assumption is met by rfrnds() and rfrnd().

hierarchize_tags_fast(
  tags_tier_1 = character(),
  tags_tier_2 = character(),
  tags_tier_3 = character()
)

Arguments

tags_tier_1
First-tier tags. A character vector.

tags_tier_2
Second-tier tags. A character vector.

tags_tier_3
Third-tier tags. A character vector.

Value

A tibble with the columns tag_tier_1, tag_tier_2 and tag_tier_3.

See also

Examples

library(magrittr)

c2d::rfrnd(id = "5bbbe26a92a21351232dd73f") %$%
  c2d:::hierarchize_tags_fast(unlist(tags_tier_1),
                              unlist(tags_tier_2),
                              unlist(tags_tier_3))
#> # A tibble: 3 × 3
#>   tag_tier_1         tag_tier_2   tag_tier_3                 
#>   <chr>              <chr>        <chr>                      
#> 1 state organisation federalism   intergovernmental relations
#> 2 state organisation legal system fundamental rights         
#> 3 public finance     taxation     tax system                 

# hierarchize the tags of all Austrian referendums
c2d::rfrnds(country_code = "AT") |>
  dplyr::group_split(id) |>
  purrr::map(~ c2d:::hierarchize_tags_fast(unlist(.x$tags_tier_1),
                                           unlist(.x$tags_tier_2),
                                           unlist(.x$tags_tier_3)))
#> [[1]]
#> # A tibble: 2 × 3
#>   tag_tier_1         tag_tier_2        tag_tier_3           
#>   <chr>              <chr>             <chr>                
#> 1 state organisation federalism        territorial questions
#> 2 state organisation national identity NA                   
#> 
#> [[2]]
#> # A tibble: 0 × 3
#> # … with 3 variables: tag_tier_1 <chr>, tag_tier_2 <chr>, tag_tier_3 <chr>
#> # ℹ Use `colnames()` to see all variable names
#> 
#> [[3]]
#> # A tibble: 1 × 3
#>   tag_tier_1         tag_tier_2        tag_tier_3
#>   <chr>              <chr>             <chr>     
#> 1 state organisation national identity NA        
#> 
#> [[4]]
#> # A tibble: 1 × 3
#>   tag_tier_1         tag_tier_2        tag_tier_3
#>   <chr>              <chr>             <chr>     
#> 1 state organisation national identity NA        
#> 
#> [[5]]
#> # A tibble: 2 × 3
#>   tag_tier_1         tag_tier_2        tag_tier_3           
#>   <chr>              <chr>             <chr>                
#> 1 state organisation federalism        territorial questions
#> 2 state organisation national identity NA                   
#> 
#> [[6]]
#> # A tibble: 2 × 3
#>   tag_tier_1         tag_tier_2        tag_tier_3           
#>   <chr>              <chr>             <chr>                
#> 1 state organisation federalism        territorial questions
#> 2 state organisation national identity NA                   
#> 
#> [[7]]
#> # A tibble: 3 × 3
#>   tag_tier_1         tag_tier_2        tag_tier_3           
#>   <chr>              <chr>             <chr>                
#> 1 state organisation federalism        territorial questions
#> 2 state organisation national identity NA                   
#> 3 state organisation political system  NA                   
#> 
#> [[8]]
#> # A tibble: 1 × 3
#>   tag_tier_1 tag_tier_2     tag_tier_3
#>   <chr>      <chr>          <chr>     
#> 1 energy     nuclear energy NA        
#> 
#> [[9]]
#> # A tibble: 1 × 3
#>   tag_tier_1     tag_tier_2      tag_tier_3
#>   <chr>          <chr>           <chr>     
#> 1 foreign policy European policy EU        
#> 
#> [[10]]
#> # A tibble: 1 × 3
#>   tag_tier_1      tag_tier_2 tag_tier_3           
#>   <chr>           <chr>      <chr>                
#> 1 security policy army       military organisation
#> 
#> [[11]]
#> # A tibble: 1 × 3
#>   tag_tier_1      tag_tier_2 tag_tier_3           
#>   <chr>           <chr>      <chr>                
#> 1 security policy army       military organisation
#>