Hierarchize topics

Reconstructs the hierarchical relations between the three topic variables topics_tier_1, topics_tier_2 and topics_tier_3. Can also be used to simply determine the parent topic(s) of any topic.

hierarchize_topics(x)

Arguments

x

The topics to hierarchize. Either a character vector of topics or a single-row data frame containing at least the columns topics_tier_1, topics_tier_2 and topics_tier_3.

Value

A tibble with the columns topic_tier_1, topic_tier_2 and topic_tier_3.

See also

Examples

c2d::hierarchize_topics("territorial questions")
#> # A tibble: 1 × 3
#>   topic_tier_1       topic_tier_2 topic_tier_3         
#>   <chr>              <chr>        <chr>                
#> 1 state organisation federalism   territorial questions

# hierarchize the topics of all Austrian referendums
c2d::rfrnds(country_code = "AT") |>
  dplyr::group_split(id) |>
  purrr::map(c2d::hierarchize_topics)
#> [[1]]
#> # A tibble: 2 × 3
#>   topic_tier_1       topic_tier_2      topic_tier_3         
#>   <chr>              <chr>             <chr>                
#> 1 state organisation federalism        territorial questions
#> 2 state organisation national identity NA                   
#> 
#> [[2]]
#> # A tibble: 0 × 3
#> # … with 3 variables: topic_tier_1 <chr>, topic_tier_2 <chr>,
#> #   topic_tier_3 <chr>
#> 
#> [[3]]
#> # A tibble: 1 × 3
#>   topic_tier_1       topic_tier_2      topic_tier_3
#>   <chr>              <chr>             <chr>       
#> 1 state organisation national identity NA          
#> 
#> [[4]]
#> # A tibble: 1 × 3
#>   topic_tier_1       topic_tier_2      topic_tier_3
#>   <chr>              <chr>             <chr>       
#> 1 state organisation national identity NA          
#> 
#> [[5]]
#> # A tibble: 2 × 3
#>   topic_tier_1       topic_tier_2      topic_tier_3         
#>   <chr>              <chr>             <chr>                
#> 1 state organisation federalism        territorial questions
#> 2 state organisation national identity NA                   
#> 
#> [[6]]
#> # A tibble: 2 × 3
#>   topic_tier_1       topic_tier_2      topic_tier_3         
#>   <chr>              <chr>             <chr>                
#> 1 state organisation federalism        territorial questions
#> 2 state organisation national identity NA                   
#> 
#> [[7]]
#> # A tibble: 3 × 3
#>   topic_tier_1       topic_tier_2      topic_tier_3         
#>   <chr>              <chr>             <chr>                
#> 1 state organisation federalism        territorial questions
#> 2 state organisation national identity NA                   
#> 3 state organisation political system  NA                   
#> 
#> [[8]]
#> # A tibble: 1 × 3
#>   topic_tier_1 topic_tier_2   topic_tier_3
#>   <chr>        <chr>          <chr>       
#> 1 energy       nuclear energy NA          
#> 
#> [[9]]
#> # A tibble: 1 × 3
#>   topic_tier_1   topic_tier_2    topic_tier_3
#>   <chr>          <chr>           <chr>       
#> 1 foreign policy European policy EU          
#> 
#> [[10]]
#> # A tibble: 1 × 3
#>   topic_tier_1    topic_tier_2 topic_tier_3         
#>   <chr>           <chr>        <chr>                
#> 1 security policy army         military organisation
#> 
#> [[11]]
#> # A tibble: 1 × 3
#>   topic_tier_1    topic_tier_2 topic_tier_3         
#>   <chr>           <chr>        <chr>                
#> 1 security policy army         military organisation
#>