List referendum data from sudd.ch

Lists the referendum data from sudd.ch in various ways its list.php endpoint allows. The output of this function can be directly fed to sudd_referendums().

list_sudd_referendums(
  mode = c("by_date", "by_mod_date", "filter", "random"),
  order = c("ascending", "descending"),
  filter = list(territory_name_de = NULL, title_de = NULL, year_min = NULL, year_max =
    NULL)
)

Arguments

mode
The listing mode. One of

"by_date": Lists all referendums in the sudd.ch database by date, together with their id_sudd, country_code, territory_name_de and title_de. Specifying the sorting order of the results via the order parameter is supported.

"by_mod_date": Lists all referendums in the sudd.ch database by date_last_edited, together with their id_sudd, country_code, territory_name_de, date and title_de.

"filter": Allows to provide additional arguments (via the filter parameter) that limit the results accordingly. Specifying the sorting order of the results via the order parameter is supported.

"random": Lists the id_sudd of five randomly selected referendums, together with their country_code.

order
The sorting order of the results. Only relevant if mode is either "by_date" or "filter".

filter
A list with valid filtering arguments. Only relevant if mode = "filter". The supported filtering arguments include

territory_name_de: A string that must be (partially) matched by the referendums' territory_name_de. Matching is case-insensitive and no fuzzy matching is supported.

title_de: A string that must be (partially) matched by the referendums' title_de. Matching is case-insensitive and no fuzzy matching is supported.

year_min: The lower year limit of the referendums' date. A positive integer.

year_max: The upper year limit of the referendums' date. A positive integer.

Value

A tibble containing at least an id_sudd column.

About sudd.ch

sudd stands for Suchmaschine für direkte Demokratie (German) and is operated by Beat Müller. Its database content is licensed under Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International (CC BY-NC-SA 4.0).

See also

Examples

# list all referendums by modification date
c2d::list_sudd_referendums(mode = "by_mod_date")

#> # A tibble: 3,045 × 10
#>    id_sudd  country_code country_name_de territory_name_… date        year month
#>    <chr>    <chr>        <chr>           <chr>            <date>     <int> <int>
#>  1 fr021793 FR           Frankreich      Rheinbecken (Rh… NA          1793     3
#>  2 fr031798 FR           Frankreich      Cisrhenanische … NA          1798     6
#>  3 gu011946 GU           Guam            Guam (USA)       1946-06-08  1946     6
#>  4 gu011982 GU           Guam            Guam (USA)       1982-01-30  1982     1
#>  5 gu021982 GU           Guam            Guam (USA)       1982-09-04  1982     9
#>  6 lv012008 LV           Lettland        Lettland         2008-08-02  2008     8
#>  7 lv012012 LV           Lettland        Lettland         2012-02-18  2012     2
#>  8 as011966 AS           Samoa (US)      Amerikanisch-Sa… 1966-11-08  1966    11
#>  9 as011970 AS           Samoa (US)      Amerikanisch-Sa… 1970-11-03  1970    11
#> 10 as021970 AS           Samoa (US)      Amerikanisch-Sa… 1970-11-03  1970    11
#> # … with 3,035 more rows, and 3 more variables: day <int>, title_de <chr>,
#> #   date_last_edited <date>

# list all referendums whose title matches "AHV"
c2d::list_sudd_referendums(mode = "filter",
                           filter = list(title_de = "AHV"))

#> # A tibble: 15 × 9
#>    id_sudd  country_code country_name_de territory_name_… date        year month
#>    <chr>    <chr>        <chr>           <chr>            <date>     <int> <int>
#>  1 ch021978 CH           Schweiz         Schweiz          1978-02-26  1978     2
#>  2 ch031978 CH           Schweiz         Schweiz          1978-02-26  1978     2
#>  3 ch021988 CH           Schweiz         Schweiz          1988-06-12  1988     6
#>  4 ch051995 CH           Schweiz         Schweiz          1995-06-25  1995     6
#>  5 ch061995 CH           Schweiz         Schweiz          1995-06-25  1995     6
#>  6 ch061998 CH           Schweiz         Schweiz          1998-09-27  1998     9
#>  7 ch082001 CH           Schweiz         Schweiz          2001-12-02  2001    12
#>  8 ch052002 CH           Schweiz         Schweiz          2002-09-22  2002     9
#>  9 ch042004 CH           Schweiz         Schweiz          2004-05-16  2004     5
#> 10 ch052004 CH           Schweiz         Schweiz          2004-05-16  2004     5
#> 11 ch022006 CH           Schweiz         Schweiz          2006-09-24  2006     9
#> 12 ch072008 CH           Schweiz         Schweiz          2008-11-30  2008    11
#> 13 ch112016 CH           Schweiz         Schweiz          2016-09-25  2016     9
#> 14 ch062017 CH           Schweiz         Schweiz          2017-09-24  2017     9
#> 15 ch022019 CH           Schweiz         Schweiz          2019-05-19  2019     5
#> # … with 2 more variables: day <int>, title_de <chr>

# get sudd.ch referendum data from all referendums from 2020 onwards
c2d::list_sudd_referendums(mode = "filter",
                           filter = list(year_min = 2020)) |>
  c2d::sudd_referendums()

#> # A tibble: 73 × 41
#>    id_sudd  country_code country_name_de territory_name_de territory_type_de 
#>    <chr>    <chr>        <chr>           <chr>             <chr>             
#>  1 ch012020 CH           Schweiz         Schweiz           unabhängiger Staat
#>  2 ch022020 CH           Schweiz         Schweiz           unabhängiger Staat
#>  3 fr012020 FR           Frankreich      Frankreich        unabhängiger Staat
#>  4 gn012020 GN           Guinea          Guinea            unabhängiger Staat
#>  5 pw012020 PW           Palau           Palau             unabhängiger Staat
#>  6 ru012020 RU           Russland        Russland          unabhängiger Staat
#>  7 ru022020 RU           Russland        Russland          unabhängiger Staat
#>  8 ru032020 RU           Russland        Russland          unabhängiger Staat
#>  9 ru042020 RU           Russland        Russland          unabhängiger Staat
#> 10 ru052020 RU           Russland        Russland          unabhängiger Staat
#> # … with 63 more rows, and 36 more variables: date <date>, year <int>,
#> #   month <int>, day <int>, title_de <chr>, question_type_de <chr>,
#> #   types <list>, result_de <chr>, adoption_requirements_de <list>,
#> #   electorate_total <int>, electorate_abroad <int>, polling_cards <int>,
#> #   votes_total <int>, votes_empty <int>, votes_void <int>,
#> #   votes_invalid <int>, votes_valid <int>, votes_yes <int>, votes_no <int>,
#> #   states <dbl>, states_yes <dbl>, states_no <dbl>, vote_channel_de <chr>, …