List referendum data from sudd.ch

Lists the referendum data from sudd.ch in various ways its list.php endpoint allows. The output of this function can be directly fed to sudd_rfrnds().

list_sudd_rfrnds(
  mode = c("by_date", "by_mod_date", "filter", "random"),
  order = c("ascending", "descending"),
  filter = list(territory_name_de = NULL, title_de = NULL, year_min = NULL, year_max =
    NULL),
  use_cache = TRUE,
  cache_lifespan = "1 week"
)

Arguments

mode
The listing mode. One of

"by_date": Lists all referendums in the sudd.ch database by date, together with their id_sudd, country_code, territory_name_de and title_de. Specifying the sorting order of the results via the order parameter is supported.

"by_mod_date": Lists all referendums in the sudd.ch database by date_last_edited, together with their id_sudd, country_code, territory_name_de, date and title_de.

"filter": Allows to provide additional arguments (via the filter parameter) that limit the results accordingly. Specifying the sorting order of the results via the order parameter is supported.

"random": Lists the id_sudd of five randomly selected referendums, together with their country_code.

order
The sorting order of the results. Only relevant if mode is either "by_date" or "filter".

filter
A list with valid filtering arguments. Only relevant if mode = "filter". The supported filtering arguments include

territory_name_de: A string that must be (partially) matched by the referendums' territory_name_de. Matching is case-insensitive and no fuzzy matching is supported.

title_de: A string that must be (partially) matched by the referendums' title_de. Matching is case-insensitive and no fuzzy matching is supported.

year_min: The lower year limit of the referendums' date. A positive integer.

year_max: The upper year limit of the referendums' date. A positive integer.

use_cache
Return cached results if possible. If FALSE, results are always newly fetched regardless of cache_lifespan.

cache_lifespan
The duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Only relevant if use_cache = TRUE.

Value

A tibble containing at least an id_sudd column.

About sudd.ch

sudd stands for Suchmaschine für direkte Demokratie (German) and is operated by Beat Müller. Its database content is licensed under Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International (CC BY-NC-SA 4.0).

See also

Examples

# list all referendums by modification date
c2d::list_sudd_rfrnds(mode = "by_mod_date")
#> # A tibble: 3,070 × 12
#>    id_sudd  count…¹ count…² count…³ is_pa…⁴ terri…⁵ date        year month   day
#>    <chr>    <chr>   <chr>   <chr>   <lgl>   <chr>   <date>     <int> <int> <int>
#>  1 ck012022 CK      Cook I… NA      FALSE   Cook-I… 2022-08-01  2022     8     1
#>  2 cl012022 CL      Chile   NA      FALSE   Chile   2022-09-04  2022     9     4
#>  3 nu011992 NU      Niue    NA      FALSE   Niue    1992-06-13  1992     6    13
#>  4 tn012022 TN      Tunisia NA      FALSE   Tunesi… 2022-07-25  2022     7    25
#>  5 li012022 LI      Liecht… NA      FALSE   Liecht… 2022-06-26  2022     6    26
#>  6 it022022 IT      Italy   NA      FALSE   Italien 2022-06-12  2022     6    12
#>  7 it032022 IT      Italy   NA      FALSE   Italien 2022-06-12  2022     6    12
#>  8 it042022 IT      Italy   NA      FALSE   Italien 2022-06-12  2022     6    12
#>  9 it052022 IT      Italy   NA      FALSE   Italien 2022-06-12  2022     6    12
#> 10 it062022 IT      Italy   NA      FALSE   Italien 2022-06-12  2022     6    12
#> # … with 3,060 more rows, 2 more variables: title_de <chr>,
#> #   date_last_edited <date>, and abbreviated variable names ¹​country_code,
#> #   ²​country_name, ³​country_code_historical, ⁴​is_past_jurisdiction,
#> #   ⁵​territory_name_de
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names

# list all referendums whose title matches "AHV"
c2d::list_sudd_rfrnds(mode = "filter",
                      filter = list(title_de = "AHV"))
#> # A tibble: 15 × 11
#>    id_sudd  count…¹ count…² count…³ is_pa…⁴ terri…⁵ date        year month   day
#>    <chr>    <chr>   <chr>   <chr>   <lgl>   <chr>   <date>     <int> <int> <int>
#>  1 ch021978 CH      Switze… NA      FALSE   Schweiz 1978-02-26  1978     2    26
#>  2 ch031978 CH      Switze… NA      FALSE   Schweiz 1978-02-26  1978     2    26
#>  3 ch021988 CH      Switze… NA      FALSE   Schweiz 1988-06-12  1988     6    12
#>  4 ch051995 CH      Switze… NA      FALSE   Schweiz 1995-06-25  1995     6    25
#>  5 ch061995 CH      Switze… NA      FALSE   Schweiz 1995-06-25  1995     6    25
#>  6 ch061998 CH      Switze… NA      FALSE   Schweiz 1998-09-27  1998     9    27
#>  7 ch082001 CH      Switze… NA      FALSE   Schweiz 2001-12-02  2001    12     2
#>  8 ch052002 CH      Switze… NA      FALSE   Schweiz 2002-09-22  2002     9    22
#>  9 ch042004 CH      Switze… NA      FALSE   Schweiz 2004-05-16  2004     5    16
#> 10 ch052004 CH      Switze… NA      FALSE   Schweiz 2004-05-16  2004     5    16
#> 11 ch022006 CH      Switze… NA      FALSE   Schweiz 2006-09-24  2006     9    24
#> 12 ch072008 CH      Switze… NA      FALSE   Schweiz 2008-11-30  2008    11    30
#> 13 ch112016 CH      Switze… NA      FALSE   Schweiz 2016-09-25  2016     9    25
#> 14 ch062017 CH      Switze… NA      FALSE   Schweiz 2017-09-24  2017     9    24
#> 15 ch022019 CH      Switze… NA      FALSE   Schweiz 2019-05-19  2019     5    19
#> # … with 1 more variable: title_de <chr>, and abbreviated variable names
#> #   ¹​country_code, ²​country_name, ³​country_code_historical,
#> #   ⁴​is_past_jurisdiction, ⁵​territory_name_de
#> # ℹ Use `colnames()` to see all variable names

# get sudd.ch referendum data from all referendums from 2020 onwards
c2d::list_sudd_rfrnds(mode = "filter",
                      filter = list(year_min = 2020)) |>
  c2d::sudd_rfrnds()
#> # A tibble: 122 × 37
#>    id_sudd  count…¹ count…² count…³ terri…⁴ terri…⁵ date        year month   day
#>    <chr>    <chr>   <chr>   <chr>   <chr>   <chr>   <date>     <int> <int> <int>
#>  1 ch012020 CH      NA      Switze… Schweiz unabhä… 2020-02-09  2020     2     9
#>  2 ch022020 CH      NA      Switze… Schweiz unabhä… 2020-02-09  2020     2     9
#>  3 fr012020 FR      NA      France  Frankr… unabhä… 2020-03-15  2020     3    15
#>  4 gn012020 GN      NA      Guinea  Guinea  unabhä… 2020-03-22  2020     3    22
#>  5 pw012020 PW      NA      Palau   Palau   unabhä… 2020-05-01  2020     5     1
#>  6 ru012020 RU      NA      Russia… Russla… unabhä… 2020-07-01  2020     7     1
#>  7 ru022020 RU      NA      Russia… Russla… unabhä… 2020-07-01  2020     7     1
#>  8 ru032020 RU      NA      Russia… Russla… unabhä… 2020-07-01  2020     7     1
#>  9 ru042020 RU      NA      Russia… Russla… unabhä… 2020-07-01  2020     7     1
#> 10 ru052020 RU      NA      Russia… Russla… unabhä… 2020-07-01  2020     7     1
#> # … with 112 more rows, 27 more variables: title_de <chr>,
#> #   question_type_de <chr>, types <list>, result_de <chr>,
#> #   result_status_de <chr>, adoption_requirements_de <list>,
#> #   electorate_total <int>, electorate_abroad <int>, polling_cards <int>,
#> #   votes_total <int>, votes_empty <int>, votes_void <int>,
#> #   votes_invalid <int>, votes_valid <int>, votes_yes <int>, votes_no <int>,
#> #   votes_option_1 <int>, votes_option_2 <int>, votes_option_none <int>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names

# get sudd.ch referendum data from five randomly picked referendums
c2d::list_sudd_rfrnds(mode = "random") |>
  c2d::sudd_rfrnds()
#> ℹ Fetching raw HTML data from <sudd.ch>...
#> ✔ Fetching raw HTML data from <sudd.ch>... done [381ms]
#> 
#> ℹ Parsing and tidying raw HTML data...
#> ✔ Parsing and tidying raw HTML data... done [102ms]
#> 
#> # A tibble: 5 × 40
#>   id_sudd  countr…¹ count…² count…³ terri…⁴ terri…⁵ date        year month   day
#>   <chr>    <chr>    <chr>   <chr>   <chr>   <chr>   <date>     <int> <int> <int>
#> 1 ch011906 CH       NA      Switze… Schweiz unabhä… 1906-06-10  1906     6    10
#> 2 ch091985 CH       NA      Switze… Schweiz unabhä… 1985-09-22  1985     9    22
#> 3 ch021989 CH       NA      Switze… Schweiz unabhä… 1989-11-26  1989    11    26
#> 4 si012011 SI       NA      Sloven… Slowen… unabhä… 2011-04-10  2011     4    10
#> 5 gg012018 GG       NA      Guerns… Guerns… abhäng… 2018-10-10  2018    10    10
#> # … with 30 more variables: title_de <chr>, question_type_de <chr>,
#> #   types <list>, result_de <chr>, result_status_de <chr>,
#> #   adoption_requirements_de <list>, electorate_total <int>,
#> #   electorate_abroad <int>, polling_cards <int>, votes_total <int>,
#> #   votes_empty <int>, votes_void <int>, votes_invalid <int>,
#> #   votes_valid <int>, votes_yes <int>, votes_no <int>, votes_option_1 <int>,
#> #   votes_option_2 <int>, votes_option_3 <int>, votes_option_4 <int>, …
#> # ℹ Use `colnames()` to see all variable names