List referendum territories from sudd.ch

Lists all referendum territories from sudd.ch, which means each country_name_de together with all the associated territory_name_de, their search URL and their number of occurrences.

list_sudd_territories()

Value

A tibble.

Details

Note that the values in the territory_name_de column returned by this function can differ from those in the territory_name_de column of sudd_rfrnds() and list_sudd_rfrnds(). The latter is often more extensive and usually includes the country_name_de (in parentheses) for subnational referendums.

About sudd.ch

sudd stands for Suchmaschine für direkte Demokratie (German) and is operated by Beat Müller. Its database content is licensed under Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International (CC BY-NC-SA 4.0).

See also

Examples

c2d::list_sudd_territories()
#> # A tibble: 427 × 4
#>    country_name_de     territory_name_de    filter_url                     n    
#>    <chr>               <chr>                <chr>                          <chr>
#>  1 Albanien            Albanien             https://sudd.ch/list.php?de=d… 3    
#>  2 Algerien            Algerien             https://sudd.ch/list.php?de=d… 13   
#>  3 Andorra             Andorra              https://sudd.ch/list.php?de=d… 5    
#>  4 Anguilla            Anguilla             https://sudd.ch/list.php?de=d… 2    
#>  5 Antigua und Barbuda Antigua und Barbuda  https://sudd.ch/list.php?de=d… 1    
#>  6 Argentinien         Argentinien          https://sudd.ch/list.php?de=d… 2    
#>  7 Argentinien         Provinz Buenos Aires https://sudd.ch/list.php?de=d… 1    
#>  8 Argentinien         Provinz Corrientes   https://sudd.ch/list.php?de=d… 1    
#>  9 Argentinien         Valle 16 de Octubre  https://sudd.ch/list.php?de=d… 1    
#> 10 Armenien            Armenien             https://sudd.ch/list.php?de=d… 5    
#> # … with 417 more rows