Tag segmentation sunburst chart

Creates a Plotly sunburst chart that visualizes the hierarchical segmentation of referendum tag occurences.

plot_tag_segmentation(
  data,
  method = c("per_rfrnd", "per_tag_lineage", "naive")
)

Arguments

data
C2D referendum data as returned by rfrnds(). A data frame that at minimum contains the columns tags_tier_1, tags_tier_2 and tags_tier_3.

method
Applied method to count the number of tag occurences. One of

"per_rfrnd": All referendums have the same weight. For a referendum with n different tags of the same tier, every tag is counted 1/n.

"per_tag_lineage": All tag lineages have the same weight. For a referendum with n different tags of the same tier, every tag is fully counted, meaning that e.g. a referendum with three different tier-3 tags has a tripled impact on the result compared to a referendum that only has a single one. Noticeably faster than "per_rfrnd".

"naive": Naive procedure which doesn't properly reflect tag proportions on tier 2 and 3. Based on the (wrong) assumptions that a) all referendums have the same number of tag lineages assigned and b) tags are not deduplicated per tier. By far the fastest method, though.

Value

A plotly object.

Details

A tag lineage is the hierarchical compound of a tag_tier_1 and optionally a grandchild tag_tier_3 and/or a child tag_tier_2.

Note that tags can be assigned on any tier to referendums (i.e. in one case, a tag_tier_1 plus a child tag_tier_2 is assigned, and in another case only a tag_tier_1 without any further child tag).

Furthermore, it should be noted that not every tag_tier_2 has potential child tag_tier_3s. See the full tag hierarchy for details.

See also

Examples

# count each referendum equally
c2d::rfrnds(country_code = "AT") |> c2d::plot_tag_segmentation(method = "per_rfrnd")

# count each tag lineage equally
c2d::rfrnds(country_code = "AT") |> c2d::plot_tag_segmentation(method = "per_tag_lineage")

# naive count (way faster, but with misleading proportions on tier 2 and 3)
c2d::rfrnds(country_code = "AT") |> c2d::plot_tag_segmentation(method = "naive")