Printify referendum data column names

Renames referendum data column names to be ready to print. Useful e.g. to create publication tables.

printify_col_names(data)

Arguments

data

C2D referendum data as returned by rfrnds().

Value

A tibble.

Details

Column names unknown to this function are not changed.

See also

Examples

c2d::rfrnds(country_code = "IT") |> c2d::printify_col_names()
#> # A tibble: 129 × 69
#>    `C2D ID`   Offic…¹ sudd.…² ISO 3…³ Country Subna…⁴ Munic…⁵ Polit…⁶ Date      
#>    <chr>      <chr>   <chr>   <fct>   <fct>   <chr>   <chr>   <ord>   <date>    
#>  1 63208effa… NA      it0620… IT      Italy   NA      NA      nation… 2022-06-12
#>  2 63208da3a… NA      it0520… IT      Italy   NA      NA      nation… 2022-06-12
#>  3 63208be1a… NA      it0420… IT      Italy   NA      NA      nation… 2022-06-12
#>  4 63208a15a… NA      it0320… IT      Italy   NA      NA      nation… 2022-06-12
#>  5 632085cda… NA      it0220… IT      Italy   NA      NA      nation… 2022-06-12
#>  6 5f6c6d3fd… NA      NA      IT      Italy   NA      NA      nation… 2020-09-21
#>  7 5d1cc245d… NA      NA      IT      Italy   Lombar… NA      subnat… 2017-10-22
#>  8 5d1cba4ed… NA      NA      IT      Italy   Venetia NA      subnat… 2017-10-22
#>  9 5d19ffd1f… NA      NA      IT      Italy   NA      NA      nation… 2016-12-04
#> 10 5cb86b29c… NA      NA      IT      Italy   NA      NA      nation… 2016-03-15
#> # … with 119 more rows, 60 more variables: `Former country?` <lgl>,
#> #   Title <chr>, Titel <chr>, Titre <chr>, `Original question` <chr>,
#> #   Question <chr>, `Committee name` <chr>, Result <fct>,
#> #   `Approving subterritories` <dbl>, `Rejecting subterritories` <dbl>,
#> #   `Total electorate` <int>, `Electorate abroad` <int>, `Yes votes` <int>,
#> #   `No votes` <int>, `Empty votes` <int>, `Invalid votes` <int>,
#> #   `Subterritorial votes` <list>, `Lower house yes votes` <int>, …