Get a single referendum's data

Downloads a single referendum's data from the C2D Database. See the codebook for a detailed description of all variables.

referendum(id, incl_archive = FALSE, tidy = TRUE)

Arguments

id
The referendum's unique identifier.

incl_archive
Whether or not to include an archive column containing data from an earlier, obsolete state of the C2D database.

tidy
Whether or not to tidy the referendum data, i.e. apply various data cleansing tasks and add additional variables. If FALSE, the raw MongoDB referendum data will only be modified just enough to be able to return it as a tibble. Note that untidy data doesn't conform to the codebook (i.a. different variable names).

Value

A tibble.

See also

Examples

c2d::referendum(id = "5bbbe26a92a21351232dd73f")
#> # A tibble: 1 × 69
#>   id            id_official id_sudd country_code country_code_hist… country_name
#>   <chr>         <chr>       <chr>   <fct>        <chr>              <fct>       
#> 1 5bbbe26a92a2… 5           NA      CH           NA                 Switzerland 
#> # … with 63 more variables: subnational_entity_name <fct>, municipality <fct>,
#> #   level <ord>, is_past_jurisdiction <lgl>, date <date>, title_de <chr>,
#> #   title_fr <chr>, title_en <chr>, question <chr>, question_en <chr>,
#> #   committee_name <chr>, result <fct>, subterritories_yes <dbl>,
#> #   subterritories_no <dbl>, electorate_total <int>, electorate_abroad <int>,
#> #   votes_yes <int>, votes_no <int>, votes_empty <int>, votes_invalid <int>,
#> #   votes_per_subterritory <list>, lower_house_yes <int>, …