Get referendum data

Downloads the referendum data from the C2D Database.

referendums(
  use_cache = TRUE,
  cache_lifespan = "1 week",
  incl_archive = FALSE,
  is_draft = FALSE,
  country_code = NULL,
  subnational_entity_name = NULL,
  municipality = NULL,
  level = NULL,
  type = NULL,
  date = NULL,
  date_time_created_min = NULL,
  date_time_created_max = NULL,
  query_filter = NULL
)

Arguments

use_cache
Return cached results if possible. If FALSE, results are always newly fetched regardless of cache_lifespan.

cache_lifespan
The duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Only relevant if use_cache = TRUE.

incl_archive
Whether or not to include an archive column containing data from an earlier, obsolete state of the C2D database.

is_draft
TRUE means to include only referendum entries with draft status, FALSE to include only normal entries. Set to NULL in order to include both draft and normal entries.

country_code
The country_code(s) to be included. A character vector.

subnational_entity_name
The subnational_entity_name(s) to be included. A character vector.

municipality
The municipality(s) to be included. A character vector.

level
The level(s) to be included. A character vector.

type
The type(s) to be included. A character vector.

date
The date(s) to be included. A character vector.

date_time_created_min
The minimum date_time_created to be included. A date, a datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05").

date_time_created_max
The maximum date_time_created to be included. A date, a datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05").

query_filter
A valid MongoDB JSON query filter document which allows for maximum control over what data is included. This takes precedence over all the other parameters, i.e. if query_filter is provided, all the other parameters are ignored.

Value

A tibble.

Examples

if (FALSE) {
# get all referendums (excl. drafts)
c2d::referendums()}

# get only referendums in Austria and Australia on subnational level
c2d::referendums(country_code = c("AT", "AU"),
                 level = "subnational")

#> # A tibble: 31 x 52
#>    id           id_official id_sudd country_code country_name subnational_entit…
#>    <chr>        <chr>       <chr>   <fct>        <fct>        <fct>             
#>  1 5bbbfb4f92a… 12707       NA      AU           Australia    Northern Territory
#>  2 5bbbf9b192a… 1219        NA      AU           Australia    Christmas Island  
#>  3 5bbbf92f92a… 12706       NA      AU           Australia    Northern Territory
#>  4 5bbbf82092a… 1710        NA      AU           Australia    Northern Territory
#>  5 5bbbf7ff92a… 12705       NA      AU           Australia    Northern Territory
#>  6 5bbbf70392a… 12704       NA      AU           Australia    Northern Territory
#>  7 5bbbf59192a… 12703       NA      AU           Australia    Northern Territory
#>  8 5bbbf4d492a… 872         NA      AU           Australia    Christmas Island  
#>  9 5bbbf4d492a… 871         NA      AU           Australia    Christmas Island  
#> 10 5bbbf29e92a… 12694       NA      AU           Australia    Northern Territory
#> # … with 21 more rows, and 46 more variables: municipality <fct>, level <chr>,
#> #   date <date>, title_de <chr>, title_fr <chr>, title_en <chr>,
#> #   committee_name <chr>, result <chr>, electorate_total <int>,
#> #   electorate_abroad <int>, votes_yes <int>, votes_no <int>,
#> #   votes_empty <int>, votes_invalid <int>, votes_per_subterritory <list>,
#> #   tags <list>, remarks <chr>, files <list>, url_sudd <chr>,
#> #   url_swissvotes <chr>, is_draft <lgl>, date_time_created <dttm>, type <chr>,
#> #   inst_legal_basis_type <chr>, inst_has_urgent_legal_basis <lgl>,
#> #   inst_is_binding <lgl>, inst_is_counter_proposal <lgl>,
#> #   inst_is_assembly <lgl>, inst_trigger_type <chr>, inst_trigger_actor <chr>,
#> #   inst_trigger_actor_level <chr>, inst_trigger_threshold <chr>,
#> #   inst_trigger_time_limit <chr>, inst_object_type <chr>,
#> #   inst_object_author <chr>, inst_object_legal_level <chr>,
#> #   inst_object_revision_extent <chr>, inst_object_revision_modes <list>,
#> #   inst_object_revision_divisibility <chr>, inst_topics_only <list>,
#> #   inst_topics_excluded <list>, inst_quorum_turnout <chr>,
#> #   inst_quorum_approval <chr>, inst_has_precondition <chr>,
#> #   inst_precondition_actor <chr>, inst_precondition_decision <chr>

# provide custom `query_filter` for more complex queries like regex matches
# cf. https://docs.mongodb.com/manual/reference/operator/query/regex/
c2d::referendums(query_filter = '{"country_code":{"$regex":"A."}}')

#> # A tibble: 225 x 52
#>    id          id_official id_sudd country_code country_name   subnational_enti…
#>    <chr>       <chr>       <chr>   <fct>        <fct>          <fct>            
#>  1 5d2dd694dc… NA          NA      AS           American Samoa NA               
#>  2 5d2dbe4fdc… NA          NA      AG           Antigua and B… NA               
#>  3 5d1cc881dc… NA          NA      AU           Australia      NA               
#>  4 5d1a183bfe… NA          NA      AZ           Azerbaijan     Nagorno-Karabakh 
#>  5 5cc9a4cccb… NA          NA      AZ           Azerbaijan     NA               
#>  6 5cc9a3b061… NA          NA      AZ           Azerbaijan     NA               
#>  7 5cc9a2d161… NA          NA      AZ           Azerbaijan     NA               
#>  8 5cc9a1e861… NA          NA      AZ           Azerbaijan     NA               
#>  9 5cc9a05fcb… NA          NA      AZ           Azerbaijan     NA               
#> 10 5cc99f2bcb… NA          NA      AZ           Azerbaijan     NA               
#> # … with 215 more rows, and 46 more variables: municipality <fct>, level <chr>,
#> #   date <date>, title_de <chr>, title_fr <chr>, title_en <chr>,
#> #   committee_name <chr>, result <chr>, electorate_total <int>,
#> #   electorate_abroad <int>, votes_yes <int>, votes_no <int>,
#> #   votes_empty <int>, votes_invalid <int>, votes_per_subterritory <list>,
#> #   tags <list>, remarks <chr>, files <list>, url_sudd <chr>,
#> #   url_swissvotes <chr>, is_draft <lgl>, date_time_created <dttm>, type <chr>,
#> #   inst_legal_basis_type <chr>, inst_has_urgent_legal_basis <lgl>,
#> #   inst_is_binding <lgl>, inst_is_counter_proposal <lgl>,
#> #   inst_is_assembly <lgl>, inst_trigger_type <chr>, inst_trigger_actor <chr>,
#> #   inst_trigger_actor_level <chr>, inst_trigger_threshold <chr>,
#> #   inst_trigger_time_limit <chr>, inst_object_type <chr>,
#> #   inst_object_author <chr>, inst_object_legal_level <chr>,
#> #   inst_object_revision_extent <chr>, inst_object_revision_modes <list>,
#> #   inst_object_revision_divisibility <chr>, inst_topics_only <list>,
#> #   inst_topics_excluded <list>, inst_quorum_turnout <chr>,
#> #   inst_quorum_approval <chr>, inst_has_precondition <chr>,
#> #   inst_precondition_actor <chr>, inst_precondition_decision <chr>