Test if referendum ID exists

Tests whether the referendum with the supplied id exists or not.

rfrnd_exists(
  id,
  .use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg)
)

Arguments

id
The referendum's unique identifier.

.use_testing_server
Whether or not to use the testing servers instead of the production servers for C2D Services API calls etc.

Value

A logical scalar.

Examples

c2d::rfrnd_exists("6303a4cba52c3995043a8c24")
#> [1] TRUE