Get referendum data

Downloads the referendum data from the C2D Database. See the codebook for a detailed description of all variables.

rfrnds(
  country_code = NULL,
  subnational_entity_name = NULL,
  municipality = NULL,
  level = NULL,
  type = NULL,
  date_min = NULL,
  date_max = NULL,
  is_draft = FALSE,
  date_time_created_min = NULL,
  date_time_created_max = NULL,
  date_time_last_edited_min = NULL,
  date_time_last_edited_max = NULL,
  query_filter = NULL,
  incl_archive = FALSE,
  tidy = TRUE,
  use_cache = TRUE,
  cache_lifespan = "1 week",
  use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg),
  quiet = FALSE
)

Arguments

country_code

The country_code(s) to be included. A character vector.

subnational_entity_name

The subnational_entity_name(s) to be included. A character vector.

municipality

The municipality(s) to be included. A character vector.

level

The level(s) to be included. A character vector.

type

The type(s) to be included. A character vector.

date_min

The minimum date to be included. A date or something coercible to.

date_max

The maximum date to be included. A date or something coercible to.

is_draft

TRUE means to include only referendum entries with draft status, FALSE to include only normal entries. Set to NULL in order to include both draft and normal entries.

date_time_created_min

The minimum date_time_created to be included. A datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05Z"; assumed to be in UTC if no timezone is given).

date_time_created_max

The maximum date_time_created to be included. A datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05Z"; assumed to be in UTC if no timezone is given).

date_time_last_edited_min

The minimum date_time_last_edited to be included. A datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05Z"; assumed to be in UTC if no timezone is given).

date_time_last_edited_max

The maximum date_time_last_edited to be included. A datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05Z"; assumed to be in UTC if no timezone is given).

query_filter

A valid MongoDB JSON query filter document which allows for maximum control over what data is included. This takes precedence over all of the above listed parameters, i.e. if query_filter is provided, the parameters country_code, subnational_entity_name, municipality, level, type, date_min, date_max, is_draft, date_time_created_min, date_time_created_max, date_time_last_edited_min and date_time_last_edited_max are ignored.

incl_archive

Whether or not to include an archive column containing data from an earlier, obsolete state of the C2D database.

tidy

Whether or not to tidy the referendum data, i.e. apply various data cleansing tasks and add additional variables. If FALSE, the raw MongoDB referendum data will only be modified just enough to be able to return it as a tibble. Note that untidy data doesn't conform to the codebook (i.a. different variable names).

use_cache

Return cached results if possible. If FALSE, results are always newly fetched regardless of cache_lifespan.

cache_lifespan

The duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Only relevant if use_cache = TRUE.

use_testing_server

Whether or not to use the testing servers instead of the production servers for C2D Services API calls etc.

quiet

Whether or not to suppress printing status output from internal processing.

Value

A tibble.

See also

Examples

if (FALSE) {
# get all referendums (excl. drafts)
c2d::rfrnds()}

# get only referendums in Austria and Australia on subnational level
c2d::rfrnds(country_code = c("AT", "AU"),
            level = "subnational")
#> ℹ Fetching JSON data from C2D API...
#> Downloading: 16 kB     Downloading: 16 kB     Downloading: 16 kB     Downloading: 16 kB     Downloading: 16 kB     Downloading: 16 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 65 kB     Downloading: 65 kB     Downloading: 65 kB     Downloading: 65 kB     Downloading: 65 kB     Downloading: 65 kB     Downloading: 75 kB     Downloading: 75 kB     Downloading: 75 kB     Downloading: 75 kB     
#> ✔ Fetching JSON data from C2D API... done [140ms]
#> 
#> ℹ Converting JSON to list data...
#> ✔ Converting JSON to list data... done [14ms]
#> 
#> ℹ Tidying data...
#> ✔ Tidying data... done [490ms]
#> 
#> # A tibble: 31 × 69
#>    id           id_of…¹ id_sudd count…² count…³ subna…⁴ munic…⁵ level date      
#>    <chr>        <chr>   <chr>   <fct>   <fct>   <chr>   <chr>   <ord> <date>    
#>  1 5bbbfb4f92a… 12707   NA      AU      Austra… Northe… NA      subn… 2001-10-24
#>  2 5bbbf9b192a… 1219    NA      AU      Austra… Christ… NA      subn… 1999-11-06
#>  3 5bbbf92f92a… 12706   NA      AU      Austra… Northe… NA      subn… 1999-05-12
#>  4 5bbbf82092a… 1710    NA      AU      Austra… Northe… NA      subn… 1998-08-26
#>  5 5bbbf7ff92a… 12705   NA      AU      Austra… Northe… NA      subn… 1998-05-14
#>  6 5bbbf70392a… 12704   NA      AU      Austra… Northe… NA      subn… 1996-10-17
#>  7 5bbbf59192a… 12703   NA      AU      Austra… Northe… NA      subn… 1995-03-15
#>  8 5bbbf4d492a… 872     NA      AU      Austra… Christ… NA      subn… 1994-05-07
#>  9 5bbbf4d492a… 871     NA      AU      Austra… Christ… NA      subn… 1994-05-07
#> 10 5bbbf29e92a… 12694   NA      AU      Austra… Northe… NA      subn… 1991-01-02
#> # … with 21 more rows, 60 more variables: is_former_country <lgl>,
#> #   title_en <chr>, title_de <chr>, title_fr <chr>, question <chr>,
#> #   question_en <chr>, committee_name <chr>, result <fct>,
#> #   subterritories_yes <dbl>, subterritories_no <dbl>, electorate_total <int>,
#> #   electorate_abroad <int>, votes_yes <int>, votes_no <int>,
#> #   votes_empty <int>, votes_invalid <int>, votes_per_subterritory <list>,
#> #   lower_house_yes <int>, lower_house_no <int>, …

# get referendums in 2020
c2d::rfrnds(date_min = "2020-01-01",
            date_max = "2020-12-31")
#> ℹ Fetching JSON data from C2D API...
#> Downloading: 16 kB     Downloading: 16 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 65 kB     Downloading: 65 kB     Downloading: 81 kB     Downloading: 81 kB     Downloading: 81 kB     Downloading: 81 kB     Downloading: 81 kB     Downloading: 81 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 150 kB     Downloading: 150 kB     Downloading: 160 kB     Downloading: 160 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 200 kB     Downloading: 200 kB     Downloading: 200 kB     Downloading: 200 kB     Downloading: 200 kB     Downloading: 200 kB     Downloading: 200 kB     Downloading: 200 kB     Downloading: 210 kB     Downloading: 210 kB     Downloading: 210 kB     Downloading: 210 kB     Downloading: 210 kB     Downloading: 210 kB     Downloading: 230 kB     Downloading: 230 kB     Downloading: 240 kB     Downloading: 240 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 280 kB     Downloading: 280 kB     Downloading: 280 kB     Downloading: 280 kB     Downloading: 280 kB     Downloading: 280 kB     Downloading: 290 kB     Downloading: 290 kB     Downloading: 290 kB     Downloading: 290 kB     Downloading: 310 kB     Downloading: 310 kB     Downloading: 310 kB     Downloading: 310 kB     Downloading: 310 kB     Downloading: 310 kB     Downloading: 310 kB     Downloading: 310 kB     
#> ✔ Fetching JSON data from C2D API... done [169ms]
#> 
#> ℹ Converting JSON to list data...
#> ✔ Converting JSON to list data... done [15ms]
#> 
#> ℹ Tidying data...
#> ✔ Tidying data... done [831ms]
#> 
#> # A tibble: 103 × 69
#>    id           id_of…¹ id_sudd count…² count…³ subna…⁴ munic…⁵ level date      
#>    <chr>        <chr>   <chr>   <fct>   <fct>   <chr>   <chr>   <ord> <date>    
#>  1 6010898bd12… NA      NA      VE      Venezu… NA      NA      nati… 2020-12-12
#>  2 60108788d12… NA      NA      VE      Venezu… NA      NA      nati… 2020-12-12
#>  3 60108540d12… NA      NA      VE      Venezu… NA      NA      nati… 2020-12-12
#>  4 601022ddd12… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08
#>  5 6010202dd12… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08
#>  6 60101d90d12… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08
#>  7 60101b85d12… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08
#>  8 6010195ad12… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08
#>  9 60101462d12… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08
#> 10 60101127d12… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08
#> # … with 93 more rows, 60 more variables: is_former_country <lgl>,
#> #   title_en <chr>, title_de <chr>, title_fr <chr>, question <chr>,
#> #   question_en <chr>, committee_name <chr>, result <fct>,
#> #   subterritories_yes <dbl>, subterritories_no <dbl>, electorate_total <int>,
#> #   electorate_abroad <int>, votes_yes <int>, votes_no <int>,
#> #   votes_empty <int>, votes_invalid <int>, votes_per_subterritory <list>,
#> #   lower_house_yes <int>, lower_house_no <int>, …

# get referendums added to the database during the last 30 days
c2d::rfrnds(date_time_created_min = clock::date_today(zone = "UTC") |> clock::add_days(-30L),
            date_time_created_max = clock::date_today(zone = "UTC"))
#> ℹ Fetching JSON data from C2D API...
#> Downloading: 9.5 kB     Downloading: 9.5 kB     Downloading: 9.5 kB     Downloading: 9.5 kB     
#> ✔ Fetching JSON data from C2D API... done [92ms]
#> 
#> ℹ Converting JSON to list data...
#> ✔ Converting JSON to list data... done [12ms]
#> 
#> ℹ Tidying data...
#> ✔ Tidying data... done [482ms]
#> 
#> # A tibble: 4 × 69
#>   id    id_of…¹ id_sudd count…² count…³ subna…⁴ munic…⁵ level date       is_fo…⁶
#>   <chr> <chr>   <chr>   <fct>   <fct>   <chr>   <chr>   <ord> <date>     <lgl>  
#> 1 6384… NA      NA      CH      Switze… Ticino  NA      subn… 2022-10-30 FALSE  
#> 2 6384… NA      NA      CH      Switze… Ticino  NA      subn… 2022-10-30 FALSE  
#> 3 6384… NA      NA      CH      Switze… Ticino  NA      subn… 2022-10-30 FALSE  
#> 4 6384… NA      NA      CH      Switze… Ticino  NA      subn… 2022-10-30 FALSE  
#> # … with 59 more variables: title_en <chr>, title_de <chr>, title_fr <chr>,
#> #   question <chr>, question_en <chr>, committee_name <chr>, result <fct>,
#> #   subterritories_yes <dbl>, subterritories_no <dbl>, electorate_total <int>,
#> #   electorate_abroad <int>, votes_yes <int>, votes_no <int>,
#> #   votes_empty <int>, votes_invalid <int>, votes_per_subterritory <list>,
#> #   lower_house_yes <int>, lower_house_no <int>, lower_house_abstentions <int>,
#> #   upper_house_yes <int>, upper_house_no <int>, …

# provide custom `query_filter` for more complex queries like regex matches
# cf. https://docs.mongodb.com/manual/reference/operator/query/regex/
c2d::rfrnds(query_filter = '{"country_code":{"$regex":"A."}}')
#> ℹ Fetching JSON data from C2D API...
#> Downloading: 16 kB     Downloading: 16 kB     Downloading: 16 kB     Downloading: 16 kB     Downloading: 16 kB     Downloading: 16 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 65 kB     Downloading: 65 kB     Downloading: 65 kB     Downloading: 65 kB     Downloading: 65 kB     Downloading: 65 kB     Downloading: 65 kB     Downloading: 65 kB     Downloading: 65 kB     Downloading: 65 kB     Downloading: 81 kB     Downloading: 81 kB     Downloading: 81 kB     Downloading: 81 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 150 kB     Downloading: 150 kB     Downloading: 150 kB     Downloading: 150 kB     Downloading: 150 kB     Downloading: 150 kB     Downloading: 160 kB     Downloading: 160 kB     Downloading: 160 kB     Downloading: 160 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 200 kB     Downloading: 200 kB     Downloading: 200 kB     Downloading: 200 kB     Downloading: 200 kB     Downloading: 200 kB     Downloading: 200 kB     Downloading: 200 kB     Downloading: 210 kB     Downloading: 210 kB     Downloading: 230 kB     Downloading: 230 kB     Downloading: 230 kB     Downloading: 230 kB     Downloading: 240 kB     Downloading: 240 kB     Downloading: 240 kB     Downloading: 240 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 280 kB     Downloading: 280 kB     Downloading: 280 kB     Downloading: 280 kB     Downloading: 280 kB     Downloading: 280 kB     Downloading: 280 kB     Downloading: 280 kB     Downloading: 290 kB     Downloading: 290 kB     Downloading: 290 kB     Downloading: 290 kB     Downloading: 310 kB     Downloading: 310 kB     Downloading: 310 kB     Downloading: 310 kB     Downloading: 310 kB     Downloading: 310 kB     Downloading: 330 kB     Downloading: 330 kB     Downloading: 330 kB     Downloading: 330 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 360 kB     Downloading: 360 kB     Downloading: 360 kB     Downloading: 360 kB     Downloading: 360 kB     Downloading: 360 kB     Downloading: 380 kB     Downloading: 380 kB     Downloading: 380 kB     Downloading: 380 kB     Downloading: 380 kB     Downloading: 380 kB     Downloading: 380 kB     Downloading: 380 kB     Downloading: 390 kB     Downloading: 390 kB     Downloading: 390 kB     Downloading: 390 kB     Downloading: 390 kB     Downloading: 390 kB     Downloading: 410 kB     Downloading: 410 kB     Downloading: 410 kB     Downloading: 410 kB     Downloading: 410 kB     Downloading: 410 kB     Downloading: 420 kB     Downloading: 420 kB     Downloading: 420 kB     Downloading: 420 kB     Downloading: 420 kB     Downloading: 420 kB     Downloading: 440 kB     Downloading: 440 kB     Downloading: 440 kB     Downloading: 440 kB     Downloading: 460 kB     Downloading: 460 kB     Downloading: 460 kB     Downloading: 460 kB     Downloading: 460 kB     Downloading: 460 kB     Downloading: 470 kB     Downloading: 470 kB     Downloading: 470 kB     Downloading: 470 kB     Downloading: 470 kB     Downloading: 470 kB     Downloading: 490 kB     Downloading: 490 kB     Downloading: 490 kB     Downloading: 490 kB     Downloading: 490 kB     Downloading: 490 kB     Downloading: 510 kB     Downloading: 510 kB     Downloading: 510 kB     Downloading: 510 kB     Downloading: 510 kB     Downloading: 510 kB     Downloading: 510 kB     Downloading: 510 kB     Downloading: 520 kB     Downloading: 520 kB     Downloading: 520 kB     Downloading: 520 kB     Downloading: 520 kB     Downloading: 520 kB     Downloading: 520 kB     Downloading: 520 kB     Downloading: 540 kB     Downloading: 540 kB     Downloading: 540 kB     Downloading: 540 kB     Downloading: 540 kB     Downloading: 540 kB     Downloading: 560 kB     Downloading: 560 kB     Downloading: 560 kB     Downloading: 560 kB     Downloading: 560 kB     Downloading: 560 kB     Downloading: 560 kB     Downloading: 560 kB     Downloading: 570 kB     Downloading: 570 kB     Downloading: 570 kB     Downloading: 570 kB     Downloading: 570 kB     Downloading: 570 kB     Downloading: 570 kB     Downloading: 570 kB     Downloading: 590 kB     Downloading: 590 kB     Downloading: 590 kB     Downloading: 590 kB     Downloading: 590 kB     Downloading: 590 kB     Downloading: 590 kB     Downloading: 590 kB     Downloading: 600 kB     Downloading: 600 kB     Downloading: 600 kB     Downloading: 600 kB     Downloading: 600 kB     Downloading: 600 kB     Downloading: 600 kB     Downloading: 600 kB     Downloading: 620 kB     Downloading: 620 kB     Downloading: 620 kB     Downloading: 620 kB     Downloading: 620 kB     Downloading: 620 kB     Downloading: 620 kB     Downloading: 620 kB     Downloading: 620 kB     Downloading: 620 kB     Downloading: 640 kB     Downloading: 640 kB     Downloading: 640 kB     Downloading: 640 kB     Downloading: 650 kB     Downloading: 650 kB     Downloading: 650 kB     Downloading: 650 kB     Downloading: 650 kB     Downloading: 650 kB     Downloading: 670 kB     Downloading: 670 kB     Downloading: 670 kB     Downloading: 670 kB     Downloading: 670 kB     Downloading: 670 kB     Downloading: 670 kB     Downloading: 670 kB     Downloading: 690 kB     Downloading: 690 kB     Downloading: 690 kB     Downloading: 690 kB     Downloading: 690 kB     Downloading: 690 kB     Downloading: 690 kB     Downloading: 690 kB     Downloading: 700 kB     Downloading: 700 kB     Downloading: 700 kB     Downloading: 700 kB     Downloading: 700 kB     Downloading: 700 kB     Downloading: 720 kB     Downloading: 720 kB     Downloading: 720 kB     Downloading: 720 kB     Downloading: 720 kB     Downloading: 720 kB     Downloading: 720 kB     Downloading: 720 kB     Downloading: 740 kB     Downloading: 740 kB     Downloading: 740 kB     Downloading: 740 kB     Downloading: 740 kB     Downloading: 740 kB     Downloading: 750 kB     Downloading: 750 kB     Downloading: 750 kB     Downloading: 750 kB     Downloading: 750 kB     Downloading: 750 kB     Downloading: 770 kB     Downloading: 770 kB     Downloading: 770 kB     Downloading: 770 kB     Downloading: 770 kB     Downloading: 770 kB     Downloading: 780 kB     Downloading: 780 kB     Downloading: 780 kB     Downloading: 780 kB     Downloading: 780 kB     Downloading: 780 kB     Downloading: 800 kB     Downloading: 800 kB     Downloading: 800 kB     Downloading: 800 kB     Downloading: 800 kB     Downloading: 800 kB     Downloading: 820 kB     Downloading: 820 kB     Downloading: 820 kB     Downloading: 820 kB     Downloading: 820 kB     Downloading: 820 kB     Downloading: 830 kB     Downloading: 830 kB     Downloading: 830 kB     Downloading: 830 kB     Downloading: 830 kB     Downloading: 830 kB     Downloading: 850 kB     Downloading: 850 kB     Downloading: 850 kB     Downloading: 850 kB     Downloading: 850 kB     Downloading: 850 kB     Downloading: 870 kB     Downloading: 870 kB     Downloading: 870 kB     Downloading: 870 kB     Downloading: 870 kB     Downloading: 870 kB     Downloading: 880 kB     Downloading: 880 kB     Downloading: 880 kB     Downloading: 880 kB     Downloading: 880 kB     Downloading: 880 kB     Downloading: 900 kB     Downloading: 900 kB     Downloading: 900 kB     Downloading: 900 kB     Downloading: 900 kB     Downloading: 900 kB     
#> ✔ Fetching JSON data from C2D API... done [271ms]
#> 
#> ℹ Converting JSON to list data...
#> ✔ Converting JSON to list data... done [27ms]
#> 
#> ℹ Tidying data...
#> ✔ Tidying data... done [894ms]
#> 
#> # A tibble: 232 × 69
#>    id           id_of…¹ id_sudd count…² count…³ subna…⁴ munic…⁵ level date      
#>    <chr>        <chr>   <chr>   <fct>   <fct>   <chr>   <chr>   <ord> <date>    
#>  1 630e49ffa52… solala  au?     BQAQ    Britis… NA      NA      nati… 1970-02-02
#>  2 630e4611a52… 123     ff3     AT      Austria NA      NA      nati… 2222-02-22
#>  3 611d89b4c72… NA      NA      AT      Austria NA      NA      nati… 2222-08-18
#>  4 611d82c4c72… NA      NA      AT      Austria NA      NA      nati… 2222-08-18
#>  5 611d5acac72… NA      NA      AT      Austria Aargau  NA      subn… 2222-08-18
#>  6 611d524fc72… NA      NA      AT      Austria Aargau  NA      subn… 2222-08-18
#>  7 5d2dd694dc0… NA      NA      AS      Americ… NA      NA      nati… 2018-11-06
#>  8 5d2dbe4fdc0… NA      NA      AG      Antigu… NA      NA      nati… 2018-11-06
#>  9 5d1cc881dc0… NA      NA      AU      Austra… NA      NA      nati… 2017-11-07
#> 10 5d1a183bfe6… NA      NA      AZ      Azerba… Nagorn… NA      subn… 2017-02-20
#> # … with 222 more rows, 60 more variables: is_former_country <lgl>,
#> #   title_en <chr>, title_de <chr>, title_fr <chr>, question <chr>,
#> #   question_en <chr>, committee_name <chr>, result <fct>,
#> #   subterritories_yes <dbl>, subterritories_no <dbl>, electorate_total <int>,
#> #   electorate_abroad <int>, votes_yes <int>, votes_no <int>,
#> #   votes_empty <int>, votes_invalid <int>, votes_per_subterritory <list>,
#> #   lower_house_yes <int>, lower_house_no <int>, …