Count number of referendums per period

Counts the number of C2D referendums per desired period, optionally by an additional column specified via group_by.

rfrnds_per_period(
  data,
  period = c("week", "month", "quarter", "year", "decade", "century"),
  group_by = NULL
)

Arguments

data
C2D referendum data as returned by rfrnds(). A data frame that at minimum contains the column specified in period or the column date (to compute the period column), plus the one specified via group_by (if any).

period
Type of period to count referendums by. One of "week","month","quarter","year","decade" or "century".

group_by
Optional data column to group by before counting number of referendums. A symbol.

Value

A tibble.

Examples

c2d::rfrnds(country_code = "AT") |> c2d::rfrnds_per_period()
#> # A tibble: 53 × 2
#>     week     n
#>    <int> <int>
#>  1     1     0
#>  2     2     0
#>  3     3     2
#>  4     4     0
#>  5     5     0
#>  6     6     0
#>  7     7     0
#>  8     8     0
#>  9     9     0
#> 10    10     0
#> # … with 43 more rows
c2d::rfrnds(country_code = "AT") |> c2d::rfrnds_per_period(group_by = level)
#> # A tibble: 106 × 3
#>    level        week     n
#>    <ord>       <int> <int>
#>  1 subnational     1     0
#>  2 subnational     2     0
#>  3 subnational     3     0
#>  4 subnational     4     0
#>  5 subnational     5     0
#>  6 subnational     6     0
#>  7 subnational     7     0
#>  8 subnational     8     0
#>  9 subnational     9     0
#> 10 subnational    10     0
#> # … with 96 more rows
c2d::rfrnds(country_code = "AT") |> c2d::rfrnds_per_period(period = "decade", group_by = level)
#> # A tibble: 22 × 3
#>    level       decade     n
#>    <ord>        <int> <int>
#>  1 subnational   1910     1
#>  2 subnational   1920     5
#>  3 subnational   1930     0
#>  4 subnational   1940     0
#>  5 subnational   1950     0
#>  6 subnational   1960     0
#>  7 subnational   1970     0
#>  8 subnational   1980     0
#>  9 subnational   1990     0
#> 10 subnational   2000     0
#> # … with 12 more rows