Count number of referendums per year

Counts number of C2D referendums per year, optionally by an additional column specified via group_by.

rfrnds_per_year(data, group_by = NULL)

Arguments

data
C2D referendum data as returned by rfrnds(). A data frame that at minimum contains the column date plus the one specified via group_by (if any).

group_by
Optional data column to group by before counting number of referendums. A symbol.

Value

A tibble.

Examples

c2d::rfrnds(country_code = "AT") |> c2d::rfrnds_per_year()
#> # A tibble: 95 × 2
#>     year     n
#>    <int> <int>
#>  1  1919     1
#>  2  1920     3
#>  3  1921     2
#>  4  1922     0
#>  5  1923     0
#>  6  1924     0
#>  7  1925     0
#>  8  1926     0
#>  9  1927     0
#> 10  1928     0
#> # … with 85 more rows
c2d::rfrnds(country_code = "AT") |> c2d::rfrnds_per_year(group_by = level)
#> # A tibble: 190 × 3
#>    level        year     n
#>    <ord>       <int> <int>
#>  1 subnational  1919     1
#>  2 subnational  1920     3
#>  3 subnational  1921     2
#>  4 subnational  1922     0
#>  5 subnational  1923     0
#>  6 subnational  1924     0
#>  7 subnational  1925     0
#>  8 subnational  1926     0
#>  9 subnational  1927     0
#> 10 subnational  1928     0
#> # … with 180 more rows