Search in English referendum titles

Allows to use the C2D API's primitive search functionality. Note that

the search is not case-sensitive and no fuzzy search is performed (i.e. only exact matches are returned).

only up to the first 5 matching results will be returned.

search_rfrnds(
  term,
  use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg)
)

Arguments

term

Search term. A character scalar.

use_testing_server

Whether or not to use the testing servers instead of the production servers for C2D Services API calls etc.

Value

A character vector of English referendum titles matching the search term.

Details

Note that this function is probably not of much use since it doesn't return any additional information about the matched referendums but only the English titles.

See also

Examples

c2d::search_rfrnds("freedom")
#> [1] "Amendment of the law on demolition, conversion and renovation of residential buildings (\"more freedom for renters\") of 13 November 2015"
#> [2] "Constitutional proposal 2-267: Guarantee of academic freedom in any institution of higher learning"                                       
#> [3] "Constitutional reform (freedom of information)"                                                                                           
#> [4] "Constitutional reform (human and civil rights and freedoms)"                                                                              
#> [5] "Extension of personal rights and freedoms"