Get referendum data from sudd.ch

Downloads referendum data from sudd.ch.

sudd_referendums(ids_sudd)

Arguments

ids_sudd
The referendum identifiers assigned by sudd.ch. Either as a character vector or a data frame containing a column id_sudd. NAs are ignored.

Value

A tibble. The column names are aligned with those of referendums() as closely as possible.

About sudd.ch

sudd stands for Suchmaschine für direkte Demokratie (German) and is operated by Beat Müller. Its database content is licensed under Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International (CC BY-NC-SA 4.0).

See also

Examples

c2d::referendum(id = "5bbc045192a21351232e596f")$id_sudd |> c2d::sudd_referendums()

#> # A tibble: 1 × 29
#>   id_sudd  country_code country_code_historical country_name  territory_name_de
#>   <chr>    <chr>        <chr>                   <chr>         <chr>            
#> 1 li012015 LI           NA                      Liechtenstein Liechtenstein    
#> # … with 24 more variables: territory_type_de <chr>, date <date>, year <int>,
#> #   month <int>, day <int>, title_de <chr>, question_type_de <chr>,
#> #   types <list>, result_de <chr>, result_status_de <chr>,
#> #   adoption_requirements_de <list>, electorate_total <int>,
#> #   polling_cards <int>, votes_total <int>, votes_empty <int>,
#> #   votes_void <int>, votes_invalid <int>, votes_valid <int>, votes_yes <int>,
#> #   votes_no <int>, files <list>, remarks <list>, sources <list>, …

c2d::referendums(country_code = "AT") |> c2d::sudd_referendums()

#> # A tibble: 5 × 28
#>   id_sudd  country_code country_code_historical country_name territory_name_de
#>   <chr>    <chr>        <chr>                   <chr>        <chr>            
#> 1 at012013 AT           NA                      Austria      Österreich       
#> 2 at012013 AT           NA                      Austria      Österreich       
#> 3 at011994 AT           NA                      Austria      Österreich       
#> 4 at011978 AT           NA                      Austria      Österreich       
#> 5 at011938 AT           NA                      Austria      Österreich       
#> # … with 23 more variables: territory_type_de <chr>, date <date>, year <int>,
#> #   month <int>, day <int>, title_de <chr>, question_type_de <chr>,
#> #   types <list>, result_de <chr>, result_status_de <chr>,
#> #   adoption_requirements_de <list>, electorate_total <int>, votes_total <int>,
#> #   votes_invalid <int>, votes_valid <int>, votes_yes <int>, votes_no <int>,
#> #   votes_option_1 <int>, votes_option_2 <int>, files <list>, remarks <list>,
#> #   sources <list>, date_last_edited <date>