Tabulate number of referendums per period

Creates a ready-to-print gt table with the number of referendums per period, optionally counted by one or more additional columns.

tbl_n_rfrnds_per_period(
  data,
  by = NULL,
  period = c("week", "month", "quarter", "year", "decade", "century"),
  fill_gaps = TRUE,
  period_floor = NULL,
  period_ceiling = NULL,
  squeeze_zero_rows = TRUE,
  descending = TRUE,
  add_total_row = TRUE,
  add_total_col = NULL
)

Arguments

data
C2D referendum data as returned by rfrnds(). A data frame that at minimum contains the column specified in period or the column date (to compute the period column), plus the one(s) specified via by (if any).

by
Up to two additional data column(s) to group by before counting number of referendums. Tidy selections are supported.

period
Type of period to count referendums by. One of "week","month","quarter","year","decade" or "century".

fill_gaps
Whether or not to add zero-value rows to the result for period gaps in data.

period_floor
Lower period limit up to which gaps are filled. If NULL, the lower limit is set to the minimum of period present in data. Only relevant if fill_gaps = TRUE and period is set to a unique timespan type ("year", "decade" or "century").

period_ceiling
Upper period limit up to which gaps are filled. If NULL, the upper limit is set to the maximum of period present in data. Only relevant if fill_gaps = TRUE and period is set to a unique timespan type ("year", "decade" or "century").

squeeze_zero_rows
Whether or not to compress consecutive zero-sum rows into single period span rows.

descending
Whether to sort the resulting table by period in descending or in ascending order.

add_total_row
Whether or not to add a summary row at the very end of the table containing column totals.

add_total_col
Whether or not to add a summary column at the very end of the table containing row totals. If NULL, a total column is added only if by is non-empty.

Value

A gt_tbl object.

Details

Note that the period types "year", "decade" and "century" are unique timespans, while "week", "month" and "quarter" are recurring timespans (e.g. every year has a week 1).

Examples

c2d::rfrnds(country_code = "AT") |> c2d::tbl_n_rfrnds_per_period(period = "decade")
#> <div id="efoodfieex" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #efoodfieex .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #efoodfieex .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #efoodfieex .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #efoodfieex .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #efoodfieex .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #efoodfieex .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #efoodfieex .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #efoodfieex .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #efoodfieex .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #efoodfieex .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #efoodfieex .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #efoodfieex .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #efoodfieex .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #efoodfieex .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #efoodfieex .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #efoodfieex .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #efoodfieex .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #efoodfieex .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #efoodfieex .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #efoodfieex .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #efoodfieex .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #efoodfieex .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #efoodfieex .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #efoodfieex .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #efoodfieex .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #efoodfieex .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #efoodfieex .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #efoodfieex .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #efoodfieex .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-left: 4px;
#>   padding-right: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #efoodfieex .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #efoodfieex .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #efoodfieex .gt_left {
#>   text-align: left;
#> }
#> 
#> #efoodfieex .gt_center {
#>   text-align: center;
#> }
#> 
#> #efoodfieex .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #efoodfieex .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #efoodfieex .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #efoodfieex .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #efoodfieex .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #efoodfieex .gt_footnote_marks {
#>   font-style: italic;
#>   font-weight: normal;
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#> }
#> 
#> #efoodfieex .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #efoodfieex .gt_indent_1 {
#>   text-indent: 5px;
#> }
#> 
#> #efoodfieex .gt_indent_2 {
#>   text-indent: 10px;
#> }
#> 
#> #efoodfieex .gt_indent_3 {
#>   text-indent: 15px;
#> }
#> 
#> #efoodfieex .gt_indent_4 {
#>   text-indent: 20px;
#> }
#> 
#> #efoodfieex .gt_indent_5 {
#>   text-indent: 25px;
#> }
#> </style>
#>   <table class="gt_table">
#>   
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" style="text-align: left; vertical-align: middle;" scope="col"></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">n</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">2010s</th>
#> <td class="gt_row gt_right">2</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">2000s</th>
#> <td class="gt_row gt_right">0</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">1990s</th>
#> <td class="gt_row gt_right">1</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">1980s</th>
#> <td class="gt_row gt_right">0</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">1970s</th>
#> <td class="gt_row gt_right">1</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">1940s–1960s</th>
#> <td class="gt_row gt_right">0</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">1930s</th>
#> <td class="gt_row gt_right">1</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">1920s</th>
#> <td class="gt_row gt_right">5</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">1910s</th>
#> <td class="gt_row gt_right">1</td></tr>
#>     <tr><td class="gt_row gt_left gt_stub gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">Total</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">11</td></tr>
#>   </tbody>
#>   
#>   
#> </table>
#> </div>

# grouped by a single additional column
c2d::rfrnds(country_code = "AT") |> c2d::tbl_n_rfrnds_per_period(by = level,
                                                                 period = "decade")
#> <div id="ombwkkeulb" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #ombwkkeulb .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #ombwkkeulb .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ombwkkeulb .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #ombwkkeulb .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #ombwkkeulb .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #ombwkkeulb .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ombwkkeulb .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #ombwkkeulb .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #ombwkkeulb .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #ombwkkeulb .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #ombwkkeulb .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #ombwkkeulb .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #ombwkkeulb .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #ombwkkeulb .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #ombwkkeulb .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #ombwkkeulb .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #ombwkkeulb .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ombwkkeulb .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #ombwkkeulb .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #ombwkkeulb .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ombwkkeulb .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #ombwkkeulb .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #ombwkkeulb .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #ombwkkeulb .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ombwkkeulb .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #ombwkkeulb .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #ombwkkeulb .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #ombwkkeulb .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ombwkkeulb .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-left: 4px;
#>   padding-right: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ombwkkeulb .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ombwkkeulb .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ombwkkeulb .gt_left {
#>   text-align: left;
#> }
#> 
#> #ombwkkeulb .gt_center {
#>   text-align: center;
#> }
#> 
#> #ombwkkeulb .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #ombwkkeulb .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #ombwkkeulb .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #ombwkkeulb .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #ombwkkeulb .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #ombwkkeulb .gt_footnote_marks {
#>   font-style: italic;
#>   font-weight: normal;
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#> }
#> 
#> #ombwkkeulb .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #ombwkkeulb .gt_indent_1 {
#>   text-indent: 5px;
#> }
#> 
#> #ombwkkeulb .gt_indent_2 {
#>   text-indent: 10px;
#> }
#> 
#> #ombwkkeulb .gt_indent_3 {
#>   text-indent: 15px;
#> }
#> 
#> #ombwkkeulb .gt_indent_4 {
#>   text-indent: 20px;
#> }
#> 
#> #ombwkkeulb .gt_indent_5 {
#>   text-indent: 25px;
#> }
#> </style>
#>   <table class="gt_table">
#>   
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" style="text-align: left; vertical-align: middle;" scope="col"><em>Political level</em></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">local</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">subnational</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">national</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">Total</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">2010s</th>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">2</td>
#> <td class="gt_row gt_right">2</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">2000s</th>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">1990s</th>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">1</td>
#> <td class="gt_row gt_right">1</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">1980s</th>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">1970s</th>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">1</td>
#> <td class="gt_row gt_right">1</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">1940s–1960s</th>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">1930s</th>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">1</td>
#> <td class="gt_row gt_right">1</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">1920s</th>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">5</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">5</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">1910s</th>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">1</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">1</td></tr>
#>     <tr><td class="gt_row gt_left gt_stub gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">Total</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">0</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">6</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">5</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">11</td></tr>
#>   </tbody>
#>   
#>   
#> </table>
#> </div>

# grouped by two addtional columns
c2d::rfrnds(country_code = "AT") |> c2d::tbl_n_rfrnds_per_period(by = c(level, type),
                                                                 period = "decade")
#> <div id="fdpbcyvkgu" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #fdpbcyvkgu .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #fdpbcyvkgu .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #fdpbcyvkgu .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #fdpbcyvkgu .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #fdpbcyvkgu .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #fdpbcyvkgu .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #fdpbcyvkgu .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #fdpbcyvkgu .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #fdpbcyvkgu .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #fdpbcyvkgu .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #fdpbcyvkgu .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #fdpbcyvkgu .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #fdpbcyvkgu .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #fdpbcyvkgu .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #fdpbcyvkgu .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #fdpbcyvkgu .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #fdpbcyvkgu .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #fdpbcyvkgu .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #fdpbcyvkgu .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #fdpbcyvkgu .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #fdpbcyvkgu .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #fdpbcyvkgu .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #fdpbcyvkgu .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #fdpbcyvkgu .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #fdpbcyvkgu .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #fdpbcyvkgu .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #fdpbcyvkgu .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #fdpbcyvkgu .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #fdpbcyvkgu .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-left: 4px;
#>   padding-right: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #fdpbcyvkgu .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #fdpbcyvkgu .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #fdpbcyvkgu .gt_left {
#>   text-align: left;
#> }
#> 
#> #fdpbcyvkgu .gt_center {
#>   text-align: center;
#> }
#> 
#> #fdpbcyvkgu .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #fdpbcyvkgu .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #fdpbcyvkgu .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #fdpbcyvkgu .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #fdpbcyvkgu .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #fdpbcyvkgu .gt_footnote_marks {
#>   font-style: italic;
#>   font-weight: normal;
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#> }
#> 
#> #fdpbcyvkgu .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #fdpbcyvkgu .gt_indent_1 {
#>   text-indent: 5px;
#> }
#> 
#> #fdpbcyvkgu .gt_indent_2 {
#>   text-indent: 10px;
#> }
#> 
#> #fdpbcyvkgu .gt_indent_3 {
#>   text-indent: 15px;
#> }
#> 
#> #fdpbcyvkgu .gt_indent_4 {
#>   text-indent: 20px;
#> }
#> 
#> #fdpbcyvkgu .gt_indent_5 {
#>   text-indent: 25px;
#> }
#> </style>
#>   <table class="gt_table">
#>   
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="2" colspan="1" style="text-align: left; vertical-align: middle;" scope="col"><em>Political level</em><br><br><em>Type</em></th>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="8" scope="colgroup">
#>         <span class="gt_column_spanner">local</span>
#>       </th>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="8" scope="colgroup">
#>         <span class="gt_column_spanner">subnational</span>
#>       </th>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="8" scope="colgroup">
#>         <span class="gt_column_spanner">national</span>
#>       </th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="2" colspan="1" scope="col">Total</th>
#>     </tr>
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">mandatory referendum</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">optional referendum</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">counter proposal</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">citizens' initiative</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">governmental referendum</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">citizens' assembly</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">recall</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">N/A</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">mandatory referendum</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">optional referendum</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">counter proposal</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">citizens' initiative</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">governmental referendum</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">citizens' assembly</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">recall</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">N/A</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">mandatory referendum</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">optional referendum</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">counter proposal</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">citizens' initiative</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">governmental referendum</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">citizens' assembly</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">recall</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col">N/A</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">2010s</th>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">2</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">2</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">2000s</th>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">1990s</th>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">1</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">1</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">1980s</th>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">1970s</th>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">1</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">1</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">1940s–1960s</th>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">1930s</th>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">1</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">1</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">1920s</th>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">4</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">1</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">5</td></tr>
#>     <tr><th scope="row" class="gt_row gt_left gt_stub">1910s</th>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">1</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">0</td>
#> <td class="gt_row gt_right">1</td></tr>
#>     <tr><td class="gt_row gt_left gt_stub gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">Total</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">0</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">0</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">0</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">0</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">0</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">0</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">0</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">0</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">0</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">0</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">0</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">0</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">5</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">0</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">0</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">1</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">1</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">0</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">0</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">0</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">4</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">0</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">0</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">0</td>
#> <td class="gt_row gt_right gt_grand_summary_row gt_first_grand_summary_row gt_last_summary_row">11</td></tr>
#>   </tbody>
#>   
#>   
#> </table>
#> </div>