Validate referendums

Validate referendums

validate_referendums(
  data,
  check_applicability_constraint = TRUE,
  check_id_sudd_prefix = TRUE
)

Arguments

data
The referendum data to validate, as returned by referendums().

check_applicability_constraint
Whether or not to check that no applicability constraints as defined in the codebook are violated.

check_id_sudd_prefix
Whether or not to check that all id_sudd prefixes are valid.

Value

data, invisibly.