Add long country code to referendum data

Augments data with an additional column holding the three-letter ISO 3166-1 alpha-3 code of the country in which the referendum took place.

add_country_code_long(data)

Arguments

data
C2D referendum data as returned by rfrnds(). A data frame that at minimum contains the column country_code (with ISO 3166-1 alpha-2 or ISO 3166-3 alpha-4 codes).

Value

A tibble.

See also

Examples

c2d::rfrnds(date_max = clock::date_today(zone = "UTC"),
            date_min = clock::date_today(zone = "UTC") |> clock::add_years(-1L)) |>
  c2d:::add_country_code_long() |>
  dplyr::select(any_of("id"),
                starts_with("country_code"))
#> # A tibble: 83 × 4
#>    id                       country_code country_code_long country_code_histor…¹
#>    <chr>                    <chr>        <chr>             <chr>                
#>  1 62d7d8b3a52c3995043a8ad4 UY           URY               NA                   
#>  2 62d7bd48a52c3995043a8ace BY           BLR               NA                   
#>  3 62d6760ca52c3995043a8a1e CH           CHE               NA                   
#>  4 62d67203a52c3995043a8a16 CH           CHE               NA                   
#>  5 62d66e97a52c3995043a8a0f CH           CHE               NA                   
#>  6 62d66ce0a52c3995043a8a08 CH           CHE               NA                   
#>  7 62d016a313bed420d6ffbeb8 CH           CHE               NA                   
#>  8 62d0138d13bed420d6ffbeb5 CH           CHE               NA                   
#>  9 62d0025e13bed420d6ffbeb2 CH           CHE               NA                   
#> 10 62cfefee13bed420d6ffbead CH           CHE               NA                   
#> # … with 73 more rows, and abbreviated variable name ¹​country_code_historical