Add short country name to referendum data

Augments data with an additional column holding the common English name of the country in which the referendum took place.

Usage

add_country_name(data)

Arguments

data

C2D referendum data as returned by rfrnds(). A data frame that at minimum contains the column country_code (with ISO 3166-1 alpha-2 or ISO 3166-3 alpha-4 codes).

Value

A tibble.

See also

Examples

c2d::rfrnds(date_max = clock::date_today(zone = "UTC"),
            date_min = clock::date_today(zone = "UTC") |> clock::add_years(-1L)) |>
  c2d:::add_country_name() |>
  dplyr::select(any_of("id"),
                starts_with("country_"))
#> # A tibble: 110 × 3
#>    id                       country_code country_name  
#>    <chr>                    <fct>        <fct>         
#>  1 63bd5f7ca52c3995043a93f7 SI           Slovenia      
#>  2 63bd5ec0a52c3995043a93f3 SI           Slovenia      
#>  3 63bd5d92a52c3995043a93ee SI           Slovenia      
#>  4 63bd47b4a52c3995043a93ea TW           Taiwan        
#>  5 63bd3905a52c3995043a93e3 AS           American Samoa
#>  6 63bd2a63a52c3995043a93dd AS           American Samoa
#>  7 63bd293ba52c3995043a93d7 AS           American Samoa
#>  8 63bd268ba52c3995043a93d2 AS           American Samoa
#>  9 63bd25dea52c3995043a93cd AS           American Samoa
#> 10 63bd256ba52c3995043a93c7 AS           American Samoa
#> # … with 100 more rows