Add period to referendum data

Augments data with an additional column holding the specified period in which the referendum took place. The new column is named after period and its values are always of type integer.

add_period(
  data,
  period = c("week", "month", "quarter", "year", "decade", "century")
)

Arguments

data
C2D referendum data as returned by rfrnds(). A data frame that at minimum contains the column date.

period
Type of period to add. One of "week","month","quarter","year","decade" or "century".

Value

A tibble.

Details

Note that the period types "year", "decade" and "century" are unique timespans, while "week", "month" and "quarter" are recurring timespans (e.g. every year has a week 1).

See also

Examples

c2d::rfrnds(country_code = "AT") |>
  c2d::add_period() |>
  dplyr::select(id, date, week)
#> # A tibble: 11 × 3
#>    id                       date        week
#>    <chr>                    <date>     <int>
#>  1 5bbc025f92a21351232e5685 2013-01-20     3
#>  2 5bbc025f92a21351232e5683 2013-01-20     3
#>  3 5bbbf50592a21351232e2c93 1994-06-12    23
#>  4 5bbbe98892a21351232e0877 1978-11-05    44
#>  5 5bbbe2a492a21351232de709 1938-04-10    14
#>  6 5bbbe28792a21351232ddf99 1921-05-29    21
#>  7 5bbbe28792a21351232ddf93 1921-04-24    16
#>  8 5bbbe28592a21351232ddf1b 1920-10-10    41
#>  9 5bbbe28592a21351232ddf19 1920-10-10    41
#> 10 5bbbe28592a21351232ddf17 1920-10-10    41
#> 11 5bbbe28292a21351232dde3f 1919-05-11    19

c2d::rfrnds(country_code = "AT") |>
  c2d::add_period("year") |>
  dplyr::select(id, date, year)
#> # A tibble: 11 × 3
#>    id                       date        year
#>    <chr>                    <date>     <int>
#>  1 5bbc025f92a21351232e5685 2013-01-20  2013
#>  2 5bbc025f92a21351232e5683 2013-01-20  2013
#>  3 5bbbf50592a21351232e2c93 1994-06-12  1994
#>  4 5bbbe98892a21351232e0877 1978-11-05  1978
#>  5 5bbbe2a492a21351232de709 1938-04-10  1938
#>  6 5bbbe28792a21351232ddf99 1921-05-29  1921
#>  7 5bbbe28792a21351232ddf93 1921-04-24  1921
#>  8 5bbbe28592a21351232ddf1b 1920-10-10  1920
#>  9 5bbbe28592a21351232ddf19 1920-10-10  1920
#> 10 5bbbe28592a21351232ddf17 1920-10-10  1920
#> 11 5bbbe28292a21351232dde3f 1919-05-11  1919