Add new referendums to the C2D database

Adds new referendum entries to the C2D database via its API.

Usage

add_rfrnds(
  data,
  email = pal::pkg_config_val(key = "api_username", pkg = this_pkg),
  password = pal::pkg_config_val(key = "api_password", pkg = this_pkg),
  use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg),
  quiet = FALSE
)

Arguments

data

The new referendum data. A tibble that in any case must contain the columns

country_code

level

date

title_en

result

electorate_total

electorate_abroad

votes_yes

votes_no

votes_empty

votes_invalid

is_draft

type

plus the column subnational_entity_name for referendums of level below "national", and the column municipality for referendums of level = "local",

plus any additional valid columns containing the values for the corresponding database fields.

email

The e-mail address of the C2D API user account to be used for authentication. A character scalar.

password

The password of the C2D API user account to be used for authentication. A character scalar.

use_testing_server

Whether or not to use the testing servers instead of the production servers for C2D Services API calls etc.

quiet

Whether or not to print the newly created referendum IDs to console.

Value

A character vector of newly created referendum IDs.

Details

Note that adding/editing the column files is not supported, i.e. it is simply dropped from data.

See also