Add turnout to referendum data

Augments data with an additional column turnout containing the voter turnout calculated as:

votes_yes+votes_no+votes_empty+votes_invalidelectorate_total\frac{votes\_yes+votes\_no+votes\_empty+votes\_invalid}{electorate\_total}electorate_totalvotes_yes+votes_no+votes_empty+votes_invalid​

add_turnout(data, rough = TRUE, excl_dubious = TRUE)

Arguments

data
C2D referendum data as returned by rfrnds(). A data frame that at minimum contains the columns electorate_total, votes_yes, votes_no, votes_empty and votes_invalid.

rough
Whether to fall back on a "rough" calculation of the turnout in case any of the variables votes_empty or votes_invalid is unknown (NA), or to be strict and return NA in such a case.

excl_dubious
Whether or not to exclude obviously dubious turnout numbers (those > 1.0) by setting them to NA. Such numbers stem either from data errors or (officially) tampered numbers.

Value

A tibble.

See also

Examples

# rough turnout numbers
c2d::rfrnds(country_code = "IT") |>
  c2d::add_turnout() |>
  dplyr::select(any_of("id"),
                electorate_total,
                starts_with("votes_"),
                turnout)
#> # A tibble: 124 × 8
#>    id                    elect…¹ votes…² votes…³ votes…⁴ votes…⁵ votes…⁶ turnout
#>    <chr>                   <int>   <int>   <int>   <int>   <int> <list>    <dbl>
#>  1 5f6c6d3fd1291cc3961f…  5.10e7  1.79e7  7.69e6  218093  226569 <NULL>    0.511
#>  2 5d1cc245dc0aeb467bfc…  7.90e6  2.88e6  1.19e5      NA   23151 <NULL>    0.383
#>  3 5d1cba4edc0aeb467bfc…  4.07e6  2.27e6  4.39e4    5165    5865 <NULL>    0.572
#>  4 5d19ffd1fe65c2bfeba0…  5.08e7  1.34e7  1.94e7   83418  308728 <NULL>    0.655
#>  5 5cb86b29cb4865239961…  5.07e7  1.33e7  2.20e6  104413  168753 <NULL>    0.312
#>  6 5bbc039292a21351232e… NA      NA      NA           NA      NA <NULL>   NA    
#>  7 5bbc032492a21351232e… NA       4.77e5  2.63e5      NA      NA <NULL>   NA    
#>  8 5bbc032492a21351232e… NA       4.72e5  4.47e5      NA      NA <NULL>   NA    
#>  9 5bbc032492a21351232e… NA       4.65e5  3.69e5      NA      NA <NULL>   NA    
#> 10 5bbc032492a21351232e… NA       2.10e6  2.57e5      NA      NA <NULL>   NA    
#> # … with 114 more rows, and abbreviated variable names ¹​electorate_total,
#> #   ²​votes_yes, ³​votes_no, ⁴​votes_empty, ⁵​votes_invalid,
#> #   ⁶​votes_per_subterritory

# strict turnout numbers
c2d::rfrnds(country_code = "IT") |>
  c2d::add_turnout(rough = FALSE) |>
  dplyr::select(any_of("id"),
                electorate_total,
                starts_with("votes_"),
                turnout)
#> # A tibble: 124 × 8
#>    id                    elect…¹ votes…² votes…³ votes…⁴ votes…⁵ votes…⁶ turnout
#>    <chr>                   <int>   <int>   <int>   <int>   <int> <list>    <dbl>
#>  1 5f6c6d3fd1291cc3961f…  5.10e7  1.79e7  7.69e6  218093  226569 <NULL>    0.511
#>  2 5d1cc245dc0aeb467bfc…  7.90e6  2.88e6  1.19e5      NA   23151 <NULL>   NA    
#>  3 5d1cba4edc0aeb467bfc…  4.07e6  2.27e6  4.39e4    5165    5865 <NULL>    0.572
#>  4 5d19ffd1fe65c2bfeba0…  5.08e7  1.34e7  1.94e7   83418  308728 <NULL>    0.655
#>  5 5cb86b29cb4865239961…  5.07e7  1.33e7  2.20e6  104413  168753 <NULL>    0.312
#>  6 5bbc039292a21351232e… NA      NA      NA           NA      NA <NULL>   NA    
#>  7 5bbc032492a21351232e… NA       4.77e5  2.63e5      NA      NA <NULL>   NA    
#>  8 5bbc032492a21351232e… NA       4.72e5  4.47e5      NA      NA <NULL>   NA    
#>  9 5bbc032492a21351232e… NA       4.65e5  3.69e5      NA      NA <NULL>   NA    
#> 10 5bbc032492a21351232e… NA       2.10e6  2.57e5      NA      NA <NULL>   NA    
#> # … with 114 more rows, and abbreviated variable names ¹​electorate_total,
#> #   ²​votes_yes, ³​votes_no, ⁴​votes_empty, ⁵​votes_invalid,
#> #   ⁶​votes_per_subterritory