Add turnout to referendum data

Augments data with an additional column turnout containing the voter turnout calculated as (votes_yes + votes_no + votes_empty + votes_invalid) / electorate_total.

add_turnout(data, rough = TRUE, excl_dubious = TRUE)

Arguments

data
C2D referendum data as returned by rfrnds(). A data frame that at minimum contains the columns electorate_total, votes_yes, votes_no, votes_empty and votes_invalid.

rough
Whether to fall back on a "rough" calculation of the turnout in case any of the variables votes_empty or votes_invalid is unknown (NA), or to be strict and return NA in such a case.

excl_dubious
Whether or not to exclude obviously dubious turnout numbers (those > 1.0) by setting them to NA. Such numbers stem either from data errors or (officially) tampered numbers.

Value

A tibble.

See also

Examples

# rough turnout numbers
c2d::rfrnds(country_code = "IT") |>
  c2d::add_turnout() |>
  dplyr::select(id, electorate_total, starts_with("votes_"), turnout)
#> # A tibble: 128 × 8
#>    id              electorate_total votes_yes votes_no votes_empty votes_invalid
#>    <chr>                      <int>     <int>    <int>       <int>         <int>
#>  1 5f6c6d3fd1291c…         50955985  17913259  7691837      218093        226569
#>  2 5d1cc245dc0aeb…          7897056   2882531   119420          NA         23151
#>  3 5d1cba4edc0aeb…          4068560   2273985    43938        5165          5865
#>  4 5d19ffd1fe65c2…         50773284  13431087 19421025       83418        308728
#>  5 5cb86b29cb4865…         50681765  13334607  2198715      104413        168753
#>  6 5bbc039292a213…               NA        NA       NA          NA            NA
#>  7 5bbc032492a213…               NA    477312   263119          NA            NA
#>  8 5bbc032492a213…               NA    472409   447189          NA            NA
#>  9 5bbc032492a213…               NA    464534   369016          NA            NA
#> 10 5bbc032492a213…               NA   2102969   257266          NA            NA
#> # … with 118 more rows, and 2 more variables: votes_per_subterritory <list>,
#> #   turnout <dbl>

# strict turnout numbers
c2d::rfrnds(country_code = "IT") |>
  c2d::add_turnout(rough = FALSE) |>
  dplyr::select(id, electorate_total, starts_with("votes_"), turnout)
#> # A tibble: 128 × 8
#>    id              electorate_total votes_yes votes_no votes_empty votes_invalid
#>    <chr>                      <int>     <int>    <int>       <int>         <int>
#>  1 5f6c6d3fd1291c…         50955985  17913259  7691837      218093        226569
#>  2 5d1cc245dc0aeb…          7897056   2882531   119420          NA         23151
#>  3 5d1cba4edc0aeb…          4068560   2273985    43938        5165          5865
#>  4 5d19ffd1fe65c2…         50773284  13431087 19421025       83418        308728
#>  5 5cb86b29cb4865…         50681765  13334607  2198715      104413        168753
#>  6 5bbc039292a213…               NA        NA       NA          NA            NA
#>  7 5bbc032492a213…               NA    477312   263119          NA            NA
#>  8 5bbc032492a213…               NA    472409   447189          NA            NA
#>  9 5bbc032492a213…               NA    464534   369016          NA            NA
#> 10 5bbc032492a213…               NA   2102969   257266          NA            NA
#> # … with 118 more rows, and 2 more variables: votes_per_subterritory <list>,
#> #   turnout <dbl>