Add UN world regions

Augments data with information about the United Nations (UN) geoscheme on three different grouping tiers based on the UN M49 area code hierarchy.

In total, six different columns are added:

un_region_tier_1_code: UN tier-1 region's M49 area code

un_region_tier_1_name: UN tier-1 region's English name

un_region_tier_2_code: UN tier-2 region's M49 area code

un_region_tier_2_name: UN tier-2 region's English name

un_region_tier_3_code: UN tier-3 region's M49 area code

un_region_tier_3_name: UN tier-3 region's English name

Tier-1 regions are the highest, i.e. most aggregated UN regions, commonly referred to as continents. Tier-2 regions are also known as "subregions" and tier-3 regions as "sub-subregions".

add_world_regions(data)

Arguments

data
C2D referendum data as returned by rfrnds(). A data frame that at minimum contains the column country_code (with two-letter ISO 3166-1 alpha-2 codes).

Value

A tibble.

Details

Note that only part of all UN tier-2 regions are further divided into UN tier-3 regions, meaning that not all countries are part of a UN tier-3 region. If a country doesn't belong to any UN tier-3 region, the corresponding un_region_tier_3_* values will simply be NA.

Examples

c2d::rfrnd(id = "5bbbe26a92a21351232dd73f") |>
  add_world_regions() |>
  dplyr::select(country_name,
                starts_with("un_"))
#> # A tibble: 1 × 8
#>   country_name un_code un_region_tier_1_name un_region_tier_1_… un_region_tier_…
#>   <chr>        <chr>   <chr>                 <chr>              <chr>           
#> 1 Switzerland  756     Europe                150                Western Europe  
#> # … with 3 more variables: un_region_tier_2_code <chr>,
#> #   un_region_tier_3_name <chr>, un_region_tier_3_code <chr>