Add UN world regions to referendum data

Augments data with information about the United Nations (UN) geoscheme on three different grouping tiers based on the UN M49 area code hierarchy.

In total, seven different columns are added:

un_country_code: UN M49 country code

un_region_tier_1_code: UN tier-1 region's M49 area code

un_region_tier_1_name: UN tier-1 region's English name

un_region_tier_2_code: UN tier-2 region's M49 area code

un_region_tier_2_name: UN tier-2 region's English name

un_region_tier_3_code: UN tier-3 region's M49 area code

un_region_tier_3_name: UN tier-3 region's English name

Tier-1 regions are the highest, i.e. most aggregated UN regions, commonly referred to as continents. Tier-2 regions are also known as "subregions" and tier-3 regions as "sub-subregions".

add_world_regions(data)

Arguments

data
C2D referendum data as returned by rfrnds(). A data frame that at minimum contains the column country_code (with two-letter ISO 3166-1 alpha-2 codes).

Value

A tibble.

Details

Note that only part of all UN tier-2 regions are further divided into UN tier-3 regions, meaning that not all countries are part of a UN tier-3 region. If a country doesn't belong to any UN tier-3 region, the corresponding un_region_tier_3_* values will simply be NA.

See also

Examples

c2d::rfrnd(id = "5bbbe26a92a21351232dd73f") |>
  add_world_regions() |>
  dplyr::select(id, country_name, starts_with("un_"))
#> # A tibble: 1 × 9
#>   id             count…¹ un_co…² un_re…³ un_re…⁴ un_re…⁵ un_re…⁶ un_re…⁷ un_re…⁸
#>   <chr>          <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 5bbbe26a92a21… Switze… 756     Europe  150     Wester… 155     NA      NA     
#> # … with abbreviated variable names ¹​country_name, ²​un_country_code,
#> #   ³​un_region_tier_1_name, ⁴​un_region_tier_1_code, ⁵​un_region_tier_2_name,
#> #   ⁶​un_region_tier_2_code, ⁷​un_region_tier_3_name, ⁸​un_region_tier_3_code