Transform to ballot-date-level observations

Transforms referendum-level observations to ones on the level of ballot date and jurisdiction via nesting of referendum-level columns. The individual values of all the referendums on a specific ballot date in a specific jurisdiction are preserved in a list column named rfrnd_data.

as_ballot_dates(data)

Arguments

data
C2D referendum data as returned by rfrnds(). A data frame that at minimum contains the column date.

Value

A tibble.

See also

Examples

c2d::rfrnds(country_code = "AT") |> c2d::as_ballot_dates()
#> # A tibble: 9 × 8
#>   country_code country_name subnatio…¹ munic…² level date       is_fo…³ rfrnd_…⁴
#>   <chr>        <chr>        <chr>      <chr>   <ord> <date>     <lgl>   <list>  
#> 1 AT           Austria      NA         NA      nati… 2013-01-20 FALSE   <tibble>
#> 2 AT           Austria      NA         NA      nati… 1994-06-12 FALSE   <tibble>
#> 3 AT           Austria      NA         NA      nati… 1978-11-05 FALSE   <tibble>
#> 4 AT           Austria      NA         NA      nati… 1938-04-10 FALSE   <tibble>
#> 5 AT           Austria      Salzburg   NA      subn… 1921-05-29 FALSE   <tibble>
#> 6 AT           Austria      Tyrol      NA      subn… 1921-04-24 FALSE   <tibble>
#> 7 AT           Austria      Carinthia  NA      subn… 1920-10-10 FALSE   <tibble>
#> 8 AT           Austria      Klagenfur… NA      subn… 1920-10-10 FALSE   <tibble>
#> 9 AT           Austria      Vorarlberg NA      subn… 1919-05-11 FALSE   <tibble>
#> # … with abbreviated variable names ¹​subnational_entity_name, ²​municipality,
#> #   ³​is_former_country, ⁴​rfrnd_data