C2D Codebook

A tibble containing the complete metadata of all referendums() variables. The Codebook below is also available online.

data_codebook

Format

A tibble.

Details

Referendum-level variables

For the following variables, the unit of observation is the individual referendum.

id

A 24-character value that uniquely identifies the referendum and corresponds to the MongoDB ObjectId in the C2D database.

Level of measurement

This variable’s scale of measure is the nominal level.

id_official

The referendum’s official identifier assigned by the authorities; this is only guaranteed to be unique per respective jurisdiction.

Level of measurement

This variable’s scale of measure is the nominal level.

id_sudd

The referendum’s identifier assigned by sudd.ch; multiple referendum’s might share the same identifier.

Level of measurement

This variable’s scale of measure is the nominal level.

country_code

The ISO 3166-1 alpha-2 code identifying the country in which the referendum took place.

Level of measurement

This variable’s scale of measure is the nominal level.

country_name

The English name of the country in which the referendum took place.

Level of measurement

This variable’s scale of measure is the nominal level.

subnational_entity_name

The English name of the subnational political entity in which the referendum took place.

Applicability constraint

This variable is only applicable under the following condition:

level %in% c('local', 'subnational')

Level of measurement

This variable’s scale of measure is the nominal level.

municipality

The municipality in which the referendum took place.

Applicability constraint

This variable is only applicable under the following condition:

level == 'local'

Level of measurement

This variable’s scale of measure is the nominal level.

level

The political level at which the referendum took place.

Possible values

Value	Meaning

"local"	The referendum took place at the local political level.

"subnational"	The referendum took place at the subnational political level.

"national"	The referendum took place at the national political level.

Level of measurement

This variable’s scale of measure is the ordinal (ascending) level.

date

The date on which the referendum took place.

Level of measurement

This variable’s scale of measure is the interval level.

title_de

The referendum’s German title.

Level of measurement

This variable’s scale of measure is the nominal level.

title_fr

The referendum’s French title.

Level of measurement

This variable’s scale of measure is the nominal level.

title_en

The referendum’s English title.

Level of measurement

This variable’s scale of measure is the nominal level.

committee_name

The name of the committee that initiated the referendum.

Level of measurement

This variable’s scale of measure is the nominal level.

result

The result of the referendum.

Possible values

Value	Meaning

"yes"	The referendum proposal was approved.

"no"	The referendum proposal was rejected.

Level of measurement

This variable’s scale of measure is the nominal (binary) level.

subterritories_yes

The number of administrative subdivisions one level below level that approved the referendum.

Applicability constraint

This variable is only applicable under the following condition:

level != 'local'

Level of measurement

This variable’s scale of measure is the ratio level.

subterritories_no

The number of administrative subdivisions one level below level that rejected the referendum.

Applicability constraint

This variable is only applicable under the following condition:

level != 'local'

Level of measurement

This variable’s scale of measure is the ratio level.

electorate_total

The total number of people entitled to vote in the referendum.

Level of measurement

This variable’s scale of measure is the ratio level.

electorate_abroad

The number of people entitled to vote in the referendum who are living abroad; by definition a subset of electorate_total.

Level of measurement

This variable’s scale of measure is the ratio level.

votes_yes

The number of votes approving the referendum proposal.

Level of measurement

This variable’s scale of measure is the ratio level.

votes_no

The number of votes rejecting the referendum proposal.

Level of measurement

This variable’s scale of measure is the ratio level.

votes_empty

The number of votes that did not specify any preference on the referendum proposal.

Level of measurement

This variable’s scale of measure is the ratio level.

votes_invalid

The number of votes that were invalid.

Level of measurement

This variable’s scale of measure is the ratio level.

votes_per_subterritory

The number of votes_yes, votes_no, votes_empty and votes_invalid per administrative subdivision one level below level; currently only filled for Swiss referendums on the national level.

Applicability constraint

This variable is only applicable under the following condition:

level != 'local'

lower_house_yes

The number of parliamentarians that approved the referendum proposal in the corresponding parliamentary voting; in bicameral legislatures, this number includes only members of the lower house, in unicameral legislatures all parliamentarians; currently only filled for Swiss referendums on the national level (National Council).

Level of measurement

This variable’s scale of measure is the ratio level.

lower_house_no

The number of parliamentarians that rejected the referendum proposal in the corresponding parliamentary voting; in bicameral legislatures, this number includes only members of the lower house, in unicameral legislatures all parliamentarians; currently only filled for Swiss referendums on the national level (National Council).

Level of measurement

This variable’s scale of measure is the ratio level.

lower_house_abstentions

The number of parliamentarians that abstained from the corresponding parliamentary voting on the referendum proposal; in bicameral legislatures, this number includes only members of the lower house, in unicameral legislatures all parliamentarians; currently only filled for Swiss referendums on the national level (National Council).

Level of measurement

This variable’s scale of measure is the ratio level.

upper_house_yes

The number of parliamentarians that approved the referendum proposal in the corresponding parliamentary voting; applies only to bicameral legislatures and includes only members of the upper house; currently only filled for Swiss referendums on the national level (Council of States).

Level of measurement

This variable’s scale of measure is the ratio level.

upper_house_no

The number of parliamentarians that rejected the referendum proposal in the corresponding parliamentary voting; applies only to bicameral legislatures and includes only members of the upper house; currently only filled for Swiss referendums on the national level (Council of States).

Level of measurement

This variable’s scale of measure is the ratio level.

upper_house_abstentions

The number of parliamentarians that abstained from the corresponding parliamentary voting on the referendum proposal; applies only to bicameral legislatures and includes only members of the upper house; currently only filled for Swiss referendums on the national level (Council of States).

Level of measurement

This variable’s scale of measure is the ratio level.

position_government

The government’s recommendation on how to vote in the referendum.

Possible values

Value	Meaning

"yes"	The government recommends to approve the referendum.

"no"	The government recommends to reject the referendum.

Level of measurement

This variable’s scale of measure is the nominal (binary) level.

tags

A maximum of 3 predefined political topics that best classify the referendum.

Possible values

Value

"state organisation"

"national identity"

"political system"

"constitution"

"procedure for constitutional reform"

"legislative procedure"

"electoral system"

"institutions"

"government, administration"

"parliament"

"courts"

"national bank"

"citizens' rights"

"citizen's initiatives"

"referendum"

"voting rights"

"federalism"

"territoral questions"

"intergovernmental relations"

"division of tasks"

"legal system"

"international law"

"fundamental rights"

"citizenship"

"private law"

"criminal law"

"data protection"

"foreign policy"

"position on foreign policy"

"neutrality"

"independence"

"good offices"

"european policy"

"EFTA"

"EU"

"EEA"

"other european organisations"

"international organisations"

"UN"

"other international organisations"

"development cooperation"

"bilateral treaties"

"foreign trade policy"

"export promotion"

"customs"

"diplomacy"

"swiss abroad"

"security policy"

"public security"

"civil protection"

"state security"

"police"

"army"

"armed forces - general"

"military organisation"

"arms"

"military facilities"

"conscientious objection, civilian service"

"abolition of the armed forces"

"military training"

"international deployment"

"national economic supply"

"economy"

"economic policy"

"short-term economic policy"

"competition policy"

"structural policy"

"pricing policy"

"consumer protection"

"employment"

"employment conditions"

"working hours"

"industrial relations"

"employment policy"

"finance"

"monetary policy"

"banks, stock exchange, insurance"

"tourism and leisure"

"tourism"

"hospitality"

"lottery and gambling"

"agriculture"

"agricultural policy"

"livestock"

"crop production"

"forestry"

"fishing, hunting and pets"

"public finance"

"taxation"

"tax policy"

"tax system"

"direct taxation"

"indirect taxation"

"financial system"

"public expenditure"

"budget cuts and remediation measures"

"energy"

"energy policy"

"nuclear energy"

"hydro-electric power"

"alternative energy"

"oil and gas"

"transport and infrastructure"

"transport policy"

"urban transport"

"transit traffic"

"road transport"

"road construction"

"heavy traffic"

"rail transport"

"goods traffic"

"passenger traffic"

"air transport"

"shipping"

"post"

"telecommunications"

"environment and living space"

"soil"

"spatial planning"

"land law"

"housing"

"tenancy issues"

"building of residential housing, property ownership"

"environment"

"environmental policy"

"noise protection"

"air quality control"

"water pollution control"

"soil protection"

"waste"

"protection of nature and cultural heritage"

"animal protection"

"social policy"

"health"

"health policy"

"medical research and technology"

"medicines"

"addictive substances"

"reproductive medicine"

"social security"

"pension insurance"

"disability insurance"

"occupational benefits provision"

"health and accident insurance"

"maternity insurance"

"unemployment insurance"

"compensation for loss of earnings for persons on military service or civil protection duty"

"welfare"

"social groups"

"immigration policy"

"refugees"

"status of women"

"family policy"

"children and young people"

"senior citizens"

"persons with disabilities"

"homosexuals"

"education and research"

"education policy"

"schools"

"universities"

"research"

"genetic engineering"

"animal testing"

"professional and vocational education"

"culture, religion and media"

"cultural policy"

"language policy"

"religion - churches"

"sport"

"media and communication"

"media policy"

"press"

"radio, television and electronic media"

"freedom of the media"

This is a multi-value variable and thus of type list.

Level of measurement

This variable’s scale of measure is the nominal level.

remarks

A free text field with additional information about the referendum.

files

Metadata about files attached to the referendum, including

date_time_attached: the date and time on which the file was attached to the referendum’s data entry

is_deleted: whether or not the file has been deleted since it was attached

file_name: the filename as it appears when the file is downloaded

file_size: the file’s size in bytes

s3_object_key: the key uniquely identifying the file in the C2D Amazon S3 bucket; the file can be downloaded by feeding this key to c2d::download_file_attachment().

url_sudd

The URL of the corresponding entry on sudd.ch.

Level of measurement

This variable’s scale of measure is the nominal level.

url_swissvotes

The URL of the corresponding entry on swissvotes.ch.

Level of measurement

This variable’s scale of measure is the nominal level.

is_draft

Whether or not the entry is marked as a draft in the database.

Possible values

Value	Meaning

TRUE	The entry is marked as a draft.

FALSE	The entry is not marked as a draft.

Level of measurement

This variable’s scale of measure is the nominal (binary) level.

date_time_created

The date and time on which the referendum’s data entry was created.

Level of measurement

This variable’s scale of measure is the interval level.

archive

The referendum’s data from an earlier, obsolete state of the C2D database.

type

The institutional type of direct democracy which the referendum is an instance of.

Possible values

Value	Meaning

"mandatory referendum"	A referendum that was automatically triggered by certain legal conditions, usually found in the constitution.

"optional referendum"	A referendum against a law passed by the parliament that came about because the quorum of citizen opposition was met (usually a certain number of signatures) within a specified period after the resolution was passed.

"counter proposal"	A counter proposal by the government or parliament to another referendum.

"citizens' initiative"	A referendum launched by citizens, usually via a signature collection.

"governmental referendum"	A referendum launched by the government.

"citizens' assembly"	A citizens’ assembly, usually in person where people vote by raising their hands.

"recall"	A recall of all members of a political institution/body.

Level of measurement

This variable’s scale of measure is the nominal level.

Institution-level variables

For the following variables, the unit of observation is the referendum’s institutional instrument, a combination of the referendum’s type and the jurisdiction it took place.

The set of variables that constitute the jurisdiction is dependent on the referendum’s level:

level	Jurisdiction is given by

national	country_code

subnational	country_code + subnational_entity_name

local	country_code + subnational_entity_name + municipality

inst_legal_basis_type

The type of the legal basis of the referendum type.

Possible values

Value	Meaning

"non-official"	The referendum type has no legal basis.

"ad-hoc"	The referendum type has a legal basis which was specifically created for it.

"official"	The referendum type has a legal basis that wasn’t specifically created for it.

Level of measurement

This variable’s scale of measure is the ordinal (ascending) level.

inst_has_urgent_legal_basis

Whether or not the legal basis of the referendum type was adopted urgently, i.e. in an abridged procedure.

Possible values

Value	Meaning

TRUE	The legal basis of the referendum type was adopted in an abridged procedure (urgent).

FALSE	The legal basis of the referendum type was adopted in a normal procedure.

Applicability constraint

This variable is only applicable under the following condition:

trigger_actor %in% c('government', 'president', 'monarch', 'parliament')

Level of measurement

This variable’s scale of measure is the nominal (binary) level.

inst_is_binding

Whether the result of the referendum type has binding force or the decision of the voters may be disregarded.

Possible values

Value	Meaning

TRUE	The result of the referendum type is binding for the legislator.

FALSE	The result of the referendum type is only consultative, i.e. the legislator does not have to follow the decision of the voters.

Level of measurement

This variable’s scale of measure is the nominal (binary) level.

inst_is_counter_proposal

Whether or not the referendum type is a counterproposal to a citizen’s (or other) initiative.

Possible values

Value	Meaning

TRUE	The referendum type is a counterproposal to a citizen’s (or other) initiative.

FALSE	The referendum type is not a counterproposal to a citizen’s (or other) initiative.

Level of measurement

This variable’s scale of measure is the nominal (binary) level.

inst_is_assembly

Whether or not the referendum type is held as an assembly (all voters at the same time and place; physical presence at the assembly is required to be able to vote; usually voting means raising one’s hand).

Possible values

Value	Meaning

TRUE	The referendum type is a counterproposal to a citizen’s (or other) initiative.

FALSE	The referendum type is not a counterproposal to a citizen’s (or other) initiative.

Level of measurement

This variable’s scale of measure is the nominal (binary) level.

inst_trigger_type

The way the referendum type is triggered; completely dependent on inst_trigger_actor.

Possible values

Value	Meaning

"automatic"	The referendum type is triggered by a constitutional/legal requirement.

"top down"	The referendum type is triggered by an institution of the political elite like the monarch/president/government, the parliament, a territorial unit, the UN or another institution.

"bottom up"	The referendum type is triggered by citizen demand (e.g. a signature collection).

Level of measurement

This variable’s scale of measure is the nominal level.

inst_trigger_actor

The actor that triggers the referendum type.

Possible values

Value	Meaning

"citizens"	The referendum type is triggered by a direct citizen demand (usually a signature collection).

"constitution"	The referendum type is triggered by a constitutional requirement.

"law"	The referendum type is triggered by a legal requirement.

"parliament"	The referendum type is triggered by the parliament.

"government"	The referendum type is triggered by the government or prime minister.

"president"	The referendum type is triggered by the president.

"monarch"	The referendum type is triggered by the monarch.

"territorial unit"	The referendum type is triggered by a province or municipality.

"UN"	The referendum type is triggered by the United Nations (UN) or one of its organs.

"other institution"	The referendum type is triggered by another institution such as the church or a state agency.

Level of measurement

This variable’s scale of measure is the nominal level.

inst_trigger_actor_level

The political level of inst_trigger_actor in comparison to level.

Possible values

Value	Meaning

"lower"	A lower political level triggers a referendum on one of the higher political levels.

"equal"	The trigger actor and the referendum are on the same political level.

"higher"	A higher political level triggers a referendum on one of the lower political levels.

Level of measurement

This variable’s scale of measure is the ordinal (ascending) level.

inst_trigger_threshold

The threshold of support among inst_trigger_actor that has to be reached in order to trigger the referendum type.

Applicability constraint

This variable is only applicable under the following condition:

trigger_actor %in% c('citizens', 'parliament', 'territorial unit')

Level of measurement

This variable’s scale of measure is the nominal level.

inst_trigger_time_limit

The time limit in which inst_trigger_threshold (usually a minimum number of signatures) has to be reached in order to trigger the referendum type.

Applicability constraint

This variable is only applicable under the following condition:

trigger_actor %in% c('citizens', 'territorial unit')

Level of measurement

This variable’s scale of measure is the nominal level.

inst_object_type

The type of the object the referendum type refers to.

Possible values

Value	Meaning

"question"	A question which can be answered with yes or no (used as fallback option if nothing else applies).

"legal text (general proposal)"	A general proposal of a legal text, meaning the proposal is not formulated in a definite way, and therefore, if approved, needs to be implemented and concretized in a further legislative process.

"legal text (formulated proposal)"	A precisely formulated legal text which, if approved, will directly enter into force.

"institution"	A recall of all members of a particular institution/body.

"person"	An appointment of a single person.

"principle"	A question about what direction to take on a particular issue/topic, but without a proposed legal text; e.g., “Should there be a total revision of our constitution?”.

Level of measurement

This variable’s scale of measure is the nominal level.

inst_object_author

The actor that defines the text/content of inst_object_type.

Possible values

Value	Meaning

"citizens"	Citizens (usually a small group initiating a citizen demand).

"parliament"	The parliament.

"government"	The government or prime minister.

"president"	The president.

"monarch"	The monarch.

"territorial unit"	A province or municipality.

"UN"	The United Nations (UN) or one of its organs.

"other institution"	Another institution such as the church or a state agency.

Applicability constraint

This variable is only applicable under the following condition:

object_type %in% c('question', 'legal text (general proposal)', 'legal text (formulated proposal)', 'principle')

Level of measurement

This variable’s scale of measure is the nominal level.

inst_object_legal_level

The legal level of inst_object_type.

Possible values

Value	Meaning

"ordinance"	An ordinance, i.e. a legal norm based on a law or the constitution.

"treaty"	A domestic treaty, e.g. an agreement between the constituent states of a republic.

"law"	A law, i.e. a legal norm based on the constitution and created by the legislative.

"constitution"	The constitution, i.e. the aggregate of fundamental principles or established precedents that constitute the legal basis of a state and commonly determine how that state is to be governed.

"international treaty"	An international treaty, i.e. a bilateral or multilateral agreement or an accession to and international organization.

Applicability constraint

This variable is only applicable under the following condition:

object_type %in% c('legal text (general proposal)', 'legal text (formulated proposal)')

Level of measurement

This variable’s scale of measure is the nominal level.

inst_object_revision_extent

The extent of revision of inst_object_type which can be achieved within one referendum.

Possible values

Value	Meaning

"partial"	Only parts of the concerned legal text can be revised at the same time.

"total"	Only the entirety of the legal text can be changed at the same time.

"both"	Both, parts and the entirety of the legal text, can be changed at the same time.

Applicability constraint

This variable is only applicable under the following condition:

object_type %in% c('legal text (general proposal)', 'legal text (formulated proposal)')

Level of measurement

This variable’s scale of measure is the nominal level.

inst_object_revision_modes

The different modes of revision that inst_object_type allows.

Possible values

Value	Meaning

"introduce"	With this instrument, it is possible to introduce an entirely new text/concept.

"revise"	With this instrument, it is possible to change some parts of the legal text/…

"abrogate"	With this instrument, it is possible to strike parts of the legal text / to recall a person/institution.

This is a multi-value variable and thus of type list.

Level of measurement

This variable’s scale of measure is the nominal level.

inst_object_revision_divisibility

The divisibility of inst_object_type.

Possible values

Value	Meaning

"as a whole"	The proposed legal text can only be accepted/rejected as a whole.

"split up"	The proposed legal text can be split up, the voters may vote upon these parts separately.

"as variants"	The proposed legal text can be presented as two or more (alternative) variants.

"both split up and as variants"	The proposed legal text can be both split up as well as presented as two or more (alternative) variants.

Applicability constraint

This variable is only applicable under the following condition:

object_type %in% c('question', 'legal text (general proposal)', 'legal text (formulated proposal)')

Level of measurement

This variable’s scale of measure is the nominal level.

inst_topics_only

The legal topics to which the referendum type is limited.

Possible values

Value	Meaning

"infrastructural acts"	The referendum type is limited to TODO.

"territorial questions"	The referendum type is limited to TODO.

"competence shifts"	The referendum type is limited to TODO.

"financial acts"	The referendum type is limited to TODO.

"financial acts (expenses)"	The referendum type is limited to TODO.

"financial acts (taxes)"	The referendum type is limited to TODO.

"financial acts (obligations)"	The referendum type is limited to TODO.

"treaties"	The referendum type is limited to TODO (automatically limited to if inst_object_legal_level == "law").

"total revisions of the constitution"	The referendum type is limited to TODO.

This is a multi-value variable and thus of type list.

Applicability constraint

This variable is only applicable under the following condition:

inst_topics_excluded %>% purrr::map_lgl(rlang::is_empty)

Level of measurement

This variable’s scale of measure is the nominal level.

inst_topics_excluded

The legal topics that are excluded from the referendum type.

Possible values

Value	Meaning

"jus cogens"	TODO is exluded from the referendum type.

"international treaties"	TODO is exluded from the referendum type.

"pardons"	TODO is exluded from the referendum type.

"personnel appointments"	TODO is exluded from the referendum type.

"budgets"	TODO is exluded from the referendum type.

"financial matters"	TODO is exluded from the referendum type.

"everything within parliamentary competence"	TODO is exluded from the referendum type.

This is a multi-value variable and thus of type list.

Applicability constraint

This variable is only applicable under the following condition:

inst_topics_only %>% purrr::map_lgl(rlang::is_empty)

Level of measurement

This variable’s scale of measure is the nominal level.

inst_quorum_turnout

The minimum percentage of the electorate_total that have to participate in the referendum type in order for it to be passed.

Level of measurement

This variable’s scale of measure is the nominal level.

inst_quorum_approval

The minimum number of voters that have to approve the referendum type in order for it to be passed.

Possible values

Value	Meaning

"simple majority"	> 50 % of the voters have to approve the referendum proposal.

"qualified majority"	A percentage higher than > 50 % of the voters has to approve the referendum proposal.

"double majority"	Two simultaneous quorums of > 50 % or higher of the voters apply.

Level of measurement

This variable’s scale of measure is the nominal level.

inst_has_precondition

Whether or not an institutional actor must take a preconditional decision in order for the referendum type to be held.

Possible values

Value	Meaning

TRUE	An institutional actor must take a preconditional decision in order for the referendum type to be held.

FALSE	No institutional actor must take a preconditional decision in order for the referendum type to be held.

Level of measurement

This variable’s scale of measure is the nominal (binary) level.

inst_precondition_actor

The actor that must take a preconditional decision in order for the referendum type to be held.

Possible values

Value	Meaning

"parliament"	The parliament.

"government"	The government or prime minister.

"president"	The president.

"UN"	The United Nations (UN) or one of its organs.

"parliament and government"	The parliament and the government or prime minister.

"parliament and president"	The parliament and the president.

"central government and SPLM"	The central government and the Sudan People’s Liberation Movement.

Applicability constraint

This variable is only applicable under the following condition:

isTRUE(inst_has_precondition)

Level of measurement

This variable’s scale of measure is the nominal level.

inst_precondition_decision

The decision that inst_precondition_actor has taken.

Possible values

Value	Meaning

"yes"	Affirmative answer (e.g. yes, we will hold a vote that has been requested by citizens).

"no"	Negative answer (e.g. no, we will not implement the demands of a citizens’ initiative voluntarily).

Level of measurement

This variable’s scale of measure is the nominal (binary) level.

See also