List referendum data from sudd.ch

Lists the referendum data from sudd.ch in various ways its list.php endpoint allows. The output of this function can be directly fed to sudd_rfrnds().

list_sudd_rfrnds(
  mode = c("by_date", "by_mod_date", "filter", "random"),
  order = c("ascending", "descending"),
  filter = list(territory_name_de = NULL, title_de = NULL, year_min = NULL, year_max =
    NULL),
  use_cache = TRUE,
  cache_lifespan = "1 week"
)

Arguments

mode

The listing mode. One of

"by_date": Lists all referendums in the sudd.ch database by date, together with their id_sudd, country_code, territory_name_de and title_de. Specifying the sorting order of the results via the order parameter is supported.

"by_mod_date": Lists all referendums in the sudd.ch database by date_last_edited, together with their id_sudd, country_code, territory_name_de, date and title_de.

"filter": Allows to provide additional arguments (via the filter parameter) that limit the results accordingly. Specifying the sorting order of the results via the order parameter is supported.

"random": Lists the id_sudd of five randomly selected referendums, together with their country_code.

order

The sorting order of the results. Only relevant if mode is either "by_date" or "filter".

filter

A list with valid filtering arguments. Only relevant if mode = "filter". The supported filtering arguments include

territory_name_de: A string that must be (partially) matched by the referendums' territory_name_de. Matching is case-insensitive and no fuzzy matching is supported.

title_de: A string that must be (partially) matched by the referendums' title_de. Matching is case-insensitive and no fuzzy matching is supported.

year_min: The lower year limit of the referendums' date. A positive integer.

year_max: The upper year limit of the referendums' date. A positive integer.

use_cache

Return cached results if possible. If FALSE, results are always newly fetched regardless of cache_lifespan.

cache_lifespan

The duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Only relevant if use_cache = TRUE.

Value

A tibble containing at least an id_sudd column.

About sudd.ch

sudd stands for Suchmaschine für direkte Demokratie (German) and is operated by Beat Müller. Its database content is licensed under Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International (CC BY-NC-SA 4.0).

See also

Examples

# list all referendums by modification date (takes a while)
if (FALSE) {
c2d::list_sudd_rfrnds(mode = "by_mod_date")}

# list all referendums whose title matches "AHV"
c2d::list_sudd_rfrnds(mode = "filter",
                      filter = list(title_de = "AHV"))
#> ℹ Fetching raw HTML data from <sudd.ch>...
#> ✔ Fetching raw HTML data from <sudd.ch>... done [242ms]
#> 
#> ℹ Parsing and tidying raw HTML data...
#> ✔ Parsing and tidying raw HTML data... done [620ms]
#> 
#> # A tibble: 17 × 11
#>    id_sudd  count…¹ count…² is_fo…³ subna…⁴ terri…⁵ date        year month   day
#>    <chr>    <chr>   <chr>   <lgl>   <chr>   <chr>   <date>     <int> <int> <int>
#>  1 ch021978 CH      Switze… FALSE   NA      Schweiz 1978-02-26  1978     2    26
#>  2 ch031978 CH      Switze… FALSE   NA      Schweiz 1978-02-26  1978     2    26
#>  3 ch021988 CH      Switze… FALSE   NA      Schweiz 1988-06-12  1988     6    12
#>  4 ch051995 CH      Switze… FALSE   NA      Schweiz 1995-06-25  1995     6    25
#>  5 ch061995 CH      Switze… FALSE   NA      Schweiz 1995-06-25  1995     6    25
#>  6 ch061998 CH      Switze… FALSE   NA      Schweiz 1998-09-27  1998     9    27
#>  7 ch082001 CH      Switze… FALSE   NA      Schweiz 2001-12-02  2001    12     2
#>  8 ch052002 CH      Switze… FALSE   NA      Schweiz 2002-09-22  2002     9    22
#>  9 ch042004 CH      Switze… FALSE   NA      Schweiz 2004-05-16  2004     5    16
#> 10 ch052004 CH      Switze… FALSE   NA      Schweiz 2004-05-16  2004     5    16
#> 11 ch022006 CH      Switze… FALSE   NA      Schweiz 2006-09-24  2006     9    24
#> 12 ch072008 CH      Switze… FALSE   NA      Schweiz 2008-11-30  2008    11    30
#> 13 ch112016 CH      Switze… FALSE   NA      Schweiz 2016-09-25  2016     9    25
#> 14 ch062017 CH      Switze… FALSE   NA      Schweiz 2017-09-24  2017     9    24
#> 15 ch022019 CH      Switze… FALSE   NA      Schweiz 2019-05-19  2019     5    19
#> 16 ch092022 CH      Switze… FALSE   NA      Schweiz 2022-09-25  2022     9    25
#> 17 ch102022 CH      Switze… FALSE   NA      Schweiz 2022-09-25  2022     9    25
#> # … with 1 more variable: title_de <chr>, and abbreviated variable names
#> #   ¹​country_code, ²​country_name, ³​is_former_country, ⁴​subnational_entity_code,
#> #   ⁵​territory_name_de

# get sudd.ch referendum data from all referendums from 2020 onwards
c2d::list_sudd_rfrnds(mode = "filter",
                      filter = list(year_min = 2020)) |>
  c2d::sudd_rfrnds()
#> ℹ Fetching raw HTML data from <sudd.ch>...
#> ✔ Fetching raw HTML data from <sudd.ch>... done [148ms]
#> 
#> ℹ Parsing and tidying raw HTML data...
#> ✔ Parsing and tidying raw HTML data... done [1.9s]
#> 
#> Scraping referendum data from sudd.ch ■■■■■■■■■■■■                      36% | E…
#> Scraping referendum data from sudd.ch ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■     93% | E…
#> # A tibble: 58 × 37
#>    id_sudd  country_c…¹ count…² is_fo…³ subna…⁴ terri…⁵ terri…⁶ date        year
#>    <chr>    <chr>       <chr>   <lgl>   <chr>   <chr>   <chr>   <date>     <int>
#>  1 ch012020 CH          Switze… FALSE   NA      Schweiz unabhä… 2020-02-09  2020
#>  2 ch022020 CH          Switze… FALSE   NA      Schweiz unabhä… 2020-02-09  2020
#>  3 fr012020 FR          France  FALSE   NA      Frankr… unabhä… 2020-03-15  2020
#>  4 gn012020 GN          Guinea  FALSE   NA      Guinea  unabhä… 2020-03-22  2020
#>  5 pw012020 PW          Palau   FALSE   NA      Palau   unabhä… 2020-05-01  2020
#>  6 ru012020 RU          Russia… FALSE   NA      Russla… unabhä… 2020-07-01  2020
#>  7 ru022020 RU          Russia… FALSE   NA      Russla… unabhä… 2020-07-01  2020
#>  8 ru032020 RU          Russia… FALSE   NA      Russla… unabhä… 2020-07-01  2020
#>  9 ru042020 RU          Russia… FALSE   NA      Russla… unabhä… 2020-07-01  2020
#> 10 ru052020 RU          Russia… FALSE   NA      Russla… unabhä… 2020-07-01  2020
#> # … with 48 more rows, 28 more variables: month <int>, day <int>,
#> #   title_de <chr>, question_type_de <chr>, types <list>, result_de <chr>,
#> #   result_status_de <chr>, adoption_requirements_de <list>,
#> #   electorate_total <int>, electorate_abroad <int>, polling_cards <int>,
#> #   votes_total <int>, votes_empty <int>, votes_void <int>,
#> #   votes_invalid <int>, votes_valid <int>, votes_yes <int>, votes_no <int>,
#> #   votes_option_1 <int>, votes_option_2 <int>, subterritories <dbl>, …

# get sudd.ch referendum data from five randomly picked referendums
c2d::list_sudd_rfrnds(mode = "random") |>
  c2d::sudd_rfrnds()
#> ℹ Fetching raw HTML data from <sudd.ch>...
#> ✔ Fetching raw HTML data from <sudd.ch>... done [168ms]
#> 
#> ℹ Parsing and tidying raw HTML data...
#> ✔ Parsing and tidying raw HTML data... done [166ms]
#> 
#> # A tibble: 5 × 48
#>   id_sudd count…¹ count…² is_fo…³ subna…⁴ terri…⁵ terri…⁶ date        year month
#>   <chr>   <chr>   <chr>   <lgl>   <chr>   <chr>   <chr>   <date>     <int> <int>
#> 1 nz0119… NZ      New Ze… FALSE   NA      Neusee… abhäng… 1925-11-04  1925    11
#> 2 tg0119… TG      Togo    FALSE   NA      Autono… Treuha… 1956-10-28  1956    10
#> 3 ch0119… CH      Switze… FALSE   NA      Schweiz unabhä… 1987-04-05  1987     4
#> 4 it0919… IT      Italy   FALSE   NA      Italien unabhä… 1995-06-11  1995     6
#> 5 li0120… LI      Liecht… FALSE   NA      Liecht… unabhä… 2005-11-27  2005    11
#> # … with 38 more variables: day <int>, title_de <chr>, question_type_de <chr>,
#> #   types <list>, result_de <chr>, result_status_de <chr>,
#> #   adoption_requirements_de <list>, electorate_total <int>,
#> #   electorate_abroad <int>, polling_cards <int>, votes_total <int>,
#> #   votes_empty <int>, votes_void <int>, votes_invalid <int>,
#> #   votes_valid <int>, votes_yes <int>, votes_no <int>, votes_option_1 <int>,
#> #   votes_option_2 <int>, votes_option_3 <int>, votes_proposal_invalid <int>, …