List referendum data from sudd.ch

Lists the referendum data from sudd.ch in various ways its list.php endpoint allows. The output of this function can be directly fed to sudd_rfrnds().

Usage

list_sudd_rfrnds(
  mode = c("by_date", "by_mod_date", "filter", "random"),
  order = c("ascending", "descending"),
  filter = list(territory_name_de = NULL, title_de = NULL, year_min = NULL, year_max =
    NULL),
  use_cache = TRUE,
  cache_lifespan = "1 week"
)

Arguments

mode

The listing mode. One of

"by_date": Lists all referendums in the sudd.ch database by date, together with their id_sudd, country_code, territory_name_de and title_de. Specifying the sorting order of the results via the order parameter is supported.

"by_mod_date": Lists all referendums in the sudd.ch database by date_last_edited, together with their id_sudd, country_code, territory_name_de, date and title_de.

"filter": Allows to provide additional arguments (via the filter parameter) that limit the results accordingly. Specifying the sorting order of the results via the order parameter is supported.

"random": Lists the id_sudd of five randomly selected referendums, together with their country_code.

order

The sorting order of the results. Only relevant if mode is either "by_date" or "filter".

filter

A list with valid filtering arguments. Only relevant if mode = "filter". The supported filtering arguments include

territory_name_de: A string that must be (partially) matched by the referendums' territory_name_de. Matching is case-insensitive and no fuzzy matching is supported.

title_de: A string that must be (partially) matched by the referendums' title_de. Matching is case-insensitive and no fuzzy matching is supported.

year_min: The lower year limit of the referendums' date. A positive integer.

year_max: The upper year limit of the referendums' date. A positive integer.

use_cache

Return cached results if possible. If FALSE, results are always newly fetched regardless of cache_lifespan.

cache_lifespan

The duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Only relevant if use_cache = TRUE.

Value

A tibble containing at least an id_sudd column.

About sudd.ch

sudd stands for Suchmaschine für direkte Demokratie (German) and is operated by Beat Müller. Its database content is licensed under Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International (CC BY-NC-SA 4.0).

See also

Examples

# list all referendums by modification date (takes a while)
if (FALSE) {
c2d::list_sudd_rfrnds(mode = "by_mod_date")}

# list all referendums whose title matches "AHV"
c2d::list_sudd_rfrnds(mode = "filter",
                      filter = list(title_de = "AHV"))
#> # A tibble: 17 × 11
#>    id_sudd  country_code country_name is_former_country subnational_entity_code territory_name_de date        year month   day title_de                         
#>    <chr>    <chr>        <chr>        <lgl>             <chr>                   <chr>             <date>     <int> <int> <int> <chr>                            
#>  1 ch021978 CH           Switzerland  FALSE             NA                      Schweiz           1978-02-26  1978     2    26 Alters- und Hinterlassenenversic…
#>  2 ch031978 CH           Switzerland  FALSE             NA                      Schweiz           1978-02-26  1978     2    26 Herabsetzung des AHV-Alters      
#>  3 ch021988 CH           Switzerland  FALSE             NA                      Schweiz           1988-06-12  1988     6    12 Herabsetzung des AHV-Rentenalter…
#>  4 ch051995 CH           Switzerland  FALSE             NA                      Schweiz           1995-06-25  1995     6    25 Alters- und Hinterlassenenversic…
#>  5 ch061995 CH           Switzerland  FALSE             NA                      Schweiz           1995-06-25  1995     6    25 Ausbau von AHV und IV            
#>  6 ch061998 CH           Switzerland  FALSE             NA                      Schweiz           1998-09-27  1998     9    27 10. AHV-Revision ohne Erhöhung d…
#>  7 ch082001 CH           Switzerland  FALSE             NA                      Schweiz           2001-12-02  2001    12     2 Sicherung der AHV durch Besteuer…
#>  8 ch052002 CH           Switzerland  FALSE             NA                      Schweiz           2002-09-22  2002     9    22 Überschüssige Goldreserven in de…
#>  9 ch042004 CH           Switzerland  FALSE             NA                      Schweiz           2004-05-16  2004     5    16 Alters- und Hinterlassenenversic…
#> 10 ch052004 CH           Switzerland  FALSE             NA                      Schweiz           2004-05-16  2004     5    16 Finanzierung der AHV/IV durch An…
#> 11 ch022006 CH           Switzerland  FALSE             NA                      Schweiz           2006-09-24  2006     9    24 Nationalbankgewinne für die AHV  
#> 12 ch072008 CH           Switzerland  FALSE             NA                      Schweiz           2008-11-30  2008    11    30 Flexibles AHV-Alter              
#> 13 ch112016 CH           Switzerland  FALSE             NA                      Schweiz           2016-09-25  2016     9    25 Höhere AHV-Renten (AHV plus)     
#> 14 ch062017 CH           Switzerland  FALSE             NA                      Schweiz           2017-09-24  2017     9    24 Zusatzfinanzierung der AHV durch…
#> 15 ch022019 CH           Switzerland  FALSE             NA                      Schweiz           2019-05-19  2019     5    19 Steuerreform und AHV-Finanzierung
#> 16 ch092022 CH           Switzerland  FALSE             NA                      Schweiz           2022-09-25  2022     9    25 Zusatzfinanzierung der AHV durch…
#> 17 ch102022 CH           Switzerland  FALSE             NA                      Schweiz           2022-09-25  2022     9    25 AHV-Gesetz (AHV 21)              

# get sudd.ch referendum data from all referendums from 2020 onwards
c2d::list_sudd_rfrnds(mode = "filter",
                      filter = list(year_min = 2020)) |>
  c2d::sudd_rfrnds()
#> # A tibble: 58 × 37
#>    id_sudd  country…¹ count…² is_fo…³ subna…⁴ terri…⁵ terri…⁶ date        year month   day title…⁷ quest…⁸ types resul…⁹ resul…˟ adopt…˟ elect…˟ elect…˟ polli…˟
#>    <chr>    <chr>     <chr>   <lgl>   <chr>   <chr>   <chr>   <date>     <int> <int> <int> <chr>   <chr>   <lis> <chr>   <chr>   <list>    <int>   <int>   <int>
#>  1 ch012020 CH        Switze… FALSE   NA      Schweiz unabhä… 2020-02-09  2020     2     9 Mehr b… Entsch… <chr> verwor… Enderg… <chr>    5.47e6  188259 NA     
#>  2 ch022020 CH        Switze… FALSE   NA      Schweiz unabhä… 2020-02-09  2020     2     9 Strafg… Entsch… <chr> angeno… Enderg… <chr>    5.47e6  188259 NA     
#>  3 fr012020 FR        France  FALSE   NA      Frankr… unabhä… 2020-03-15  2020     3    15 Rückzu… Entsch… <chr> angeno… Enderg… <chr>   NA           NA NA     
#>  4 gn012020 GN        Guinea  FALSE   NA      Guinea  unabhä… 2020-03-22  2020     3    22 Verfas… Entsch… <chr> angeno… Enderg… <chr>    5.03e6      NA NA     
#>  5 pw012020 PW        Palau   FALSE   NA      Palau   unabhä… 2020-05-01  2020     5     1 Festle… Entsch… <chr> angeno… vorläu… <chr>    1.45e4      NA NA     
#>  6 ru012020 RU        Russia… FALSE   NA      Russla… unabhä… 2020-07-01  2020     7     1 Verfas… Entsch… <chr> angeno… Enderg… <chr>    1.09e8      NA  7.42e7
#>  7 ru022020 RU        Russia… FALSE   NA      Russla… unabhä… 2020-07-01  2020     7     1 Staats… Entsch… <chr> angeno… Zahlen… <chr>    1.09e8      NA NA     
#>  8 ru032020 RU        Russia… FALSE   NA      Russla… unabhä… 2020-07-01  2020     7     1 Bodens… Entsch… <chr> angeno… Zahlen… <chr>    1.09e8      NA NA     
#>  9 ru042020 RU        Russia… FALSE   NA      Russla… unabhä… 2020-07-01  2020     7     1 Tiefer… Entsch… <chr> angeno… Zahlen… <chr>    1.09e8      NA NA     
#> 10 ru052020 RU        Russia… FALSE   NA      Russla… unabhä… 2020-07-01  2020     7     1 Indexi… Entsch… <chr> angeno… Zahlen… <chr>    1.09e8      NA NA     
#> # … with 48 more rows, 17 more variables: votes_total <int>, votes_empty <int>, votes_void <int>, votes_invalid <int>, votes_valid <int>, votes_yes <int>,
#> #   votes_no <int>, votes_option_1 <int>, votes_option_2 <int>, subterritories <dbl>, subterritories_yes <dbl>, subterritories_no <dbl>, files <list>,
#> #   remarks <list>, sources <list>, ids_sudd_simultaneous <list>, date_last_edited <date>, and abbreviated variable names ¹​country_code, ²​country_name,
#> #   ³​is_former_country, ⁴​subnational_entity_code, ⁵​territory_name_de, ⁶​territory_type_de, ⁷​title_de, ⁸​question_type_de, ⁹​result_de, ˟​result_status_de,
#> #   ˟​adoption_requirements_de, ˟​electorate_total, ˟​electorate_abroad, ˟​polling_cards

# get sudd.ch referendum data from five randomly picked referendums
c2d::list_sudd_rfrnds(mode = "random") |>
  c2d::sudd_rfrnds()
#> ℹ Fetching raw HTML data from <sudd.ch>...
#> ✔ Fetching raw HTML data from <sudd.ch>... done [259ms]
#> 
#> ℹ Parsing and tidying raw HTML data...
#> ✔ Parsing and tidying raw HTML data... done [236ms]
#> 
#> # A tibble: 5 × 28
#>   id_sudd  country_…¹ count…² is_fo…³ subna…⁴ terri…⁵ terri…⁶ date        year month   day title…⁷ quest…⁸ types resul…⁹ resul…˟ adopt…˟ elect…˟ votes…˟ votes…˟
#>   <chr>    <chr>      <chr>   <lgl>   <chr>   <chr>   <chr>   <date>     <int> <int> <int> <chr>   <chr>   <lis> <chr>   <chr>   <list>    <int>   <int>   <int>
#> 1 gu131987 GU         Guam    FALSE   NA      Guam (… abhäng… 1987-08-08  1987     8     8 Ausgab… Entsch… <chr> verwor… Enderg… <chr>     35846   13841     727
#> 2 ge012001 GEHH       Gilber… TRUE    NA      Südoss… unabhä… 2001-04-08  2001     4     8 Verfas… Entsch… <chr> angeno… Zahlen… <chr>     45000   23540      NA
#> 3 cg012002 CG         Congo   FALSE   NA      Republ… unabhä… 2002-01-20  2002     1    20 Verfas… Entsch… <chr> angeno… Enderg… <chr>   1701415 1318238   49908
#> 4 ie012009 IE         Ireland FALSE   NA      Irland  unabhä… 2009-10-02  2009    10     2 Vertra… Entsch… <chr> angeno… Enderg… <chr>   3078032 1816098    7224
#> 5 ua032022 UA         Ukraine FALSE   NA      Region… unabhä… 2022-09-27  2022     9    27 Anschl… Entsch… <chr> angeno… Enderg… <chr>        NA  571001    5118
#> # … with 8 more variables: votes_valid <int>, votes_yes <int>, votes_no <int>, files <list>, remarks <list>, sources <list>, ids_sudd_simultaneous <list>,
#> #   date_last_edited <date>, and abbreviated variable names ¹​country_code, ²​country_name, ³​is_former_country, ⁴​subnational_entity_code, ⁵​territory_name_de,
#> #   ⁶​territory_type_de, ⁷​title_de, ⁸​question_type_de, ⁹​result_de, ˟​result_status_de, ˟​adoption_requirements_de, ˟​electorate_total, ˟​votes_total, ˟​votes_invalid