List referendum data from sudd.ch

Lists the referendum data from sudd.ch in various ways its list.php endpoint allows. The output of this function can be directly fed to sudd_rfrnds().

list_sudd_rfrnds(
  mode = c("by_date", "by_mod_date", "filter", "random"),
  order = c("ascending", "descending"),
  filter = list(territory_name_de = NULL, title_de = NULL, year_min = NULL, year_max =
    NULL),
  use_cache = TRUE,
  cache_lifespan = "1 week"
)

Arguments

mode
The listing mode. One of

"by_date": Lists all referendums in the sudd.ch database by date, together with their id_sudd, country_code, territory_name_de and title_de. Specifying the sorting order of the results via the order parameter is supported.

"by_mod_date": Lists all referendums in the sudd.ch database by date_last_edited, together with their id_sudd, country_code, territory_name_de, date and title_de.

"filter": Allows to provide additional arguments (via the filter parameter) that limit the results accordingly. Specifying the sorting order of the results via the order parameter is supported.

"random": Lists the id_sudd of five randomly selected referendums, together with their country_code.

order
The sorting order of the results. Only relevant if mode is either "by_date" or "filter".

filter
A list with valid filtering arguments. Only relevant if mode = "filter". The supported filtering arguments include

territory_name_de: A string that must be (partially) matched by the referendums' territory_name_de. Matching is case-insensitive and no fuzzy matching is supported.

title_de: A string that must be (partially) matched by the referendums' title_de. Matching is case-insensitive and no fuzzy matching is supported.

year_min: The lower year limit of the referendums' date. A positive integer.

year_max: The upper year limit of the referendums' date. A positive integer.

use_cache
Return cached results if possible. If FALSE, results are always newly fetched regardless of cache_lifespan.

cache_lifespan
The duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Only relevant if use_cache = TRUE.

Value

A tibble containing at least an id_sudd column.

About sudd.ch

sudd stands for Suchmaschine für direkte Demokratie (German) and is operated by Beat Müller. Its database content is licensed under Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International (CC BY-NC-SA 4.0).

See also

Examples

# list all referendums by modification date
c2d::list_sudd_rfrnds(mode = "by_mod_date")
#> # A tibble: 3,067 × 12
#>    id_sudd  country_code country_name  country_code_historical is_past_jurisdic…
#>    <chr>    <chr>        <chr>         <chr>                   <lgl>            
#>  1 li012022 LI           Liechtenstein NA                      FALSE            
#>  2 it022022 IT           Italy         NA                      FALSE            
#>  3 it032022 IT           Italy         NA                      FALSE            
#>  4 it042022 IT           Italy         NA                      FALSE            
#>  5 it052022 IT           Italy         NA                      FALSE            
#>  6 it062022 IT           Italy         NA                      FALSE            
#>  7 in011951 IN           India         NA                      FALSE            
#>  8 uz012002 UZ           Uzbekistan    NA                      FALSE            
#>  9 uz022002 UZ           Uzbekistan    NA                      FALSE            
#> 10 lt011991 LT           Lithuania     NA                      FALSE            
#> # … with 3,057 more rows, and 7 more variables: territory_name_de <chr>,
#> #   date <date>, year <int>, month <int>, day <int>, title_de <chr>,
#> #   date_last_edited <date>

# list all referendums whose title matches "AHV"
c2d::list_sudd_rfrnds(mode = "filter",
                      filter = list(title_de = "AHV"))
#> # A tibble: 15 × 11
#>    id_sudd  country_code country_name country_code_historical is_past_jurisdict…
#>    <chr>    <chr>        <chr>        <chr>                   <lgl>             
#>  1 ch021978 CH           Switzerland  NA                      FALSE             
#>  2 ch031978 CH           Switzerland  NA                      FALSE             
#>  3 ch021988 CH           Switzerland  NA                      FALSE             
#>  4 ch051995 CH           Switzerland  NA                      FALSE             
#>  5 ch061995 CH           Switzerland  NA                      FALSE             
#>  6 ch061998 CH           Switzerland  NA                      FALSE             
#>  7 ch082001 CH           Switzerland  NA                      FALSE             
#>  8 ch052002 CH           Switzerland  NA                      FALSE             
#>  9 ch042004 CH           Switzerland  NA                      FALSE             
#> 10 ch052004 CH           Switzerland  NA                      FALSE             
#> 11 ch022006 CH           Switzerland  NA                      FALSE             
#> 12 ch072008 CH           Switzerland  NA                      FALSE             
#> 13 ch112016 CH           Switzerland  NA                      FALSE             
#> 14 ch062017 CH           Switzerland  NA                      FALSE             
#> 15 ch022019 CH           Switzerland  NA                      FALSE             
#> # … with 6 more variables: territory_name_de <chr>, date <date>, year <int>,
#> #   month <int>, day <int>, title_de <chr>

# get sudd.ch referendum data from all referendums from 2020 onwards
c2d::list_sudd_rfrnds(mode = "filter",
                      filter = list(year_min = 2020)) |>
  c2d::sudd_rfrnds()
#> # A tibble: 119 × 37
#>    id_sudd  country_code country_code_historical country_name   territory_name_…
#>    <chr>    <chr>        <chr>                   <chr>          <chr>           
#>  1 ch012020 CH           NA                      Switzerland    Schweiz         
#>  2 ch022020 CH           NA                      Switzerland    Schweiz         
#>  3 fr012020 FR           NA                      France         Frankreich      
#>  4 gn012020 GN           NA                      Guinea         Guinea          
#>  5 pw012020 PW           NA                      Palau          Palau           
#>  6 ru012020 RU           NA                      Russian Feder… Russland        
#>  7 ru022020 RU           NA                      Russian Feder… Russland        
#>  8 ru032020 RU           NA                      Russian Feder… Russland        
#>  9 ru042020 RU           NA                      Russian Feder… Russland        
#> 10 ru052020 RU           NA                      Russian Feder… Russland        
#> # … with 109 more rows, and 32 more variables: territory_type_de <chr>,
#> #   date <date>, year <int>, month <int>, day <int>, title_de <chr>,
#> #   question_type_de <chr>, types <list>, result_de <chr>,
#> #   result_status_de <chr>, adoption_requirements_de <list>,
#> #   electorate_total <int>, electorate_abroad <int>, polling_cards <int>,
#> #   votes_total <int>, votes_empty <int>, votes_void <int>,
#> #   votes_invalid <int>, votes_valid <int>, votes_yes <int>, votes_no <int>, …

# get sudd.ch referendum data from five randomly picked referendums
c2d::list_sudd_rfrnds(mode = "random") |>
  c2d::sudd_rfrnds()
#> ℹ Fetching raw HTML data from <sudd.ch>...
#> ✔ Fetching raw HTML data from <sudd.ch>... done [394ms]
#> 
#> ℹ Parsing and tidying raw HTML data...
#> ✔ Parsing and tidying raw HTML data... done [96ms]
#> 
#> # A tibble: 5 × 33
#>   id_sudd  country_code country_code_historical country_name territory_name_de
#>   <chr>    <chr>        <chr>                   <chr>        <chr>            
#> 1 ch091866 CH           NA                      Switzerland  Schweiz          
#> 2 ch031900 CH           NA                      Switzerland  Schweiz          
#> 3 ee021917 EE           NA                      Estonia      Narva (Estland)  
#> 4 ch072003 CH           NA                      Switzerland  Schweiz          
#> 5 ie012015 IE           NA                      Ireland      Irland           
#> # … with 28 more variables: territory_type_de <chr>, date <date>, year <int>,
#> #   month <int>, day <int>, title_de <chr>, question_type_de <chr>,
#> #   types <list>, result_de <chr>, result_status_de <chr>,
#> #   adoption_requirements_de <list>, electorate_total <int>,
#> #   electorate_abroad <int>, votes_total <int>, votes_empty <int>,
#> #   votes_void <int>, votes_invalid <int>, votes_valid <int>, votes_yes <int>,
#> #   votes_no <int>, subterritories <dbl>, subterritories_yes <dbl>, …