Count number of referendums per period

Counts the number of C2D referendums per desired period, optionally by additional columns specified via by.

Usage

n_rfrnds_per_period(
  data,
  by = NULL,
  period = c("week", "month", "quarter", "year", "decade", "century"),
  fill_gaps = TRUE,
  period_floor = NULL,
  period_ceiling = NULL,
  descending = FALSE
)

Arguments

data

C2D referendum data as returned by rfrnds(). A data frame that at minimum contains the column specified in period or the column date (to compute the period column), plus the one(s) specified via by (if any).

by

Optional data column(s) to group by before counting number of referendums. Tidy selections are supported.

period

Type of period to count referendums by. One of "week","month","quarter","year","decade" or "century".

fill_gaps

Whether or not to add zero-value rows to the result for period gaps in data.

period_floor

Lower period limit up to which gaps are filled. If NULL, the lower limit is set to the minimum of period present in data. Only relevant if fill_gaps = TRUE and period is set to a unique timespan type ("year", "decade" or "century").

period_ceiling

Upper period limit up to which gaps are filled. If NULL, the upper limit is set to the maximum of period present in data. Only relevant if fill_gaps = TRUE and period is set to a unique timespan type ("year", "decade" or "century").

descending

Whether to sort the resulting table by period in descending or in ascending order.

Value

A tibble.

Details

Note that the period types "year", "decade" and "century" are unique timespans, while "week", "month" and "quarter" are recurring timespans (e.g. every year has a week 1).

See also

Examples

c2d::rfrnds(country_code = "AT") |> c2d::n_rfrnds_per_period()
#> # A tibble: 53 × 2
#>     week     n
#>    <int> <int>
#>  1     1     0
#>  2     2     0
#>  3     3     2
#>  4     4     0
#>  5     5     0
#>  6     6     0
#>  7     7     0
#>  8     8     0
#>  9     9     0
#> 10    10     0
#> # … with 43 more rows
c2d::rfrnds(country_code = "AT") |> c2d::n_rfrnds_per_period(by = level)
#> # A tibble: 159 × 3
#>    level        week     n
#>    <ord>       <int> <int>
#>  1 local           1     0
#>  2 subnational     1     0
#>  3 national        1     0
#>  4 local           2     0
#>  5 subnational     2     0
#>  6 national        2     0
#>  7 local           3     0
#>  8 subnational     3     0
#>  9 national        3     2
#> 10 local           4     0
#> # … with 149 more rows

# without filling gaps
c2d::rfrnds(country_code = "AT") |>
  c2d::n_rfrnds_per_period(by = level,
                           fill_gaps = FALSE)
#> # A tibble: 8 × 3
#>   level        week     n
#>   <ord>       <int> <int>
#> 1 national        3     2
#> 2 national       14     1
#> 3 subnational    16     1
#> 4 subnational    19     1
#> 5 subnational    21     1
#> 6 national       23     1
#> 7 subnational    41     3
#> 8 national       44     1

# per decade and by multiple columns
c2d::rfrnds(country_code = "AT") |>
  c2d::n_rfrnds_per_period(by = c(level, type),
                           period = "decade")
#> # A tibble: 264 × 4
#>    level       type                    decade     n
#>    <ord>       <fct>                    <int> <int>
#>  1 local       mandatory referendum      1910     0
#>  2 local       optional referendum       1910     0
#>  3 local       counter proposal          1910     0
#>  4 local       citizens' initiative      1910     0
#>  5 local       governmental referendum   1910     0
#>  6 local       citizens' assembly        1910     0
#>  7 local       recall                    1910     0
#>  8 local       NA                        1910     0
#>  9 subnational mandatory referendum      1910     0
#> 10 subnational optional referendum       1910     0
#> # … with 254 more rows