Count number of referendums per period

Counts the number of C2D referendums per desired period, optionally by additional columns specified via by.

n_rfrnds_per_period(
  data,
  period = c("week", "month", "quarter", "year", "decade", "century"),
  fill_gaps = TRUE,
  by = NULL
)

Arguments

data
C2D referendum data as returned by rfrnds(). A data frame that at minimum contains the column specified in period or the column date (to compute the period column), plus the one(s) specified via by (if any).

period
Type of period to count referendums by. One of "week","month","quarter","year","decade" or "century".

fill_gaps
Whether or not to add zero-value rows to the result for period gaps in data.

by
Optional data column(s) to group by before counting number of referendums. Tidy selections are supported.

Value

A tibble.

Details

Note that the period types "year", "decade" and "century" are unique timespans, while "week", "month" and "quarter" are not, as they are recurring by nature (e.g. every year has a week 1).

See also

Examples

c2d::rfrnds(country_code = "AT") |> c2d::n_rfrnds_per_period()
#> # A tibble: 53 × 2
#>     week     n
#>    <int> <int>
#>  1     1     0
#>  2     2     0
#>  3     3     2
#>  4     4     0
#>  5     5     0
#>  6     6     0
#>  7     7     0
#>  8     8     0
#>  9     9     0
#> 10    10     0
#> # … with 43 more rows
c2d::rfrnds(country_code = "AT") |> c2d::n_rfrnds_per_period(by = level)
#> # A tibble: 159 × 3
#>    level  week     n
#>    <ord> <int> <int>
#>  1 local     1     0
#>  2 local     2     0
#>  3 local     3     0
#>  4 local     4     0
#>  5 local     5     0
#>  6 local     6     0
#>  7 local     7     0
#>  8 local     8     0
#>  9 local     9     0
#> 10 local    10     0
#> # … with 149 more rows

# without filling gaps
c2d::rfrnds(country_code = "AT") |>
  c2d::n_rfrnds_per_period(by = level,
                           fill_gaps = FALSE)
#> # A tibble: 8 × 3
#>   level        week     n
#>   <ord>       <int> <int>
#> 1 subnational    16     1
#> 2 subnational    19     1
#> 3 subnational    21     1
#> 4 subnational    41     3
#> 5 national        3     2
#> 6 national       14     1
#> 7 national       23     1
#> 8 national       44     1

# per decade and by multiple columns
c2d::rfrnds(country_code = "AT") |>
  c2d::n_rfrnds_per_period(period = "decade",
                           by = c(level, type))
#> # A tibble: 264 × 4
#>    level type                 decade     n
#>    <ord> <fct>                 <int> <int>
#>  1 local mandatory referendum   1910     0
#>  2 local mandatory referendum   1920     0
#>  3 local mandatory referendum   1930     0
#>  4 local mandatory referendum   1940     0
#>  5 local mandatory referendum   1950     0
#>  6 local mandatory referendum   1960     0
#>  7 local mandatory referendum   1970     0
#>  8 local mandatory referendum   1980     0
#>  9 local mandatory referendum   1990     0
#> 10 local mandatory referendum   2000     0
#> # … with 254 more rows