Printify referendum data column names

Renames referendum data column names to be ready to print. Useful e.g. to create publication tables.

printify_col_names(data)

Arguments

data
C2D referendum data as returned by rfrnds().

Value

A tibble.

Details

Column names unknown to this function are not changed.

See also

Examples

c2d::rfrnds(country_code = "IT") |> c2d::printify_col_names()
#> # A tibble: 124 × 69
#>    `C2D ID`   Offic…¹ sudd.…² ISO 3…³ Country Subna…⁴ Munic…⁵ Polit…⁶ Date      
#>    <chr>      <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <ord>   <date>    
#>  1 5f6c6d3fd… NA      NA      IT      Italy   NA      NA      nation… 2020-09-21
#>  2 5d1cc245d… NA      NA      IT      Italy   Lombar… NA      subnat… 2017-10-22
#>  3 5d1cba4ed… NA      NA      IT      Italy   Venetia NA      subnat… 2017-10-22
#>  4 5d19ffd1f… NA      NA      IT      Italy   NA      NA      nation… 2016-12-04
#>  5 5cb86b29c… NA      NA      IT      Italy   NA      NA      nation… 2016-03-15
#>  6 5bbc03929… NA      it0520… IT      Italy   Friaul… NA      subnat… 2014-10-31
#>  7 5bbc03249… NA      it0420… IT      Italy   Venetia NA      subnat… 2014-03-21
#>  8 5bbc03249… NA      it0320… IT      Italy   Venetia NA      subnat… 2014-03-21
#>  9 5bbc03249… NA      it0220… IT      Italy   Venetia NA      subnat… 2014-03-21
#> 10 5bbc03249… NA      it0120… IT      Italy   Venetia NA      subnat… 2014-03-21
#> # … with 114 more rows, 60 more variables: `Former country?` <lgl>,
#> #   Title <chr>, Titel <chr>, Titre <chr>, `Original question` <chr>,
#> #   Question <chr>, `Committee name` <chr>, Result <fct>,
#> #   `Approving subterritories` <dbl>, `Rejecting subterritories` <dbl>,
#> #   `Total electorate` <int>, `Electorate abroad` <int>, `Yes votes` <int>,
#> #   `No votes` <int>, `Empty votes` <int>, `Invalid votes` <int>,
#> #   `Subterritorial votes` <list>, `Lower house yes votes` <int>, …