Printify referendum data column names

Renames referendum data column names to be ready to print. Useful e.g. to create publication tables.

printify_col_names(data)

Arguments

data
C2D referendum data as returned by rfrnds().

Value

A tibble.

Details

Column names unknown to this function are not changed.

See also

Examples

c2d::rfrnds(country_code = "IT") |> c2d::printify_col_names()
#> # A tibble: 128 × 69
#>    `C2D ID`     `Official ID` `sudd.ch ID` `ISO 3166-1 al…` `ISO 3166-3` Country
#>    <chr>        <chr>         <chr>        <chr>            <chr>        <chr>  
#>  1 5f6c6d3fd12… NA            NA           IT               NA           Italy  
#>  2 5d1cc245dc0… NA            NA           IT               NA           Italy  
#>  3 5d1cba4edc0… NA            NA           IT               NA           Italy  
#>  4 5d19ffd1fe6… NA            NA           IT               NA           Italy  
#>  5 5cb86b29cb4… NA            NA           IT               NA           Italy  
#>  6 5bbc039292a… NA            it052014     IT               NA           Italy  
#>  7 5bbc032492a… NA            it042014     IT               NA           Italy  
#>  8 5bbc032492a… NA            it032014     IT               NA           Italy  
#>  9 5bbc032492a… NA            it022014     IT               NA           Italy  
#> 10 5bbc032492a… NA            it012014     IT               NA           Italy  
#> # … with 118 more rows, and 63 more variables: `Subnational entity name` <chr>,
#> #   Municipality <chr>, `Political level` <ord>, `Past jurisdiction?` <lgl>,
#> #   Date <date>, Titel <chr>, Titre <chr>, Title <chr>,
#> #   `Original question` <chr>, Question <chr>, `Committee name` <chr>,
#> #   Result <fct>, `Approving subterritories` <dbl>,
#> #   `Rejecting subterritories` <dbl>, `Total electorate` <int>,
#> #   `Electorate abroad` <int>, `Yes votes` <int>, `No votes` <int>, …