Get referendum data

Downloads the referendum data from the C2D Database. See the codebook for a detailed description of all variables.

referendums(
  country_code = NULL,
  subnational_entity_name = NULL,
  municipality = NULL,
  level = NULL,
  type = NULL,
  date = NULL,
  is_draft = FALSE,
  date_time_created_min = NULL,
  date_time_created_max = NULL,
  query_filter = NULL,
  incl_archive = FALSE,
  tidy = TRUE,
  use_cache = TRUE,
  cache_lifespan = "1 week"
)

Arguments

country_code
The country_code(s) to be included. A character vector.

subnational_entity_name
The subnational_entity_name(s) to be included. A character vector.

municipality
The municipality(s) to be included. A character vector.

level
The level(s) to be included. A character vector.

type
The type(s) to be included. A character vector.

date
The date(s) to be included. A character vector.

is_draft
TRUE means to include only referendum entries with draft status, FALSE to include only normal entries. Set to NULL in order to include both draft and normal entries.

date_time_created_min
The minimum date_time_created to be included. A date, a datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05").

date_time_created_max
The maximum date_time_created to be included. A date, a datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05").

query_filter
A valid MongoDB JSON query filter document which allows for maximum control over what data is included. This takes precedence over all of the above listed parameters, i.e. if query_filter is provided, the parameters country_code, subnational_entity_name, municipality, level, type, date, is_draft, date_time_created_min and date_time_created_max are ignored.

incl_archive
Whether or not to include an archive column containing data from an earlier, obsolete state of the C2D database.

tidy
Whether or not to tidy the referendum data, i.e. apply various data cleansing tasks and add additional variables. If FALSE, the raw MongoDB referendum data will only be modified just enough to be able to return it as a tibble. Note that untidy data doesn't conform to the codebook (i.a. different variable names).

use_cache
Return cached results if possible. If FALSE, results are always newly fetched regardless of cache_lifespan.

cache_lifespan
The duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Only relevant if use_cache = TRUE.

Value

A tibble.

See also

Examples

if (FALSE) {
# get all referendums (excl. drafts)
c2d::referendums()}

# get only referendums in Austria and Australia on subnational level
c2d::referendums(country_code = c("AT", "AU"),
                 level = "subnational")
#> ℹ Fetching JSON data from C2D API...
#> Downloading: 16 kB     Downloading: 16 kB     Downloading: 16 kB     Downloading: 16 kB     Downloading: 16 kB     Downloading: 16 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 65 kB     Downloading: 65 kB     Downloading: 75 kB     Downloading: 75 kB     Downloading: 75 kB     Downloading: 75 kB     
#> ✓ Fetching JSON data from C2D API... done [173ms]
#> 
#> ℹ Converting JSON to list data...
#> ✓ Converting JSON to list data... done [25ms]
#> 
#> ℹ Tidying data...
#> ✓ Tidying data... done [544ms]
#> 
#> # A tibble: 31 × 60
#>    id           id_official id_sudd country_code country_code_hist… country_name
#>    <chr>        <chr>       <chr>   <fct>        <chr>              <fct>       
#>  1 5bbbfb4f92a… 12707       NA      AU           NA                 Australia   
#>  2 5bbbf9b192a… 1219        NA      AU           NA                 Australia   
#>  3 5bbbf92f92a… 12706       NA      AU           NA                 Australia   
#>  4 5bbbf82092a… 1710        NA      AU           NA                 Australia   
#>  5 5bbbf7ff92a… 12705       NA      AU           NA                 Australia   
#>  6 5bbbf70392a… 12704       NA      AU           NA                 Australia   
#>  7 5bbbf59192a… 12703       NA      AU           NA                 Australia   
#>  8 5bbbf4d492a… 872         NA      AU           NA                 Australia   
#>  9 5bbbf4d492a… 871         NA      AU           NA                 Australia   
#> 10 5bbbf29e92a… 12694       NA      AU           NA                 Australia   
#> # … with 21 more rows, and 54 more variables: subnational_entity_name <fct>,
#> #   municipality <fct>, level <ord>, is_past_jurisdiction <lgl>, date <date>,
#> #   title_de <chr>, title_fr <chr>, title_en <chr>, question <chr>,
#> #   question_en <chr>, committee_name <chr>, result <fct>,
#> #   electorate_total <int>, electorate_abroad <int>, votes_yes <int>,
#> #   votes_no <int>, votes_empty <int>, votes_invalid <int>,
#> #   votes_per_subterritory <list>, tags_tier_1 <list>, tags_tier_2 <list>, …

# provide custom `query_filter` for more complex queries like regex matches
# cf. https://docs.mongodb.com/manual/reference/operator/query/regex/
c2d::referendums(query_filter = '{"country_code":{"$regex":"A."}}')
#> ℹ Fetching JSON data from C2D API...
#> Downloading: 16 kB     Downloading: 16 kB     Downloading: 16 kB     Downloading: 16 kB     Downloading: 16 kB     Downloading: 16 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 32 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 65 kB     Downloading: 65 kB     Downloading: 81 kB     Downloading: 81 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 150 kB     Downloading: 150 kB     Downloading: 160 kB     Downloading: 160 kB     Downloading: 160 kB     Downloading: 160 kB     Downloading: 160 kB     Downloading: 160 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 180 kB     Downloading: 200 kB     Downloading: 200 kB     Downloading: 210 kB     Downloading: 210 kB     Downloading: 230 kB     Downloading: 230 kB     Downloading: 250 kB     Downloading: 250 kB     Downloading: 250 kB     Downloading: 250 kB     Downloading: 250 kB     Downloading: 250 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 260 kB     Downloading: 280 kB     Downloading: 280 kB     Downloading: 290 kB     Downloading: 290 kB     Downloading: 310 kB     Downloading: 310 kB     Downloading: 330 kB     Downloading: 330 kB     Downloading: 330 kB     Downloading: 330 kB     Downloading: 330 kB     Downloading: 330 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 340 kB     Downloading: 360 kB     Downloading: 360 kB     Downloading: 380 kB     Downloading: 380 kB     Downloading: 390 kB     Downloading: 390 kB     Downloading: 410 kB     Downloading: 410 kB     Downloading: 410 kB     Downloading: 410 kB     Downloading: 410 kB     Downloading: 410 kB     Downloading: 430 kB     Downloading: 430 kB     Downloading: 430 kB     Downloading: 430 kB     Downloading: 430 kB     Downloading: 430 kB     Downloading: 440 kB     Downloading: 440 kB     Downloading: 440 kB     Downloading: 440 kB     Downloading: 440 kB     Downloading: 440 kB     Downloading: 440 kB     Downloading: 440 kB     Downloading: 460 kB     Downloading: 460 kB     Downloading: 470 kB     Downloading: 470 kB     Downloading: 490 kB     Downloading: 490 kB     Downloading: 510 kB     Downloading: 510 kB     Downloading: 510 kB     Downloading: 510 kB     Downloading: 510 kB     Downloading: 510 kB     Downloading: 510 kB     Downloading: 510 kB     Downloading: 520 kB     Downloading: 520 kB     Downloading: 520 kB     Downloading: 520 kB     Downloading: 520 kB     Downloading: 520 kB     Downloading: 520 kB     Downloading: 520 kB     Downloading: 520 kB     Downloading: 520 kB     Downloading: 520 kB     Downloading: 520 kB     Downloading: 540 kB     Downloading: 540 kB     Downloading: 560 kB     Downloading: 560 kB     Downloading: 570 kB     Downloading: 570 kB     Downloading: 590 kB     Downloading: 590 kB     Downloading: 590 kB     Downloading: 590 kB     Downloading: 610 kB     Downloading: 610 kB     Downloading: 610 kB     Downloading: 610 kB     Downloading: 610 kB     Downloading: 610 kB     Downloading: 620 kB     Downloading: 620 kB     Downloading: 620 kB     Downloading: 620 kB     Downloading: 620 kB     Downloading: 620 kB     Downloading: 640 kB     Downloading: 640 kB     Downloading: 650 kB     Downloading: 650 kB     Downloading: 670 kB     Downloading: 670 kB     Downloading: 670 kB     Downloading: 670 kB     Downloading: 690 kB     Downloading: 690 kB     Downloading: 700 kB     Downloading: 700 kB     Downloading: 700 kB     Downloading: 700 kB     Downloading: 700 kB     Downloading: 700 kB     Downloading: 720 kB     Downloading: 720 kB     Downloading: 720 kB     Downloading: 720 kB     Downloading: 720 kB     Downloading: 720 kB     Downloading: 720 kB     Downloading: 720 kB     Downloading: 740 kB     Downloading: 740 kB     Downloading: 750 kB     Downloading: 750 kB     Downloading: 770 kB     Downloading: 770 kB     Downloading: 790 kB     Downloading: 790 kB     Downloading: 800 kB     Downloading: 800 kB     Downloading: 820 kB     Downloading: 820 kB     Downloading: 820 kB     Downloading: 820 kB     Downloading: 820 kB     Downloading: 820 kB     Downloading: 820 kB     Downloading: 820 kB     Downloading: 820 kB     Downloading: 820 kB     Downloading: 820 kB     Downloading: 820 kB     Downloading: 830 kB     Downloading: 830 kB     Downloading: 850 kB     Downloading: 850 kB     Downloading: 850 kB     Downloading: 850 kB     Downloading: 870 kB     Downloading: 870 kB     Downloading: 880 kB     Downloading: 880 kB     Downloading: 900 kB     Downloading: 900 kB     Downloading: 900 kB     Downloading: 900 kB     
#> ✓ Fetching JSON data from C2D API... done [418ms]
#> 
#> ℹ Converting JSON to list data...
#> ✓ Converting JSON to list data... done [43ms]
#> 
#> ℹ Tidying data...
#> ✓ Tidying data... done [1.2s]
#> 
#> # A tibble: 229 × 60
#>    id          id_official id_sudd country_code country_code_hist… country_name 
#>    <chr>       <chr>       <chr>   <fct>        <chr>              <fct>        
#>  1 611d89b4c7… NA          NA      AT           NA                 Austria      
#>  2 611d82c4c7… NA          NA      AT           NA                 Austria      
#>  3 611d5acac7… NA          NA      AT           NA                 Austria      
#>  4 611d524fc7… NA          NA      AT           NA                 Austria      
#>  5 5d2dd694dc… NA          NA      AS           NA                 American Sam…
#>  6 5d2dbe4fdc… NA          NA      AG           NA                 Antigua and …
#>  7 5d1cc881dc… NA          NA      AU           NA                 Australia    
#>  8 5d1a183bfe… NA          NA      AZ           NA                 Azerbaijan   
#>  9 5cc9a4cccb… NA          NA      AZ           NA                 Azerbaijan   
#> 10 5cc9a3b061… NA          NA      AZ           NA                 Azerbaijan   
#> # … with 219 more rows, and 54 more variables: subnational_entity_name <fct>,
#> #   municipality <fct>, level <ord>, is_past_jurisdiction <lgl>, date <date>,
#> #   title_de <chr>, title_fr <chr>, title_en <chr>, question <chr>,
#> #   question_en <chr>, committee_name <chr>, result <fct>,
#> #   electorate_total <int>, electorate_abroad <int>, votes_yes <int>,
#> #   votes_no <int>, votes_empty <int>, votes_invalid <int>,
#> #   votes_per_subterritory <list>, tags_tier_1 <list>, tags_tier_2 <list>, …