Get referendum data

Downloads the referendum data from the C2D Database. See the codebook for a detailed description of all variables.

referendums(
  use_cache = TRUE,
  cache_lifespan = "1 week",
  incl_archive = FALSE,
  is_draft = FALSE,
  country_code = NULL,
  subnational_entity_name = NULL,
  municipality = NULL,
  level = NULL,
  type = NULL,
  date = NULL,
  date_time_created_min = NULL,
  date_time_created_max = NULL,
  query_filter = NULL,
  tidy = TRUE
)

Arguments

use_cache
Return cached results if possible. If FALSE, results are always newly fetched regardless of cache_lifespan.

cache_lifespan
The duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Only relevant if use_cache = TRUE.

incl_archive
Whether or not to include an archive column containing data from an earlier, obsolete state of the C2D database.

is_draft
TRUE means to include only referendum entries with draft status, FALSE to include only normal entries. Set to NULL in order to include both draft and normal entries.

country_code
The country_code(s) to be included. A character vector.

subnational_entity_name
The subnational_entity_name(s) to be included. A character vector.

municipality
The municipality(s) to be included. A character vector.

level
The level(s) to be included. A character vector.

type
The type(s) to be included. A character vector.

date
The date(s) to be included. A character vector.

date_time_created_min
The minimum date_time_created to be included. A date, a datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05").

date_time_created_max
The maximum date_time_created to be included. A date, a datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05").

query_filter
A valid MongoDB JSON query filter document which allows for maximum control over what data is included. This takes precedence over all the other parameters, i.e. if query_filter is provided, the parameters country_code, subnational_entity_name, municipality, level, type, date, is_draft, date_time_created_min and date_time_created_max)`` are ignored.

tidy
Whether or not to tidy the referendum data, i.e. apply various data cleansing tasks and add additional variables. If FALSE, the raw MongoDB referendum data will only be modified just enough to be able to return it as a tibble. Note that the latter doesn't conform to the codebook (i.a. different variable names).

Value

A tibble.

See also

Examples

if (FALSE) {
# get all referendums (excl. drafts)
c2d::referendums()}

# get only referendums in Austria and Australia on subnational level
c2d::referendums(country_code = c("AT", "AU"),
                 level = "subnational")

#> # A tibble: 31 × 54
#>    id                       id_official id_sudd country_code country_name subnational_ent…
#>    <chr>                    <chr>       <chr>   <fct>        <fct>        <fct>           
#>  1 5bbbfb4f92a21351232e3f97 12707       NA      AU           Australia    Northern Territ…
#>  2 5bbbf9b192a21351232e3ac3 1219        NA      AU           Australia    Christmas Island
#>  3 5bbbf92f92a21351232e3a0d 12706       NA      AU           Australia    Northern Territ…
#>  4 5bbbf82092a21351232e380d 1710        NA      AU           Australia    Northern Territ…
#>  5 5bbbf7ff92a21351232e3787 12705       NA      AU           Australia    Northern Territ…
#>  6 5bbbf70392a21351232e330d 12704       NA      AU           Australia    Northern Territ…
#>  7 5bbbf59192a21351232e2e13 12703       NA      AU           Australia    Northern Territ…
#>  8 5bbbf4d492a21351232e2c47 872         NA      AU           Australia    Christmas Island
#>  9 5bbbf4d492a21351232e2c45 871         NA      AU           Australia    Christmas Island
#> 10 5bbbf29e92a21351232e2345 12694       NA      AU           Australia    Northern Territ…
#> # … with 21 more rows, and 48 more variables: municipality <fct>, level <ord>,
#> #   date <date>, title_de <chr>, title_fr <chr>, title_en <chr>,
#> #   committee_name <chr>, result <fct>, electorate_total <int>,
#> #   electorate_abroad <int>, votes_yes <int>, votes_no <int>,
#> #   votes_empty <int>, votes_invalid <int>, votes_per_subterritory <list>,
#> #   tags_tier_1 <list>, tags_tier_2 <list>, tags_tier_3 <list>, remarks <chr>,
#> #   files <list>, url_sudd <chr>, url_swissvotes <chr>, is_draft <lgl>, …

# provide custom `query_filter` for more complex queries like regex matches
# cf. https://docs.mongodb.com/manual/reference/operator/query/regex/
c2d::referendums(query_filter = '{"country_code":{"$regex":"A."}}')

#> # A tibble: 229 × 54
#>    id                       id_official id_sudd country_code country_name subnational_ent…
#>    <chr>                    <chr>       <chr>   <fct>        <fct>        <fct>           
#>  1 611d89b4c72633da602299ba NA          NA      AT           Austria      NA              
#>  2 611d82c4c72633da602299b6 NA          NA      AT           Austria      NA              
#>  3 611d5acac72633da60229994 NA          NA      AT           Austria      Aargau          
#>  4 611d524fc72633da60229991 NA          NA      AT           Austria      Aargau          
#>  5 5d2dd694dc0aeb467bfcc1e6 NA          NA      AS           American Sa… NA              
#>  6 5d2dbe4fdc0aeb467bfcc1d4 NA          NA      AG           Antigua and… NA              
#>  7 5d1cc881dc0aeb467bfcbfb9 NA          NA      AU           Australia    NA              
#>  8 5d1a183bfe65c2bfeba00bc3 NA          NA      AZ           Azerbaijan   Nagorno-Karabakh
#>  9 5cc9a4cccb4865239961910c NA          NA      AZ           Azerbaijan   NA              
#> 10 5cc9a3b061af8403c235e09a NA          NA      AZ           Azerbaijan   NA              
#> # … with 219 more rows, and 48 more variables: municipality <fct>, level <ord>,
#> #   date <date>, title_de <chr>, title_fr <chr>, title_en <chr>,
#> #   committee_name <chr>, result <fct>, electorate_total <int>,
#> #   electorate_abroad <int>, votes_yes <int>, votes_no <int>,
#> #   votes_empty <int>, votes_invalid <int>, votes_per_subterritory <list>,
#> #   tags_tier_1 <list>, tags_tier_2 <list>, tags_tier_3 <list>, remarks <chr>,
#> #   files <list>, url_sudd <chr>, url_swissvotes <chr>, is_draft <lgl>, …