Get referendum data

Downloads the referendum data from the C2D Database. See the codebook for a detailed description of all variables.

Usage

rfrnds(
  country_code = NULL,
  subnational_entity_name = NULL,
  municipality = NULL,
  level = NULL,
  type = NULL,
  date_min = NULL,
  date_max = NULL,
  is_draft = FALSE,
  date_time_created_min = NULL,
  date_time_created_max = NULL,
  date_time_last_edited_min = NULL,
  date_time_last_edited_max = NULL,
  query_filter = NULL,
  incl_archive = FALSE,
  tidy = TRUE,
  use_cache = TRUE,
  cache_lifespan = "1 week",
  use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg),
  quiet = FALSE
)

Arguments

country_code

The country_code(s) to be included. A character vector.

subnational_entity_name

The subnational_entity_name(s) to be included. A character vector.

municipality

The municipality(s) to be included. A character vector.

level

The level(s) to be included. A character vector.

type

The type(s) to be included. A character vector.

date_min

The minimum date to be included. A date or something coercible to.

date_max

The maximum date to be included. A date or something coercible to.

is_draft

TRUE means to include only referendum entries with draft status, FALSE to include only normal entries. Set to NULL in order to include both draft and normal entries.

date_time_created_min

The minimum date_time_created to be included. A datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05Z"; assumed to be in UTC if no timezone is given).

date_time_created_max

The maximum date_time_created to be included. A datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05Z"; assumed to be in UTC if no timezone is given).

date_time_last_edited_min

The minimum date_time_last_edited to be included. A datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05Z"; assumed to be in UTC if no timezone is given).

date_time_last_edited_max

The maximum date_time_last_edited to be included. A datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05Z"; assumed to be in UTC if no timezone is given).

query_filter

A valid MongoDB JSON query filter document which allows for maximum control over what data is included. This takes precedence over all of the above listed parameters, i.e. if query_filter is provided, the parameters country_code, subnational_entity_name, municipality, level, type, date_min, date_max, is_draft, date_time_created_min, date_time_created_max, date_time_last_edited_min and date_time_last_edited_max are ignored.

incl_archive

Whether or not to include an archive column containing data from an earlier, obsolete state of the C2D database.

tidy

Whether or not to tidy the referendum data, i.e. apply various data cleansing tasks and add additional variables. If FALSE, the raw MongoDB referendum data will only be modified just enough to be able to return it as a tibble. Note that untidy data doesn't conform to the codebook (i.a. different variable names).

use_cache

Return cached results if possible. If FALSE, results are always newly fetched regardless of cache_lifespan.

cache_lifespan

The duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Only relevant if use_cache = TRUE.

use_testing_server

Whether or not to use the testing servers instead of the production servers for C2D Services API calls etc.

quiet

Whether or not to suppress printing status output from internal processing.

Value

A tibble.

See also

Examples

if (FALSE) {
# get all referendums (excl. drafts)
c2d::rfrnds()}

# get only referendums in Austria and Australia on subnational level
c2d::rfrnds(country_code = c("AT", "AU"),
            level = "subnational")
#> # A tibble: 31 × 69
#>    id    id_of…¹ id_sudd count…² count…³ subna…⁴ munic…⁵ level date       is_fo…⁶ title…⁷ title…⁸ title…⁹ quest…˟ quest…˟ commi…˟ result subte…˟ subte…˟ elect…˟
#>    <chr> <chr>   <chr>   <fct>   <fct>   <chr>   <chr>   <ord> <date>     <lgl>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <fct>    <dbl>   <dbl>   <int>
#>  1 5bbb… 12707   NA      AU      Austra… Northe… NA      subn… 2001-10-24 FALSE   Immedi… NA      NA      NA      NA      NA      yes         NA      NA    1140
#>  2 5bbb… 1219    NA      AU      Austra… Christ… NA      subn… 1999-11-06 FALSE   Enhanc… NA      NA      NA      NA      NA      yes         NA      NA      NA
#>  3 5bbb… 12706   NA      AU      Austra… Northe… NA      subn… 1999-05-12 FALSE   NA      NA      Consti… NA      NA      NA      no          NA      NA    1110
#>  4 5bbb… 1710    NA      AU      Austra… Northe… NA      subn… 1998-08-26 FALSE   Title … NA      NA      NA      NA      NA      no          NA      NA    1114
#>  5 5bbb… 12705   NA      AU      Austra… Northe… NA      subn… 1998-05-14 FALSE   Nomina… NA      NA      NA      NA      NA      no          NA      NA      NA
#>  6 5bbb… 12704   NA      AU      Austra… Northe… NA      subn… 1996-10-17 FALSE   Plans … NA      NA      NA      NA      NA      no          NA      NA      NA
#>  7 5bbb… 12703   NA      AU      Austra… Northe… NA      subn… 1995-03-15 FALSE   Quaran… NA      NA      NA      NA      NA      no          NA      NA      NA
#>  8 5bbb… 872     NA      AU      Austra… Christ… NA      subn… 1994-05-07 FALSE   Nation… NA      NA      NA      NA      NA      no          NA      NA      NA
#>  9 5bbb… 871     NA      AU      Austra… Christ… NA      subn… 1994-05-07 FALSE   Nation… NA      NA      NA      NA      NA      yes         NA      NA      NA
#> 10 5bbb… 12694   NA      AU      Austra… Northe… NA      subn… 1991-01-02 FALSE   NA      NA      Consti… NA      NA      NA      no          NA      NA    1051
#> # … with 21 more rows, 49 more variables: electorate_abroad <int>, votes_yes <int>, votes_no <int>, votes_empty <int>, votes_invalid <int>,
#> #   votes_per_subterritory <list>, lower_house_yes <int>, lower_house_no <int>, lower_house_abstentions <int>, upper_house_yes <int>, upper_house_no <int>,
#> #   upper_house_abstentions <int>, position_government <fct>, topics_tier_1 <list>, topics_tier_2 <list>, topics_tier_3 <list>, remarks <chr>, files <list>,
#> #   url_sudd <chr>, url_swissvotes <chr>, sources <chr>, is_draft <lgl>, date_time_created <dttm>, date_time_last_edited <dttm>, type <fct>,
#> #   inst_legal_basis_type <ord>, inst_has_urgent_legal_basis <lgl>, inst_is_binding <lgl>, inst_is_counter_proposal <lgl>, inst_is_assembly <lgl>,
#> #   inst_trigger_type <fct>, inst_trigger_actor <fct>, inst_trigger_actor_level <ord>, inst_trigger_threshold <chr>, inst_trigger_time_limit <chr>,
#> #   inst_object_type <fct>, inst_object_author <fct>, inst_object_legal_level <fct>, inst_object_revision_extent <fct>, inst_object_revision_modes <list>, …

# get referendums in 2020
c2d::rfrnds(date_min = "2020-01-01",
            date_max = "2020-12-31")
#> # A tibble: 103 × 69
#>    id    id_of…¹ id_sudd count…² count…³ subna…⁴ munic…⁵ level date       is_fo…⁶ title…⁷ title…⁸ title…⁹ quest…˟ quest…˟ commi…˟ result subte…˟ subte…˟ elect…˟
#>    <chr> <chr>   <chr>   <fct>   <fct>   <chr>   <chr>   <ord> <date>     <lgl>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <fct>    <dbl>   <dbl>   <int>
#>  1 6010… NA      NA      VE      Venezu… NA      NA      nati… 2020-12-12 FALSE   Reject… Ablehn… NA      NA      NA      NA      yes         NA      NA  2.07e7
#>  2 6010… NA      NA      VE      Venezu… NA      NA      nati… 2020-12-12 FALSE   Initia… Einlei… NA      NA      NA      NA      yes         NA      NA  2.07e7
#>  3 6010… NA      NA      VE      Venezu… NA      NA      nati… 2020-12-12 FALSE   Free e… Freie … NA      NA      NA      NA      yes         NA      NA  2.07e7
#>  4 6010… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08 FALSE   Electi… Wahlbe… NA      NA      NA      NA      yes         NA      NA  2.48e6
#>  5 6010… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08 FALSE   Electi… Wahlte… NA      NA      NA      NA      yes         NA      NA  2.48e6
#>  6 6010… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08 FALSE   Reduct… Kürzun… NA      NA      NA      NA      yes         NA      NA  2.48e6
#>  7 6010… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08 FALSE   Reduct… Kürzun… NA      NA      NA      NA      yes         NA      NA  2.48e6
#>  8 6010… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08 FALSE   Reduct… Kürzun… NA      NA      NA      NA      yes         NA      NA  2.48e6
#>  9 6010… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08 FALSE   Reduct… Kürzun… NA      NA      NA      NA      yes         NA      NA  2.48e6
#> 10 6010… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08 FALSE   Reduct… Kürzun… NA      NA      NA      NA      yes         NA      NA  2.48e6
#> # … with 93 more rows, 49 more variables: electorate_abroad <int>, votes_yes <int>, votes_no <int>, votes_empty <int>, votes_invalid <int>,
#> #   votes_per_subterritory <list>, lower_house_yes <int>, lower_house_no <int>, lower_house_abstentions <int>, upper_house_yes <int>, upper_house_no <int>,
#> #   upper_house_abstentions <int>, position_government <fct>, topics_tier_1 <list>, topics_tier_2 <list>, topics_tier_3 <list>, remarks <chr>, files <list>,
#> #   url_sudd <chr>, url_swissvotes <chr>, sources <chr>, is_draft <lgl>, date_time_created <dttm>, date_time_last_edited <dttm>, type <fct>,
#> #   inst_legal_basis_type <ord>, inst_has_urgent_legal_basis <lgl>, inst_is_binding <lgl>, inst_is_counter_proposal <lgl>, inst_is_assembly <lgl>,
#> #   inst_trigger_type <fct>, inst_trigger_actor <fct>, inst_trigger_actor_level <ord>, inst_trigger_threshold <chr>, inst_trigger_time_limit <chr>,
#> #   inst_object_type <fct>, inst_object_author <fct>, inst_object_legal_level <fct>, inst_object_revision_extent <fct>, inst_object_revision_modes <list>, …

# get referendums added to the database during the last 30 days
c2d::rfrnds(date_time_created_min = clock::date_today(zone = "UTC") |> clock::add_days(-30L),
            date_time_created_max = clock::date_today(zone = "UTC"))
#> # A tibble: 16 × 69
#>    id    id_of…¹ id_sudd count…² count…³ subna…⁴ munic…⁵ level date       is_fo…⁶ title…⁷ title…⁸ title…⁹ quest…˟ quest…˟ commi…˟ result subte…˟ subte…˟ elect…˟
#>    <chr> <chr>   <chr>   <fct>   <fct>   <chr>   <chr>   <ord> <date>     <lgl>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <fct>    <dbl>   <dbl>   <int>
#>  1 6397… NA      NA      CH      Switze… Zurich  NA      subn… 2022-11-27 FALSE   "Count… Gegenv… NA      Stimme… "Do yo… NA      yes         NA      NA  928239
#>  2 6397… NA      NA      CH      Switze… Zurich  NA      subn… 2022-11-27 FALSE   "Citiz… Volksi… NA      Stimme… "Do yo… NA      no          NA      NA  928239
#>  3 6396… NA      NA      CH      Switze… Schwyz  NA      subn… 2022-11-27 FALSE   "Law o… Gesetz… NA      Wollen… "Do yo… NA      yes         NA      NA  107228
#>  4 6396… NA      NA      CH      Switze… Nidwal… NA      subn… 2022-11-27 FALSE   "Objec… Bewill… NA      Wollen… "Do yo… NA      no          NA      NA   31171
#>  5 6396… NA      NA      CH      Switze… Basel … NA      subn… 2022-11-27 FALSE   "Count… Gegenv… NA      Wollen… "Do yo… NA      yes         NA      NA  104260
#>  6 6396… NA      NA      CH      Switze… Basel … NA      subn… 2022-11-26 FALSE   "Citiz… Kanton… NA      Wollen… "Do yo… NA      no          NA      NA  104260
#>  7 6396… NA      NA      CH      Switze… Basel … NA      subn… 2022-11-27 FALSE   "Amend… Änderu… NA      Wollen… "Do yo… NA      yes         NA      NA  190346
#>  8 638d… NA      NA      CH      Switze… Graubü… NA      subn… 2022-11-27 FALSE   "Commi… Verpfl… NA      Wollen… "Do yo… NA      yes         NA      NA  141630
#>  9 638d… NA      NA      CH      Switze… Graubü… NA      subn… 2022-11-27 FALSE   "Parti… Teilre… NA      Wollen… "Do yo… NA      yes         NA      NA  141630
#> 10 638d… NA      NA      CH      Switze… Thurgau NA      subn… 2022-11-27 FALSE   "Objec… Botsch… NA      Stimme… "Do yo… NA      yes         NA      NA  173172
#> 11 638d… NA      NA      CH      Switze… Valais  NA      subn… 2022-11-27 FALSE   "Law o… Gesetz… NA      Nehmen… "Do yo… NA      yes         NA      NA  224250
#> 12 638d… NA      NA      CH      Switze… Valais  NA      subn… 2022-11-27 FALSE   "Amend… Änderu… NA      Nehmen… "Do yo… NA      yes         NA      NA  224250
#> 13 6384… NA      NA      CH      Switze… Ticino  NA      subn… 2022-10-30 FALSE   "Offic… NA      NA      NA       NA     NA      yes         NA      NA  223627
#> 14 6384… NA      NA      CH      Switze… Ticino  NA      subn… 2022-10-30 FALSE   "Refor… NA      NA      NA       NA     NA      yes         NA      NA  223627
#> 15 6384… NA      NA      CH      Switze… Ticino  NA      subn… 2022-10-30 FALSE   "count… NA      NA      NA       NA     NA      no          NA      NA  223627
#> 16 6384… NA      NA      CH      Switze… Ticino  NA      subn… 2022-10-30 FALSE   "citiz… NA      NA      NA       NA     NA      yes         NA      NA  223627
#> # … with 49 more variables: electorate_abroad <int>, votes_yes <int>, votes_no <int>, votes_empty <int>, votes_invalid <int>, votes_per_subterritory <list>,
#> #   lower_house_yes <int>, lower_house_no <int>, lower_house_abstentions <int>, upper_house_yes <int>, upper_house_no <int>, upper_house_abstentions <int>,
#> #   position_government <fct>, topics_tier_1 <list>, topics_tier_2 <list>, topics_tier_3 <list>, remarks <chr>, files <list>, url_sudd <chr>,
#> #   url_swissvotes <chr>, sources <chr>, is_draft <lgl>, date_time_created <dttm>, date_time_last_edited <dttm>, type <fct>, inst_legal_basis_type <ord>,
#> #   inst_has_urgent_legal_basis <lgl>, inst_is_binding <lgl>, inst_is_counter_proposal <lgl>, inst_is_assembly <lgl>, inst_trigger_type <fct>,
#> #   inst_trigger_actor <fct>, inst_trigger_actor_level <ord>, inst_trigger_threshold <chr>, inst_trigger_time_limit <chr>, inst_object_type <fct>,
#> #   inst_object_author <fct>, inst_object_legal_level <fct>, inst_object_revision_extent <fct>, inst_object_revision_modes <list>, inst_is_divisible <lgl>, …

# provide custom `query_filter` for more complex queries like regex matches
# cf. https://docs.mongodb.com/manual/reference/operator/query/regex/
c2d::rfrnds(query_filter = '{"country_code":{"$regex":"A."}}')
#> # A tibble: 232 × 69
#>    id    id_of…¹ id_sudd count…² count…³ subna…⁴ munic…⁵ level date       is_fo…⁶ title…⁷ title…⁸ title…⁹ quest…˟ quest…˟ commi…˟ result subte…˟ subte…˟ elect…˟
#>    <chr> <chr>   <chr>   <fct>   <fct>   <chr>   <chr>   <ord> <date>     <lgl>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <fct>    <dbl>   <dbl>   <int>
#>  1 630e… solala  au?     BQAQ    Britis… NA      NA      nati… 1970-02-02 TRUE    test 4… NA      NA      hm      hm-en   NA      NA          NA      NA NA     
#>  2 630e… 123     ff3     AT      Austria NA      NA      nati… 2222-02-22 FALSE   test 3… NA      NA      TEST    TEST    NA      NA          NA      NA NA     
#>  3 611d… NA      NA      AT      Austria NA      NA      nati… 2222-08-18 FALSE   test s… NA      NA      NA      NA      NA      NA          NA      NA NA     
#>  4 611d… NA      NA      AT      Austria NA      NA      nati… 2222-08-18 FALSE   test s… NA      NA      NA      NA      NA      NA          NA      NA NA     
#>  5 611d… NA      NA      AT      Austria Aargau  NA      subn… 2222-08-18 FALSE   test s… NA      NA      NA      NA      NA      NA          NA      NA NA     
#>  6 611d… NA      NA      AT      Austria Aargau  NA      subn… 2222-08-18 FALSE   test s… NA      NA      NA      NA      NA      NA          NA      NA NA     
#>  7 5d2d… NA      NA      AS      Americ… NA      NA      nati… 2018-11-06 FALSE   Author… NA      NA      NA      NA      NA      no          NA      NA  1.55e4
#>  8 5d2d… NA      NA      AG      Antigu… NA      NA      nati… 2018-11-06 FALSE   Caribb… NA      NA      NA      NA      NA      no          NA      NA  5.30e4
#>  9 5d1c… NA      NA      AU      Austra… NA      NA      nati… 2017-11-07 FALSE   Introd… NA      NA      NA      NA      NA      yes         NA      NA  1.60e7
#> 10 5d1a… NA      NA      AZ      Azerba… Nagorn… NA      subn… 2017-02-20 FALSE   Consti… NA      NA      NA      NA      NA      yes         NA      NA  1.04e5
#> # … with 222 more rows, 49 more variables: electorate_abroad <int>, votes_yes <int>, votes_no <int>, votes_empty <int>, votes_invalid <int>,
#> #   votes_per_subterritory <list>, lower_house_yes <int>, lower_house_no <int>, lower_house_abstentions <int>, upper_house_yes <int>, upper_house_no <int>,
#> #   upper_house_abstentions <int>, position_government <fct>, topics_tier_1 <list>, topics_tier_2 <list>, topics_tier_3 <list>, remarks <chr>, files <list>,
#> #   url_sudd <chr>, url_swissvotes <chr>, sources <chr>, is_draft <lgl>, date_time_created <dttm>, date_time_last_edited <dttm>, type <fct>,
#> #   inst_legal_basis_type <ord>, inst_has_urgent_legal_basis <lgl>, inst_is_binding <lgl>, inst_is_counter_proposal <lgl>, inst_is_assembly <lgl>,
#> #   inst_trigger_type <fct>, inst_trigger_actor <fct>, inst_trigger_actor_level <ord>, inst_trigger_threshold <chr>, inst_trigger_time_limit <chr>,
#> #   inst_object_type <fct>, inst_object_author <fct>, inst_object_legal_level <fct>, inst_object_revision_extent <fct>, inst_object_revision_modes <list>, …