Get referendum data

Downloads the referendum data from the C2D Database. See the codebook for a detailed description of all variables.

rfrnds(
  country_code = NULL,
  subnational_entity_name = NULL,
  municipality = NULL,
  level = NULL,
  type = NULL,
  date_min = NULL,
  date_max = NULL,
  is_draft = FALSE,
  date_time_created_min = NULL,
  date_time_created_max = NULL,
  date_time_last_edited_min = NULL,
  date_time_last_edited_max = NULL,
  query_filter = NULL,
  incl_archive = FALSE,
  tidy = TRUE,
  use_cache = TRUE,
  cache_lifespan = "1 week",
  use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg),
  quiet = FALSE
)

Arguments

country_code
The country_code(s) to be included. A character vector.

subnational_entity_name
The subnational_entity_name(s) to be included. A character vector.

municipality
The municipality(s) to be included. A character vector.

level
The level(s) to be included. A character vector.

type
The type(s) to be included. A character vector.

date_min
The minimum date to be included. A date or something coercible to.

date_max
The maximum date to be included. A date or something coercible to.

is_draft
TRUE means to include only referendum entries with draft status, FALSE to include only normal entries. Set to NULL in order to include both draft and normal entries.

date_time_created_min
The minimum date_time_created to be included. A datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05Z"; assumed to be in UTC if no timezone is given).

date_time_created_max
The maximum date_time_created to be included. A datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05Z"; assumed to be in UTC if no timezone is given).

date_time_last_edited_min
The minimum date_time_last_edited to be included. A datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05Z"; assumed to be in UTC if no timezone is given).

date_time_last_edited_max
The maximum date_time_last_edited to be included. A datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05Z"; assumed to be in UTC if no timezone is given).

query_filter
A valid MongoDB JSON query filter document which allows for maximum control over what data is included. This takes precedence over all of the above listed parameters, i.e. if query_filter is provided, the parameters country_code, subnational_entity_name, municipality, level, type, date_min, date_max, is_draft, date_time_created_min, date_time_created_max, date_time_last_edited_min and date_time_last_edited_max are ignored.

incl_archive
Whether or not to include an archive column containing data from an earlier, obsolete state of the C2D database.

tidy
Whether or not to tidy the referendum data, i.e. apply various data cleansing tasks and add additional variables. If FALSE, the raw MongoDB referendum data will only be modified just enough to be able to return it as a tibble. Note that untidy data doesn't conform to the codebook (i.a. different variable names).

use_cache
Return cached results if possible. If FALSE, results are always newly fetched regardless of cache_lifespan.

cache_lifespan
The duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Only relevant if use_cache = TRUE.

use_testing_server
Whether or not to use the testing servers instead of the production servers for C2D Services API calls etc.

quiet
Whether or not to suppress printing status output from internal processing.

Value

A tibble.

See also

Examples

if (FALSE) {
# get all referendums (excl. drafts)
c2d::rfrnds()}

# get only referendums in Austria and Australia on subnational level
c2d::rfrnds(country_code = c("AT", "AU"),
            level = "subnational")
#> # A tibble: 31 × 69
#>    id           id_of…¹ id_sudd count…² count…³ subna…⁴ munic…⁵ level date      
#>    <chr>        <chr>   <chr>   <fct>   <fct>   <chr>   <chr>   <ord> <date>    
#>  1 5bbbfb4f92a… 12707   NA      AU      Austra… Northe… NA      subn… 2001-10-24
#>  2 5bbbf9b192a… 1219    NA      AU      Austra… Christ… NA      subn… 1999-11-06
#>  3 5bbbf92f92a… 12706   NA      AU      Austra… Northe… NA      subn… 1999-05-12
#>  4 5bbbf82092a… 1710    NA      AU      Austra… Northe… NA      subn… 1998-08-26
#>  5 5bbbf7ff92a… 12705   NA      AU      Austra… Northe… NA      subn… 1998-05-14
#>  6 5bbbf70392a… 12704   NA      AU      Austra… Northe… NA      subn… 1996-10-17
#>  7 5bbbf59192a… 12703   NA      AU      Austra… Northe… NA      subn… 1995-03-15
#>  8 5bbbf4d492a… 872     NA      AU      Austra… Christ… NA      subn… 1994-05-07
#>  9 5bbbf4d492a… 871     NA      AU      Austra… Christ… NA      subn… 1994-05-07
#> 10 5bbbf29e92a… 12694   NA      AU      Austra… Northe… NA      subn… 1991-01-02
#> # … with 21 more rows, 60 more variables: is_former_country <lgl>,
#> #   title_en <chr>, title_de <chr>, title_fr <chr>, question <chr>,
#> #   question_en <chr>, committee_name <chr>, result <fct>,
#> #   subterritories_yes <dbl>, subterritories_no <dbl>, electorate_total <int>,
#> #   electorate_abroad <int>, votes_yes <int>, votes_no <int>,
#> #   votes_empty <int>, votes_invalid <int>, votes_per_subterritory <list>,
#> #   lower_house_yes <int>, lower_house_no <int>, …

# get referendums in 2020
c2d::rfrnds(date_min = "2020-01-01",
            date_max = "2020-12-31")
#> # A tibble: 103 × 69
#>    id           id_of…¹ id_sudd count…² count…³ subna…⁴ munic…⁵ level date      
#>    <chr>        <chr>   <chr>   <fct>   <fct>   <chr>   <chr>   <ord> <date>    
#>  1 6010898bd12… NA      NA      VE      Venezu… NA      NA      nati… 2020-12-12
#>  2 60108788d12… NA      NA      VE      Venezu… NA      NA      nati… 2020-12-12
#>  3 60108540d12… NA      NA      VE      Venezu… NA      NA      nati… 2020-12-12
#>  4 601022ddd12… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08
#>  5 6010202dd12… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08
#>  6 60101d90d12… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08
#>  7 60101b85d12… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08
#>  8 6010195ad12… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08
#>  9 60101462d12… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08
#> 10 60101127d12… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08
#> # … with 93 more rows, 60 more variables: is_former_country <lgl>,
#> #   title_en <chr>, title_de <chr>, title_fr <chr>, question <chr>,
#> #   question_en <chr>, committee_name <chr>, result <fct>,
#> #   subterritories_yes <dbl>, subterritories_no <dbl>, electorate_total <int>,
#> #   electorate_abroad <int>, votes_yes <int>, votes_no <int>,
#> #   votes_empty <int>, votes_invalid <int>, votes_per_subterritory <list>,
#> #   lower_house_yes <int>, lower_house_no <int>, …

# get referendums added to the database during the last 30 days
c2d::rfrnds(date_time_created_min = clock::date_today(zone = "UTC") |> clock::add_days(-30L),
            date_time_created_max = clock::date_today(zone = "UTC"))
#> # A tibble: 15 × 69
#>    id           id_of…¹ id_sudd count…² count…³ subna…⁴ munic…⁵ level date      
#>    <chr>        <chr>   <chr>   <fct>   <fct>   <chr>   <chr>   <ord> <date>    
#>  1 63209b73a52… NA      tn0120… TN      Tunisia NA      NA      nati… 2022-07-25
#>  2 632092c2a52… NA      li0120… LI      Liecht… NA      NA      nati… 2022-06-26
#>  3 63208effa52… NA      it0620… IT      Italy   NA      NA      nati… 2022-06-12
#>  4 63208da3a52… NA      it0520… IT      Italy   NA      NA      nati… 2022-06-12
#>  5 63208be1a52… NA      it0420… IT      Italy   NA      NA      nati… 2022-06-12
#>  6 63208a15a52… NA      it0320… IT      Italy   NA      NA      nati… 2022-06-12
#>  7 632085cda52… NA      it0220… IT      Italy   NA      NA      nati… 2022-06-12
#>  8 6320522fa52… NA      kz0120… KZ      Kazakh… NA      NA      nati… 2022-06-05
#>  9 632035b5a52… NA      dk0120… DK      Denmark NA      NA      nati… 2022-06-01
#> 10 6319fe9fa52… NA      mx0120… MX      Mexico  NA      NA      nati… 2022-04-10
#> 11 6319f91da52… NA      hu0420… HU      Hungary NA      NA      nati… 2022-04-03
#> 12 6319f6f2a52… NA      hu0320… HU      Hungary NA      NA      nati… 2022-04-03
#> 13 6319f6e1a52… NA      hu0220… HU      Hungary NA      NA      nati… 2022-04-03
#> 14 6319f11ea52… NA      hu0120… HU      Hungary NA      NA      nati… 2022-04-03
#> 15 6319d738a52… NA      rs0120… RS      Serbia  NA      NA      nati… 2022-01-16
#> # … with 60 more variables: is_former_country <lgl>, title_en <chr>,
#> #   title_de <chr>, title_fr <chr>, question <chr>, question_en <chr>,
#> #   committee_name <chr>, result <fct>, subterritories_yes <dbl>,
#> #   subterritories_no <dbl>, electorate_total <int>, electorate_abroad <int>,
#> #   votes_yes <int>, votes_no <int>, votes_empty <int>, votes_invalid <int>,
#> #   votes_per_subterritory <list>, lower_house_yes <int>, lower_house_no <int>,
#> #   lower_house_abstentions <int>, upper_house_yes <int>, …

# provide custom `query_filter` for more complex queries like regex matches
# cf. https://docs.mongodb.com/manual/reference/operator/query/regex/
c2d::rfrnds(query_filter = '{"country_code":{"$regex":"A."}}')
#> # A tibble: 232 × 69
#>    id           id_of…¹ id_sudd count…² count…³ subna…⁴ munic…⁵ level date      
#>    <chr>        <chr>   <chr>   <fct>   <fct>   <chr>   <chr>   <ord> <date>    
#>  1 630e49ffa52… solala  au?     BQAQ    Britis… NA      NA      nati… 1970-02-02
#>  2 630e4611a52… 123     ff3     AT      Austria NA      NA      nati… 2222-02-22
#>  3 611d89b4c72… NA      NA      AT      Austria NA      NA      nati… 2222-08-18
#>  4 611d82c4c72… NA      NA      AT      Austria NA      NA      nati… 2222-08-18
#>  5 611d5acac72… NA      NA      AT      Austria Aargau  NA      subn… 2222-08-18
#>  6 611d524fc72… NA      NA      AT      Austria Aargau  NA      subn… 2222-08-18
#>  7 5d2dd694dc0… NA      NA      AS      Americ… NA      NA      nati… 2018-11-06
#>  8 5d2dbe4fdc0… NA      NA      AG      Antigu… NA      NA      nati… 2018-11-06
#>  9 5d1cc881dc0… NA      NA      AU      Austra… NA      NA      nati… 2017-11-07
#> 10 5d1a183bfe6… NA      NA      AZ      Azerba… Nagorn… NA      subn… 2017-02-20
#> # … with 222 more rows, 60 more variables: is_former_country <lgl>,
#> #   title_en <chr>, title_de <chr>, title_fr <chr>, question <chr>,
#> #   question_en <chr>, committee_name <chr>, result <fct>,
#> #   subterritories_yes <dbl>, subterritories_no <dbl>, electorate_total <int>,
#> #   electorate_abroad <int>, votes_yes <int>, votes_no <int>,
#> #   votes_empty <int>, votes_invalid <int>, votes_per_subterritory <list>,
#> #   lower_house_yes <int>, lower_house_no <int>, …