Get referendum data from sudd.ch

Downloads referendum data from sudd.ch.

sudd_rfrnds(ids_sudd, use_cache = TRUE, cache_lifespan = "1 week")

Arguments

ids_sudd
The referendum identifiers assigned by sudd.ch. Either as a character vector or a data frame containing a column id_sudd. NAs are ignored.

use_cache
Return cached results if possible. If FALSE, results are always newly fetched regardless of cache_lifespan.

cache_lifespan
The duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Only relevant if use_cache = TRUE.

Value

A tibble. The column names are aligned with those of rfrnds() as closely as possible.

About sudd.ch

sudd stands for Suchmaschine für direkte Demokratie (German) and is operated by Beat Müller. Its database content is licensed under Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International (CC BY-NC-SA 4.0).

See also

Examples

c2d::rfrnd(id = "5bbc045192a21351232e596f")$id_sudd |> c2d::sudd_rfrnds()
#> # A tibble: 1 × 30
#>   id_sudd count…¹ count…² is_fo…³ subna…⁴ terri…⁵ terri…⁶ date        year month
#>   <chr>   <chr>   <chr>   <lgl>   <chr>   <chr>   <chr>   <date>     <int> <int>
#> 1 li0120… LI      Liecht… FALSE   NA      Liecht… unabhä… 2015-12-13  2015    12
#> # … with 20 more variables: day <int>, title_de <chr>, question_type_de <chr>,
#> #   types <list>, result_de <chr>, result_status_de <chr>,
#> #   adoption_requirements_de <list>, electorate_total <int>,
#> #   polling_cards <int>, votes_total <int>, votes_empty <int>,
#> #   votes_void <int>, votes_invalid <int>, votes_valid <int>, votes_yes <int>,
#> #   votes_no <int>, files <list>, remarks <list>, sources <list>,
#> #   date_last_edited <date>, and abbreviated variable names ¹​country_code, …

c2d::rfrnds(country_code = "AT") |> c2d::sudd_rfrnds()
#> # A tibble: 5 × 29
#>   id_sudd count…¹ count…² is_fo…³ subna…⁴ terri…⁵ terri…⁶ date        year month
#>   <chr>   <chr>   <chr>   <lgl>   <chr>   <chr>   <chr>   <date>     <int> <int>
#> 1 at0120… AT      Austria FALSE   NA      Österr… unabhä… 2013-01-20  2013     1
#> 2 at0120… AT      Austria FALSE   NA      Österr… unabhä… 2013-01-20  2013     1
#> 3 at0119… AT      Austria FALSE   NA      Österr… unabhä… 1994-06-12  1994     6
#> 4 at0119… AT      Austria FALSE   NA      Österr… unabhä… 1978-11-05  1978    11
#> 5 at0119… AT      Austria FALSE   NA      Österr… unabhä… 1938-04-10  1938     4
#> # … with 19 more variables: day <int>, title_de <chr>, question_type_de <chr>,
#> #   types <list>, result_de <chr>, result_status_de <chr>,
#> #   adoption_requirements_de <list>, electorate_total <int>, votes_total <int>,
#> #   votes_invalid <int>, votes_valid <int>, votes_yes <int>, votes_no <int>,
#> #   votes_option_1 <int>, votes_option_2 <int>, files <list>, remarks <list>,
#> #   sources <list>, date_last_edited <date>, and abbreviated variable names
#> #   ¹​country_code, ²​country_name, ³​is_former_country, …