Get the set of possible value labels of a referendum data variable

Returns a character vector of value labels of a specific referendums() column, in the same order as v_vals(), or of length 0 if v_name's values are not restricted to a predefined set or no value labels are defined in the codebook.

val_lbls(v_name, incl_affixes = TRUE)

Arguments

v_name
A variable name. Must be one of the column names of data_codebook.

incl_affixes
Whether or not to add the corresponding value_label_prefix and value_label_suffix to the returned labels.

Value

A character vector. Of length 0 if v_name's values are not restricted to a predefined set or no value labels are defined in the codebook.

See also

Examples

c2d::val_lbls("result",
              incl_affixes = FALSE)

#> [1] "approved" "rejected"

c2d::val_lbls("result")

#> [1] "The referendum proposal was approved"
#> [2] "The referendum proposal was rejected"

# Convert the labels to sentence case with trailing dot
c2d::val_lbls("result") |> pal::sentenceify()

#> [1] "The referendum proposal was approved."
#> [2] "The referendum proposal was rejected."